@@ -3,7 +3,9 @@ name = "resource"
 version = "0.1.0"
 authors = ["The godot-rust developers"]
 publish = false
-edition = "2018"
+edition = "2021"
+rust-version = "1.70"
+license = "MIT"
 
 [lib]
 crate-type = ["cdylib"]
