@@ -2,7 +2,9 @@
 name = "signals"
 version = "0.1.0"
 authors = ["The godot-rust developers"]
-edition = "2018"
+edition = "2021"
+rust-version = "1.70"
+license = "MIT"
 publish = false
 
 [lib]
