@@ -5,21 +5,22 @@ description = "The Godot game engine's automatcally generated bindings to Godot 
 documentation = "https://docs.rs/crate/gdnative-bindings"
 repository = "https://github.com/godot-rust/godot-rust"
 homepage = "https://godot-rust.github.io/"
-version = "0.9.3"
+version = "0.11.3"
 license = "MIT"
 workspace = ".."
-edition = "2018"
+edition = "2021"
+rust-version = "1.70"
 
 [features]
+default = ["one-class-one-file"]
 formatted = []
-one_class_one_file = []
+one-class-one-file = []
+custom-godot = ["gdnative_bindings_generator/custom-godot"]
+ptrcall = ["gdnative_bindings_generator/ptrcall"]
 
 [dependencies]
-gdnative-sys = { path = "../gdnative-sys", version = "0.9.3" }
-gdnative-core = { path = "../gdnative-core", version = "=0.9.3" }
-libc = "0.2.98"
-bitflags = "1.2.1"
+gdnative-core = { path = "../gdnative-core", version = "=0.11.3" }
+libc = "0.2"
 
 [build-dependencies]
-heck = "0.3.3"
-gdnative_bindings_generator = { path = "../bindings_generator", version = "=0.9.3" }
+gdnative_bindings_generator = { path = "../bindings-generator", version = "=0.11.3" }
