@@ -5,28 +5,34 @@ description = "The Godot game engine's gdnative core bindings."
 documentation = "https://docs.rs/crate/gdnative-core"
 repository = "https://github.com/godot-rust/godot-rust"
 homepage = "https://godot-rust.github.io/"
-version = "0.9.3"
+version = "0.11.3"
 license = "MIT"
 workspace = ".."
-edition = "2018"
+edition = "2021"
+rust-version = "1.70"
 
 [features]
-default = ["nativescript"]
-gd_test = []
-nativescript = ["bitflags", "parking_lot"]
-type_tag_fallback = []
+default = []
+gd-test = []
+type-tag-fallback = []
+custom-godot = []
 
 [dependencies]
-gdnative-sys = { path = "../gdnative-sys", version = "0.9.3" }
-libc = "0.2.98"
-approx = "0.5.0"
-glam = "0.18.0"
-indexmap = "1.7.0"
-ahash = "0.7.4"
-once_cell = "1.8.0"
+gdnative-sys = { path = "../gdnative-sys", version = "=0.11.3" }
+gdnative-derive = { path = "../gdnative-derive", version = "=0.11.3" }
+gdnative-impl-proc-macros = { path = "../impl/proc-macros", version = "=0.11.3" }
+ahash = "0.8"
+approx = "0.5"
+atomic-take = "1"
+bitflags = "1"
+glam = "0.27"
+indexmap = "2"
+inventory = { version = "0.3", optional = true }
+libc = "0.2"
+once_cell = "1"
+parking_lot = "0.12"
+semver = "1"
+serde = { version = "1", features = ["derive"], optional = true }
 
-gdnative-impl-proc-macros = { path = "../impl/proc_macros", version = "=0.9.3" }
-
-bitflags = { version = "1.2.1", optional = true }
-parking_lot = { version = "0.11.1", optional = true }
-atomic-take = "1.0.0"
+[dev-dependencies]
+gdnative = { path = "../gdnative" } # for doc-tests
