@@ -5,15 +5,21 @@ description = "The Godot game engine's gdnative derive and procedural macros."
 documentation = "https://docs.rs/crate/gdnative-derive"
 repository = "https://github.com/godot-rust/godot-rust"
 homepage = "https://godot-rust.github.io/"
-version = "0.9.3"
+version = "0.11.3"
 license = "MIT"
 workspace = ".."
-edition = "2018"
+edition = "2021"
+rust-version = "1.70"
 
 [lib]
 proc-macro = true
 
 [dependencies]
-syn = { version = "1.0.74", features = ["full", "extra-traits", "visit"] }
-quote = "1.0.9"
-proc-macro2 = "1.0.28"
+syn = { version = "1.0.84", features = ["full", "extra-traits", "visit"] }
+quote = "1"
+proc-macro2 = "1"
+proc-macro-crate = "3"
+
+[dev-dependencies]
+# This is included for the doc tests.
+gdnative = { path = "../gdnative" }
