@@ -5,17 +5,23 @@ description = "Generated bindings to the Godot game engine's gdnative core types
 documentation = "https://docs.rs/crate/gdnative-sys"
 repository = "https://github.com/godot-rust/godot-rust"
 homepage = "https://godot-rust.github.io/"
-version = "0.9.3"
+version = "0.11.3"
 build = "build.rs"
 license = "MIT"
 workspace = ".."
-edition = "2018"
+edition = "2021"
+rust-version = "1.70"
 
 [dependencies]
-libc = "0.2.98"
+libc = "0.2"
 
 [build-dependencies]
-bindgen = { version = "0.59.1", default-features = false, features = ["runtime"] }
-proc-macro2 = "1.0.28"
-quote = "1.0.9"
-miniserde = "0.1.14"
+bindgen = { version = "0.69.0", default-features = false, features = ["runtime"] }
+proc-macro2 = "1"
+quote = "1"
+miniserde = "0.1.16"
+semver = "1"
+
+# False positives found by cargo-machete
+[package.metadata.cargo-machete]
+ignored = ["libc"]
