@@ -2,31 +2,41 @@
 name = "gdnative"
 authors = ["The godot-rust developers"]
 description = "The Godot game engine's gdnative bindings."
+keywords = ["gamedev", "godot", "engine", "bindings"]
 documentation = "https://docs.rs/crate/gdnative"
 repository = "https://github.com/godot-rust/godot-rust"
 homepage = "https://godot-rust.github.io/"
-version = "0.9.3"
+version = "0.11.3"
 license = "MIT"
 workspace = ".."
 readme = "../README.md"
-edition = "2018"
+edition = "2021"
+rust-version = "1.70"
 
 [features]
-default = ["bindings"]
-formatted = ["gdnative-bindings/formatted", "gdnative-bindings/one_class_one_file"]
+# Public
+default = []
+async = ["gdnative-async"]
+custom-godot = ["gdnative-bindings/custom-godot", "gdnative-core/custom-godot"]
+formatted = ["gdnative-bindings/formatted", "gdnative-bindings/one-class-one-file"]
+ptrcall = ["gdnative-bindings/ptrcall"]
+serde = ["gdnative-core/serde"]
+inventory = ["gdnative-core/inventory"]
 
-gd_test = ["gdnative-core/gd_test"]
-type_tag_fallback = ["gdnative-core/type_tag_fallback"]
-bindings = ["gdnative-bindings"]
+# Internal
+gd-test = ["gdnative-core/gd-test"]
+type-tag-fallback = ["gdnative-core/type-tag-fallback"]
 
 [dependencies]
-gdnative-derive = { path = "../gdnative-derive", version = "=0.9.3" }
-gdnative-core = { path = "../gdnative-core", version = "=0.9.3" }
-gdnative-bindings = { optional = true, path = "../gdnative-bindings", version = "=0.9.3" }
+gdnative-derive = { path = "../gdnative-derive", version = "=0.11.3" }
+gdnative-core = { path = "../gdnative-core", version = "=0.11.3" }
+gdnative-bindings = { path = "../gdnative-bindings", version = "=0.11.3" }
+gdnative-async = { path = "../gdnative-async", version = "=0.11.3", optional = true }
 
 [dev-dependencies]
-trybuild = "1.0.43"
-rustversion = "1.0.5"
+trybuild = "1.0.18" # earrlier versions use broken termcolor 1.0.0
+rustversion = "1"
 
+# See https://docs.rs/about/metadata
 [package.metadata.docs.rs]
-all-features = true
+features = ["async", "serde"]
