@@ -2,7 +2,9 @@
 name = "gdnative-test"
 version = "0.1.0"
 workspace = ".."
-edition = "2018"
+edition = "2021"
+rust-version = "1.70"
+license = "MIT"
 publish = false
 
 [lib]
@@ -10,9 +12,19 @@ crate-type = ["cdylib"]
 
 [features]
 default = []
-type_tag_fallback = ["gdnative/type_tag_fallback"]
+type-tag-fallback = ["gdnative/type-tag-fallback"]
+custom-godot = ["gdnative/custom-godot"]
+ptrcall = ["gdnative/ptrcall"]
+inventory = ["gdnative/inventory"]
+no-manual-register = []
 
 [dependencies]
-gdnative = { path = "../gdnative", features = ["gd_test"] }
-gdnative-derive = { path = "../gdnative-derive" }
-approx = "0.5.0"
+gdnative = { path = "../gdnative", features = ["gd-test", "serde", "async"] }
+gdnative-core = { path = "../gdnative-core" }
+approx = "0.5"
+ron = "0.8"
+serde = "1"
+serde_json = "1"
+bincode = "1"
+rmp-serde = "1"
+futures = "0.3"
