@@ -1,87 +1,65 @@
 use crate::*;
+use std::borrow::Cow;
+use std::collections::{HashMap, HashSet};
 use std::default::Default;
 use std::fmt;
+use std::hash::Hash;
 use std::mem::{forget, transmute};
 use std::ptr;
 
 use crate::core_types::*;
+use crate::object::ownership::*;
 use crate::object::*;
 use crate::private::{get_api, ManuallyManagedClassPlaceholder};
-use crate::thread_access::*;
 
-// TODO: implement Debug, PartialEq, etc.
+#[cfg(feature = "serde")]
+mod serialize;
 
-/// A `Variant` can represent many of godot's core types.
+/// A `Variant` can represent all Godot values (core types or `Object` class instances).
 ///
-/// The underlying data can be either stored inline or reference-counted,
-/// dependning on the size of the type and whether the it is trivially copyable.
+/// The underlying data is either stored inline or reference-counted on the heap,
+/// depending on the size of the type and whether the it is trivially copyable.
+///
+/// If you compile godot-rust with the `serde` feature enabled, you will have
+/// access to serialization/deserialization support: the traits `Serialize`
+/// and `Deserialize` will be automatically implemented on [`VariantDispatch`]
+/// as well as most of the types in [`core_types`].
 pub struct Variant(pub(crate) sys::godot_variant);
 
-macro_rules! variant_constructors {
-    (
-        $(
-            $(#[$attr:meta])*
-            pub fn $ctor:ident($Type:ty) -> Self;
-        )*
-    ) => (
-        $(
-            $(#[$attr])*
-            pub fn $ctor(val: $Type) -> Variant {
-                ToVariant::to_variant(val)
-            }
-        )*
-    )
-}
-
-macro_rules! variant_to_type_transmute {
-    (
-        $(
-            $(#[$to_attr:meta])*
-            pub fn $to_method:ident(&self) -> $ToType:ident : $to_gd_method:ident;
-            $(#[$try_attr:meta])*
-            pub fn $try_method:ident(&self) -> Option<$TryType:ident>;
-        )*
-    ) => (
-        $(
-            $(#[$to_attr])*
-            pub fn $to_method(&self) -> $ToType {
+macro_rules! impl_coerce_from_variant_inner {
+    (impl CoerceFromVariant for $type:path = transmute($to_gd_method:ident)) => {
+        impl private::Sealed for $type {}
+        impl CoerceFromVariant for $type {
+            #[inline]
+            fn coerce_from_variant(v: &Variant) -> Self {
                 unsafe {
                     #[allow(clippy::useless_transmute)]
-                    transmute((get_api().$to_gd_method)(&self.0))
+                    transmute((get_api().$to_gd_method)(&v.0))
                 }
             }
-
-            $(#[$try_attr])*
-            pub fn $try_method(&self) -> Option<$TryType> {
-                $TryType::from_variant(self).ok()
+        }
+    };
+    (impl CoerceFromVariant for $type:path = from_sys($to_gd_method:ident)) => {
+        impl private::Sealed for $type {}
+        impl CoerceFromVariant for $type {
+            #[inline]
+            fn coerce_from_variant(v: &Variant) -> Self {
+                unsafe { Self::from_sys((get_api().$to_gd_method)(&v.0)) }
             }
-        )*
-    )
+        }
+    };
 }
 
-macro_rules! variant_to_type_from_sys {
+macro_rules! impl_coerce_from_variant {
     (
         $(
-            $(#[$to_attr:meta])*
-            pub fn $to_method:ident(&self) -> $ToType:ty : $to_gd_method:ident;
-            $(#[$try_attr:meta])*
-            pub fn $try_method:ident(&self) -> Option<$TryType:ty>;
+            impl CoerceFromVariant for $type:path = $kind:ident ($to_gd_method:ident);
         )*
-    ) => (
+    ) => {
         $(
-            $(#[$to_attr])*
-            pub fn $to_method(&self) -> $ToType {
-                unsafe {
-                    <$ToType>::from_sys((get_api().$to_gd_method)(&self.0))
-                }
-            }
-
-            $(#[$try_attr])*
-            pub fn $try_method(&self) -> Option<$TryType> {
-                <$TryType>::from_variant(self).ok()
-            }
+            impl_coerce_from_variant_inner!(impl CoerceFromVariant for $type = $kind($to_gd_method));
         )*
-    )
+    }
 }
 
 macro_rules! variant_dispatch_arm {
@@ -103,10 +81,18 @@ macro_rules! decl_variant_type {
     ) => {
         #[repr(u32)]
         #[derive(Copy, Clone, Debug, PartialEq, Eq, Hash)]
+        #[non_exhaustive]
         pub enum VariantType {
             $(
                 $variant = $c_const as u32,
             )*
+        }
+
+        impl VariantType {
+            /// The potential names of VariantTypes. Mostly used for serialization.
+            pub const NAMES: &'static [&'static str] = &[
+                $(stringify!($variant),)*
+            ];
         }
 
         /// Rust enum associating each primitive variant type to its value.
@@ -114,6 +100,7 @@ macro_rules! decl_variant_type {
         /// For `Variant`s containing objects, the original `Variant` is returned unchanged, due to
         /// the limitations of statically-determined memory management.
         #[repr(u32)]
+        #[non_exhaustive]
         pub enum VariantDispatch {
             $(
                 $variant $( ($inner) )?,
@@ -129,6 +116,19 @@ macro_rules! decl_variant_type {
                             variant_dispatch_arm!(v, $variant $( ($inner) )?)
                         },
                     )*
+                }
+            }
+        }
+
+        impl<'a> From<&'a VariantDispatch> for Variant {
+            #[inline]
+            fn from(v: &'a VariantDispatch) -> Self {
+                match v {
+                    $($(VariantDispatch::$variant(v) => {
+                        let v: &$inner = v;
+                        v.to_variant()
+                    })?)*
+                    _ => Variant::nil()
                 }
             }
         }
@@ -157,26 +157,38 @@ decl_variant_type!(
         Object(Variant) = sys::godot_variant_type_GODOT_VARIANT_TYPE_OBJECT,
         Dictionary(Dictionary) = sys::godot_variant_type_GODOT_VARIANT_TYPE_DICTIONARY,
         VariantArray(VariantArray) = sys::godot_variant_type_GODOT_VARIANT_TYPE_ARRAY,
-        ByteArray(ByteArray) = sys::godot_variant_type_GODOT_VARIANT_TYPE_POOL_BYTE_ARRAY,
-        Int32Array(Int32Array) = sys::godot_variant_type_GODOT_VARIANT_TYPE_POOL_INT_ARRAY,
-        Float32Array(Float32Array) = sys::godot_variant_type_GODOT_VARIANT_TYPE_POOL_REAL_ARRAY,
-        StringArray(StringArray) = sys::godot_variant_type_GODOT_VARIANT_TYPE_POOL_STRING_ARRAY,
-        Vector2Array(Vector2Array) = sys::godot_variant_type_GODOT_VARIANT_TYPE_POOL_VECTOR2_ARRAY,
-        Vector3Array(Vector3Array) = sys::godot_variant_type_GODOT_VARIANT_TYPE_POOL_VECTOR3_ARRAY,
-        ColorArray(ColorArray) = sys::godot_variant_type_GODOT_VARIANT_TYPE_POOL_COLOR_ARRAY,
+        ByteArray(PoolArray<u8>) = sys::godot_variant_type_GODOT_VARIANT_TYPE_POOL_BYTE_ARRAY,
+        Int32Array(PoolArray<i32>) = sys::godot_variant_type_GODOT_VARIANT_TYPE_POOL_INT_ARRAY,
+        Float32Array(PoolArray<f32>) = sys::godot_variant_type_GODOT_VARIANT_TYPE_POOL_REAL_ARRAY,
+        StringArray(PoolArray<GodotString>) = sys::godot_variant_type_GODOT_VARIANT_TYPE_POOL_STRING_ARRAY,
+        Vector2Array(PoolArray<Vector2>) = sys::godot_variant_type_GODOT_VARIANT_TYPE_POOL_VECTOR2_ARRAY,
+        Vector3Array(PoolArray<Vector3>) = sys::godot_variant_type_GODOT_VARIANT_TYPE_POOL_VECTOR3_ARRAY,
+        ColorArray(PoolArray<Color>) = sys::godot_variant_type_GODOT_VARIANT_TYPE_POOL_COLOR_ARRAY,
     }
 );
 
 impl VariantType {
+    #[allow(clippy::unnecessary_cast)] // False positives: casts necessary for cross-platform
     #[doc(hidden)]
     #[inline]
     pub fn from_sys(v: sys::godot_variant_type) -> VariantType {
         unsafe { transmute(v as u32) }
     }
+
+    /// The `stringify!` representation of this variant. Mostly used for serialization.
+    #[inline]
+    pub const fn name(self) -> &'static str {
+        // NOTE: this assumes that the discriminants remain sequential, since any additions to the
+        // VariantType enum would require a breaking change anyway since it is not marked as non-exhaustive.
+        // See also the Deserialize implementation in the serde submodule.
+        Self::NAMES[self as usize]
+    }
 }
 
+#[allow(clippy::unnecessary_cast)] // False positives: casts necessary for cross-platform
 #[repr(u32)]
 #[derive(Copy, Clone, Debug, PartialEq, Eq, Hash)]
+#[non_exhaustive]
 pub enum CallError {
     InvalidMethod =
         sys::godot_variant_call_error_error_GODOT_CALL_ERROR_CALL_ERROR_INVALID_METHOD as u32,
@@ -191,6 +203,7 @@ pub enum CallError {
 }
 
 impl CallError {
+    #[allow(clippy::unnecessary_cast)] // False positives: casts necessary for cross-platform
     #[inline]
     fn from_sys(v: sys::godot_variant_call_error_error) -> Result<(), CallError> {
         if v == sys::godot_variant_call_error_error_GODOT_CALL_ERROR_CALL_OK {
@@ -223,8 +236,10 @@ impl std::fmt::Display for CallError {
 impl std::error::Error for CallError {}
 
 /// Godot variant operator kind.
+#[allow(clippy::unnecessary_cast)] // False positives: casts necessary for cross-platform
 #[repr(u32)]
 #[derive(Copy, Clone, Debug, PartialEq, Eq, Hash)]
+#[non_exhaustive]
 pub enum VariantOperator {
     // Comparison
     Equal = sys::godot_variant_operator_GODOT_VARIANT_OP_EQUAL as u32,
@@ -262,6 +277,7 @@ pub enum VariantOperator {
     In = sys::godot_variant_operator_GODOT_VARIANT_OP_IN as u32,
 }
 
+#[allow(clippy::unnecessary_cast)] // False positives: casts necessary for cross-platform
 impl VariantOperator {
     const MAX: u32 = sys::godot_variant_operator_GODOT_VARIANT_OP_MAX as u32;
 
@@ -288,89 +304,20 @@ impl VariantOperator {
 #[derive(Copy, Clone, Eq, PartialEq, Hash, Default, Debug)]
 pub struct InvalidOp;
 
-//fn to_godot_varianty_type(v: VariantType) -> sys::godot_variant_type {
-//    unsafe { transmute(v) }
-//}
-
 // These aliases are just here so the type name matches the VariantType's variant names
 // to make writing macros easier.
-type F64 = f64;
-type I64 = i64;
 type Bool = bool;
 
 impl Variant {
-    variant_constructors!(
-        /// Creates a `Variant` wrapping a `Vector2`.
-        #[inline]
-        pub fn from_vector2(&Vector2) -> Self;
-        /// Creates a `Variant` wrapping a `Vector3`.
-        #[inline]
-        pub fn from_vector3(&Vector3) -> Self;
-        /// Creates a `Variant` wrapping a `Quat`.
-        #[inline]
-        pub fn from_quat(&Quat) -> Self;
-        /// Creates a `Variant` wrapping a `Plane`.
-        #[inline]
-        pub fn from_plane(&Plane) -> Self;
-        /// Creates a `Variant` wrapping a `Rect2`.
-        #[inline]
-        pub fn from_rect2(&Rect2) -> Self;
-        /// Creates a `Variant` wrapping a `Transform`.
-        #[inline]
-        pub fn from_transform(&Transform) -> Self;
-        /// Creates a `Variant` wrapping a `Transform2D`.
-        #[inline]
-        pub fn from_transform2d(&Transform2D) -> Self;
-        /// Creates a `Variant` wrapping a `Basis`.
-        #[inline]
-        pub fn from_basis(&Basis) -> Self;
-        /// Creates a `Variant` wrapping a `Color`.
-        #[inline]
-        pub fn from_color(&Color) -> Self;
-        /// Creates a `Variant` wrapping an `Aabb`.
-        #[inline]
-        pub fn from_aabb(&Aabb) -> Self;
-        /// Creates a `Variant` wrapping an `Rid`.
-        #[inline]
-        pub fn from_rid(&Rid) -> Self;
-        /// Creates a `Variant` wrapping a `NodePath`.
-        #[inline]
-        pub fn from_node_path(&NodePath) -> Self;
-        /// Creates a `Variant` wrapping a `GodotString`.
-        #[inline]
-        pub fn from_godot_string(&GodotString) -> Self;
-        /// Creates an `Variant` wrapping an array of variants.
-        #[inline]
-        pub fn from_array(&VariantArray<Shared>) -> Self;
-        /// Creates a `Variant` wrapping a byte array.
-        #[inline]
-        pub fn from_byte_array(&ByteArray) -> Self;
-        /// Creates a `Variant` wrapping an array of 32bit signed integers.
-        #[inline]
-        pub fn from_int32_array(&Int32Array) -> Self;
-        /// Creates a `Variant` wrapping an array of 32bit floats.
-        #[inline]
-        pub fn from_float32_array(&Float32Array) -> Self;
-        /// Creates a `Variant` wrapping an array of godot strings.
-        #[inline]
-        pub fn from_string_array(&StringArray) -> Self;
-        /// Creates a `Variant` wrapping an array of 2d vectors.
-        #[inline]
-        pub fn from_vector2_array(&Vector2Array) -> Self;
-        /// Creates a `Variant` wrapping an array of 3d vectors.
-        #[inline]
-        pub fn from_vector3_array(&Vector3Array) -> Self;
-        /// Creates a `Variant` wrapping an array of colors.
-        #[inline]
-        pub fn from_color_array(&ColorArray) -> Self;
-        /// Creates a `Variant` wrapping a dictionary.
-        #[inline]
-        pub fn from_dictionary(&Dictionary<Shared>) -> Self;
-    );
+    /// Creates a `Variant` from a value that implements [`ToVariant`].
+    #[inline]
+    pub fn new<T: OwnedToVariant>(from: T) -> Self {
+        from.owned_to_variant()
+    }
 
     /// Creates an empty `Variant`.
     #[inline]
-    pub fn new() -> Self {
+    pub fn nil() -> Self {
         unsafe {
             let api = get_api();
             let mut dest = sys::godot_variant::default();
@@ -379,35 +326,7 @@ impl Variant {
         }
     }
 
-    /// Creates a `Variant` wrapping a string.
-    #[inline]
-    #[allow(clippy::should_implement_trait)]
-    pub fn from_str<S>(s: S) -> Variant
-    where
-        S: AsRef<str>,
-    {
-        unsafe {
-            let api = get_api();
-            let mut dest = sys::godot_variant::default();
-            let val = s.as_ref();
-            let mut godot_s =
-                (api.godot_string_chars_to_utf8_with_len)(val.as_ptr() as *const _, val.len() as _);
-            (api.godot_variant_new_string)(&mut dest, &godot_s);
-            (api.godot_string_destroy)(&mut godot_s);
-            Variant(dest)
-        }
-    }
-
-    /// Creates a `Variant` wrapping a Godot object.
-    #[inline]
-    pub fn from_object<R>(val: R) -> Variant
-    where
-        R: AsVariant,
-    {
-        unsafe { R::to_arg_variant(&val) }
-    }
-
-    /// Creats a `Variant` from a raw object pointer.
+    /// Creates a `Variant` from a raw object pointer.
     ///
     /// # Safety
     ///
@@ -419,50 +338,6 @@ impl Variant {
         let mut dest = sys::godot_variant::default();
         (api.godot_variant_new_object)(&mut dest, val);
         Variant(dest)
-    }
-
-    /// Creates a `Variant` wrapping a signed integer value.
-    #[inline]
-    pub fn from_i64(v: i64) -> Variant {
-        unsafe {
-            let api = get_api();
-            let mut dest = sys::godot_variant::default();
-            (api.godot_variant_new_int)(&mut dest, v);
-            Variant(dest)
-        }
-    }
-
-    /// Creates a `Variant` wrapping an unsigned integer value.
-    #[inline]
-    pub fn from_u64(v: u64) -> Variant {
-        unsafe {
-            let api = get_api();
-            let mut dest = sys::godot_variant::default();
-            (api.godot_variant_new_uint)(&mut dest, v);
-            Variant(dest)
-        }
-    }
-
-    /// Creates a `Variant` wrapping a double-precision float value.
-    #[inline]
-    pub fn from_f64(v: f64) -> Variant {
-        unsafe {
-            let api = get_api();
-            let mut ret = sys::godot_variant::default();
-            (api.godot_variant_new_real)(&mut ret, v);
-            Variant(ret)
-        }
-    }
-
-    /// Creates a `Variant` wrapping an boolean.
-    #[inline]
-    pub fn from_bool(v: bool) -> Variant {
-        unsafe {
-            let api = get_api();
-            let mut dest = sys::godot_variant::default();
-            (api.godot_variant_new_bool)(&mut dest, v);
-            Variant(dest)
-        }
     }
 
     #[inline]
@@ -480,219 +355,58 @@ impl Variant {
         Ok(&self.0)
     }
 
-    variant_to_type_transmute!(
-        /// Do a best effort to create a `Vector2` out of the variant, possibly returning a default value.
-        #[inline]
-        pub fn to_vector2(&self) -> Vector2 : godot_variant_as_vector2;
-        /// Returns `Some(Vector2)` if this variant is one, `None` otherwise.
-        #[inline]
-        pub fn try_to_vector2(&self) -> Option<Vector2>;
-
-        /// Do a best effort to create a `Vector3` out of the variant, possibly returning a default value.
-        #[inline]
-        pub fn to_vector3(&self) -> Vector3 : godot_variant_as_vector3;
-        /// Returns `Some(Vector3)` if this variant is one, `None` otherwise.
-        #[inline]
-        pub fn try_to_vector3(&self) -> Option<Vector3>;
-
-        /// Do a best effort to create a `Quat` out of the variant, possibly returning a default value.
-        #[inline]
-        pub fn to_quat(&self) -> Quat : godot_variant_as_quat;
-        /// Returns `Some(Quat)` if this variant is one, `None` otherwise.
-        #[inline]
-        pub fn try_to_quat(&self) -> Option<Quat>;
-
-        /// Do a best effort to create a `Rect2` out of the variant, possibly returning a default value.
-        #[inline]
-        pub fn to_rect2(&self) -> Rect2 : godot_variant_as_rect2;
-        /// Returns `Some(Rect2)` if this variant is one, `None` otherwise.
-        #[inline]
-        pub fn try_to_rect2(&self) -> Option<Rect2>;
-
-        /// Do a best effort to create a `Transform2D` out of the variant, possibly returning a default value.
-        #[inline]
-        pub fn to_transform2d(&self) -> Transform2D : godot_variant_as_transform2d;
-        /// Returns `Some(Transform2D)` if this variant is one, `None` otherwise.
-        #[inline]
-        pub fn try_to_transform2d(&self) -> Option<Transform2D>;
-
-        /// Do a best effort to create a `f64` out of the variant, possibly returning a default value.
-        #[inline]
-        pub fn to_f64(&self) -> F64 : godot_variant_as_real;
-        /// Returns `Some(f64)` if this variant is one, `None` otherwise.
-        #[inline]
-        pub fn try_to_f64(&self) -> Option<F64>;
-
-        /// Do a best effort to create an `i64` out of the variant, possibly returning a default value.
-        #[inline]
-        pub fn to_i64(&self) -> I64 : godot_variant_as_int;
-        /// Returns `Some(i64)` if this variant is one, `None` otherwise.
-        #[inline]
-        pub fn try_to_i64(&self) -> Option<I64>;
-
-        /// Do a best effort to create a `bool` out of the variant, possibly returning a default value.
-        #[inline]
-        pub fn to_bool(&self) -> Bool : godot_variant_as_bool;
-        /// Returns `Some(bool)` if this variant is one, `None` otherwise.
-        #[inline]
-        pub fn try_to_bool(&self) -> Option<Bool>;
-    );
-
-    /// Do a best effort to create a `u64` out of the variant, possibly returning a default value.
+    /// Performs a strongly-typed, structure-aware conversion to `T` from this variant, if it
+    /// is a valid representation of said type. This is the same as `T::from_variant(self).ok()`.
+    ///
+    /// This is the same conversion used to parse arguments of exported methods. See
+    /// [`FromVariant`] for more details.
     #[inline]
-    pub fn to_u64(&self) -> u64 {
-        unsafe {
-            let api = get_api();
-            (api.godot_variant_as_uint)(&self.0)
-        }
+    pub fn to<T: FromVariant>(&self) -> Option<T> {
+        self.try_to().ok()
     }
 
-    /// Returns `Some(u64)` if this variant is one, `None` otherwise.
+    /// Performs a strongly-typed, structure-aware conversion to `T` from this variant, if it
+    /// is a valid representation of said type. This is the same as `T::from_variant(self)`.
+    ///
+    /// This is the same conversion used to parse arguments of exported methods. See
+    /// [`FromVariant`] for more details.
     #[inline]
-    pub fn try_to_u64(&self) -> Option<u64> {
-        unsafe {
-            let api = get_api();
-            if (api.godot_variant_get_type)(&self.0)
-                == sys::godot_variant_type_GODOT_VARIANT_TYPE_INT
-            {
-                Some((api.godot_variant_as_uint)(&self.0))
-            } else {
-                None
-            }
-        }
+    pub fn try_to<T: FromVariant>(&self) -> Result<T, FromVariantError> {
+        T::from_variant(self)
     }
 
-    variant_to_type_from_sys!(
-        /// Do a best effort to create a `Plane` out of the variant, possibly returning a default value.
-        #[inline]
-        pub fn to_plane(&self) -> Plane : godot_variant_as_plane;
-        /// Returns `Some(Plane)` if this variant is one, `None` otherwise.
-        #[inline]
-        pub fn try_to_plane(&self) -> Option<Plane>;
-
-        /// Do a best effort to create a `Transform` out of the variant, possibly returning a default value.
-        #[inline]
-        pub fn to_transform(&self) -> Transform : godot_variant_as_transform;
-        /// Returns `Some(Transform)` if this variant is one, `None` otherwise.
-        #[inline]
-        pub fn try_to_transform(&self) -> Option<Transform>;
-
-        /// Do a best effort to create a `Color` out of the variant, possibly returning a default value.
-        #[inline]
-        pub fn to_color(&self) -> Color : godot_variant_as_color;
-        /// Returns `Some(Color)` if this variant is one, `None` otherwise.
-        #[inline]
-        pub fn try_to_color(&self) -> Option<Color>;
-
-        /// Do a best effort to create a `Basis` out of the variant, possibly returning a default value.
-        #[inline]
-        pub fn to_basis(&self) -> Basis : godot_variant_as_basis;
-        /// Returns `Some(Basis)` if this variant is one, `None` otherwise.
-        #[inline]
-        pub fn try_to_basis(&self) -> Option<Basis>;
-
-        /// Do a best effort to create an `Aabb` out of the variant, possibly returning a default value.
-        #[inline]
-        pub fn to_aabb(&self) -> Aabb : godot_variant_as_aabb;
-        /// Returns `Some(Aabb)` if this variant is one, `None` otherwise.
-        #[inline]
-        pub fn try_to_aabb(&self) -> Option<Aabb>;
-
-        /// Do a best effort to create a `NodePath` out of the variant, possibly returning a default value.
-        #[inline]
-        pub fn to_node_path(&self) -> NodePath : godot_variant_as_node_path;
-        /// Returns `Some(NodePath)` if this variant is one, `None` otherwise.
-        #[inline]
-        pub fn try_to_node_path(&self) -> Option<NodePath>;
-
-        /// Do a best effort to create a `GodotString` out of the variant, possibly returning a default value.
-        #[inline]
-        pub fn to_godot_string(&self) -> GodotString : godot_variant_as_string;
-        /// Returns `Some(GodotString)` if this variant is one, `None` otherwise.
-        #[inline]
-        pub fn try_to_godot_string(&self) -> Option<GodotString>;
-
-        /// Do a best effort to create a `Rid` out of the variant, possibly returning a default value.
-        #[inline]
-        pub fn to_rid(&self) -> Rid : godot_variant_as_rid;
-        /// Returns `Some(Rid)` if this variant is one, `None` otherwise.
-        #[inline]
-        pub fn try_to_rid(&self) -> Option<Rid>;
-
-        /// Do a best effort to create a `VariantArray` out of the variant, possibly returning a default value.
-        #[inline]
-        pub fn to_array(&self) -> VariantArray<Shared> : godot_variant_as_array;
-        /// Returns `Some(VariantArray)` if this variant is one, `None` otherwise.
-        #[inline]
-        pub fn try_to_array(&self) -> Option<VariantArray<Shared>>;
-
-        /// Do a best effort to create a `ByteArray` out of the variant, possibly returning a default value.
-        #[inline]
-        pub fn to_byte_array(&self) -> ByteArray : godot_variant_as_pool_byte_array;
-        /// Returns `Some(ByteArray)` if this variant is one, `None` otherwise.
-        #[inline]
-        pub fn try_to_byte_array(&self) -> Option<ByteArray>;
-
-        /// Do a best effort to create an `Int32Array` out of the variant, possibly returning a default value.
-        #[inline]
-        pub fn to_int32_array(&self) -> Int32Array : godot_variant_as_pool_int_array;
-        /// Returns `Some(Int32Array)` if this variant is one, `None` otherwise.
-        #[inline]
-        pub fn try_to_int32_array(&self) -> Option<Int32Array>;
-
-        /// Do a best effort to create a `Float32Array` out of the variant, possibly returning a default value.
-        #[inline]
-        pub fn to_float32_array(&self) -> Float32Array : godot_variant_as_pool_real_array;
-        /// Returns `Some(Float32Array)` if this variant is one, `None` otherwise.
-        #[inline]
-        pub fn try_to_float32_array(&self) -> Option<Float32Array>;
-
-        /// Do a best effort to create a `StringArray` out of the variant, possibly returning a default value.
-        #[inline]
-        pub fn to_string_array(&self) -> StringArray : godot_variant_as_pool_string_array;
-        /// Returns `Some(StringArray)` if this variant is one, `None` otherwise.
-        #[inline]
-        pub fn try_to_string_array(&self) -> Option<StringArray>;
-
-        /// Do a best effort to create a `Vector2Array` out of the variant, possibly returning a default value.
-        #[inline]
-        pub fn to_vector2_array(&self) -> Vector2Array : godot_variant_as_pool_vector2_array;
-        /// Returns `Some(Vector2Array)` if this variant is one, `None` otherwise.
-        #[inline]
-        pub fn try_to_vector2_array(&self) -> Option<Vector2Array>;
-
-        /// Do a best effort to create a `Vector3Array` out of the variant, possibly returning a default value.
-        #[inline]
-        pub fn to_vector3_array(&self) -> Vector3Array : godot_variant_as_pool_vector3_array;
-        /// Returns `Some(Vector3Array)` if this variant is one, `None` otherwise.
-        #[inline]
-        pub fn try_to_vector3_array(&self) -> Option<Vector3Array>;
-
-        /// Do a best effort to create a `ColorArray` out of the variant, possibly returning a default value.
-        #[inline]
-        pub fn to_color_array(&self) -> ColorArray : godot_variant_as_pool_color_array;
-        /// Returns `Some(ColorArray)` if this variant is one, `None` otherwise.
-        #[inline]
-        pub fn try_to_color_array(&self) -> Option<ColorArray>;
-
-        /// Do a best effort to create a `Dictionary` out of the variant, possibly returning a default value.
-        #[inline]
-        pub fn to_dictionary(&self) -> Dictionary<Shared> : godot_variant_as_dictionary;
-        /// Returns `Some(Dictionary)` if this variant is one, `None` otherwise.
-        #[inline]
-        pub fn try_to_dictionary(&self) -> Option<Dictionary<Shared>>;
-    );
-
+    /// Coerce a value of type `T` out of this variant, through what Godot presents as a
+    /// "best-effort" conversion, possibly returning a default value.
+    ///
+    /// See [`CoerceFromVariant`] for more details.
+    ///
+    /// See also [`Variant::to`] and [`Variant::try_to`] for strongly-typed, structure-aware
+    /// conversions into Rust types.
     #[inline]
-    pub fn try_to_object<T>(&self) -> Option<Ref<T, Shared>>
+    pub fn coerce_to<T: CoerceFromVariant>(&self) -> T {
+        T::coerce_from_variant(self)
+    }
+
+    /// Convenience method to extract a `Ref<T, Shared>` from this variant, if the type
+    /// matches. This is the same as `Ref::<T, Shared>::from_variant(self).ok()`.
+    ///
+    /// This is the same conversion used to parse arguments of exported methods. See
+    /// [`FromVariant`] for more details.
+    #[inline]
+    pub fn to_object<T>(&self) -> Option<Ref<T, Shared>>
     where
         T: GodotObject,
     {
-        self.try_to_object_with_error::<T>().ok()
+        self.try_to_object::<T>().ok()
     }
 
+    /// Convenience method to extract a `Ref<T, Shared>` from this variant, if the type
+    /// matches. This is the same as `Ref::<T, Shared>::from_variant(self)`.
+    ///
+    /// This is the same conversion used to parse arguments of exported methods. See
+    /// [`FromVariant`] for more details.
     #[inline]
-    pub fn try_to_object_with_error<T>(&self) -> Result<Ref<T, Shared>, FromVariantError>
+    pub fn try_to_object<T>(&self) -> Result<Ref<T, Shared>, FromVariantError>
     where
         T: GodotObject,
     {
@@ -712,11 +426,6 @@ impl Variant {
 
             Ok(Ref::from_sys(obj.sys()))
         }
-    }
-
-    #[inline]
-    pub fn try_to_string(&self) -> Option<String> {
-        self.try_to_godot_string().map(|s| s.to_string())
     }
 
     /// Returns this variant's type.
@@ -746,7 +455,14 @@ impl Variant {
     /// Returns true if this is an empty variant.
     #[inline]
     pub fn is_nil(&self) -> bool {
-        self.get_type() == VariantType::Nil
+        match self.get_type() {
+            VariantType::Nil => true,
+            VariantType::Object => {
+                let ptr = unsafe { (get_api().godot_variant_as_object)(&self.0) };
+                ptr.is_null()
+            }
+            _ => false,
+        }
     }
 
     #[inline]
@@ -755,27 +471,31 @@ impl Variant {
         unsafe { (get_api().godot_variant_has_method)(&self.0, &method.0) }
     }
 
+    /// Invokes a method on the held object.
+    ///
+    /// # Safety
+    /// This method may invoke [Object::call()] internally, which is unsafe, as it allows
+    /// execution of arbitrary code (including user-defined code in GDScript or unsafe Rust).
     #[inline]
-    pub fn call(
+    pub unsafe fn call(
         &mut self,
         method: impl Into<GodotString>,
         args: &[Variant],
     ) -> Result<Variant, CallError> {
         let method = method.into();
-        unsafe {
-            let api = get_api();
-            let mut err = sys::godot_variant_call_error::default();
-            let mut arg_refs = args.iter().map(Variant::sys).collect::<Vec<_>>();
-            let variant = (api.godot_variant_call)(
-                &mut self.0,
-                &method.0,
-                arg_refs.as_mut_ptr(),
-                args.len() as i32,
-                &mut err,
-            );
 
-            CallError::from_sys(err.error).map(|_| Variant::from_sys(variant))
-        }
+        let api = get_api();
+        let mut err = sys::godot_variant_call_error::default();
+        let mut arg_refs = args.iter().map(Variant::sys).collect::<Vec<_>>();
+        let variant = (api.godot_variant_call)(
+            &mut self.0,
+            &method.0,
+            arg_refs.as_mut_ptr(),
+            args.len() as i32,
+            &mut err,
+        );
+
+        CallError::from_sys(err.error).map(|_| Variant::from_sys(variant))
     }
 
     /// Evaluates a variant operator on `self` and `rhs` and returns the result on success.
@@ -787,7 +507,7 @@ impl Variant {
     pub fn evaluate(&self, op: VariantOperator, rhs: &Self) -> Result<Variant, InvalidOp> {
         unsafe {
             let api = get_api();
-            let mut ret = Variant::new();
+            let mut ret = Variant::nil();
             let mut valid = false;
 
             (api.godot_variant_evaluate)(
@@ -821,14 +541,14 @@ impl Variant {
         unsafe { &mut *(ptr as *mut variant::Variant) }
     }
 
-    /// Returns the internal ffi representation of the variant and consumes
-    /// the rust object without running the destructor.
+    /// Returns the internal FFI representation of the variant and consumes
+    /// the Rust object without running the destructor.
     ///
-    /// This should be only used when certain that the receiving side is
-    /// responsible for running the destructor for the object, otherwise
-    /// it is leaked.
+    /// The returned object has no `Drop` implementation. The caller is
+    /// responsible of manually ensuring destruction.
     #[inline]
-    pub fn forget(self) -> sys::godot_variant {
+    #[doc(hidden)]
+    pub fn leak(self) -> sys::godot_variant {
         let v = self.0;
         forget(self);
         v
@@ -863,50 +583,79 @@ impl Variant {
     }
 }
 
+impl_coerce_from_variant!(
+    impl CoerceFromVariant for Vector2 = transmute(godot_variant_as_vector2);
+    impl CoerceFromVariant for Vector3 = transmute(godot_variant_as_vector3);
+    impl CoerceFromVariant for Quat = transmute(godot_variant_as_quat);
+    impl CoerceFromVariant for Rect2 = transmute(godot_variant_as_rect2);
+    impl CoerceFromVariant for Transform2D = transmute(godot_variant_as_transform2d);
+    impl CoerceFromVariant for f64 = transmute(godot_variant_as_real);
+    impl CoerceFromVariant for i64 = transmute(godot_variant_as_int);
+    impl CoerceFromVariant for u64 = transmute(godot_variant_as_uint);
+    impl CoerceFromVariant for Bool = transmute(godot_variant_as_bool);
+    impl CoerceFromVariant for Plane = from_sys(godot_variant_as_plane);
+    impl CoerceFromVariant for Transform = from_sys(godot_variant_as_transform);
+    impl CoerceFromVariant for Color = from_sys(godot_variant_as_color);
+    impl CoerceFromVariant for Basis = from_sys(godot_variant_as_basis);
+    impl CoerceFromVariant for Aabb = from_sys(godot_variant_as_aabb);
+    impl CoerceFromVariant for NodePath = from_sys(godot_variant_as_node_path);
+    impl CoerceFromVariant for GodotString = from_sys(godot_variant_as_string);
+    impl CoerceFromVariant for Rid = from_sys(godot_variant_as_rid);
+    impl CoerceFromVariant for VariantArray<Shared> = from_sys(godot_variant_as_array);
+    impl CoerceFromVariant for PoolArray<u8> = from_sys(godot_variant_as_pool_byte_array);
+    impl CoerceFromVariant for PoolArray<i32> = from_sys(godot_variant_as_pool_int_array);
+    impl CoerceFromVariant for PoolArray<f32> = from_sys(godot_variant_as_pool_real_array);
+    impl CoerceFromVariant for PoolArray<GodotString> = from_sys(godot_variant_as_pool_string_array);
+    impl CoerceFromVariant for PoolArray<Vector2> = from_sys(godot_variant_as_pool_vector2_array);
+    impl CoerceFromVariant for PoolArray<Vector3> = from_sys(godot_variant_as_pool_vector3_array);
+    impl CoerceFromVariant for PoolArray<Color> = from_sys(godot_variant_as_pool_color_array);
+    impl CoerceFromVariant for Dictionary<Shared> = from_sys(godot_variant_as_dictionary);
+);
+
 impl_basic_traits_as_sys!(
     for Variant as godot_variant {
         Drop => godot_variant_destroy;
         Clone => godot_variant_new_copy;
         PartialEq => godot_variant_operator_equal;
+        Ord => godot_variant_operator_less;
     }
 );
 
 impl Eq for Variant {}
 
-impl ToString for Variant {
+impl fmt::Display for Variant {
     #[inline]
-    fn to_string(&self) -> String {
-        self.to_godot_string().to_string()
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        self.coerce_to::<GodotString>().fmt(f)
     }
 }
 
 impl Default for Variant {
     #[inline]
     fn default() -> Self {
-        Variant::new()
+        Variant::nil()
     }
 }
 
 impl fmt::Debug for Variant {
     #[inline]
     fn fmt(&self, f: &mut fmt::Formatter) -> Result<(), fmt::Error> {
-        write!(f, "{:?}({})", self.get_type(), self.to_string())
+        write!(f, "{:?}({})", self.get_type(), self)
     }
 }
 
 godot_test!(
     test_variant_nil {
-        let nil = Variant::new();
-        assert_eq!(nil.get_type(), VariantType::Nil);
+        let nil = Variant::nil();
         assert!(nil.is_nil());
 
-        assert!(nil.try_to_array().is_none());
-        assert!(nil.try_to_rid().is_none());
-        assert!(nil.try_to_i64().is_none());
-        assert!(nil.try_to_bool().is_none());
-        assert!(nil.try_to_aabb().is_none());
-        assert!(nil.try_to_vector2().is_none());
-        assert!(nil.try_to_basis().is_none());
+        assert!(nil.try_to::<VariantArray>().is_err());
+        assert!(nil.try_to::<Rid>().is_err());
+        assert!(nil.try_to::<i64>().is_err());
+        assert!(nil.try_to::<bool>().is_err());
+        assert!(nil.try_to::<Aabb>().is_err());
+        assert!(nil.try_to::<Vector2>().is_err());
+        assert!(nil.try_to::<Basis>().is_err());
 
         assert!(!nil.has_method("foo"));
 
@@ -915,39 +664,39 @@ godot_test!(
     }
 
     test_variant_i64 {
-        let v_42 = Variant::from_i64(42);
+        let v_42 = Variant::new(42);
         assert_eq!(v_42.get_type(), VariantType::I64);
 
         assert!(!v_42.is_nil());
-        assert_eq!(v_42.try_to_i64(), Some(42));
-        assert!(v_42.try_to_f64().is_none());
-        assert!(v_42.try_to_array().is_none());
+        assert_eq!(v_42.try_to::<i64>(), Ok(42));
+        assert!(v_42.try_to::<f64>().is_err());
+        assert!(v_42.try_to::<VariantArray>().is_err());
 
-        let v_m1 = Variant::from_i64(-1);
+        let v_m1 = Variant::new(-1);
         assert_eq!(v_m1.get_type(), VariantType::I64);
 
         assert!(!v_m1.is_nil());
-        assert_eq!(v_m1.try_to_i64(), Some(-1));
-        assert!(v_m1.try_to_f64().is_none());
-        assert!(v_m1.try_to_array().is_none());
+        assert_eq!(v_m1.try_to::<i64>(), Ok(-1));
+        assert!(v_m1.try_to::<f64>().is_err());
+        assert!(v_m1.try_to::<VariantArray>().is_err());
     }
 
     test_variant_bool {
-        let v_true = Variant::from_bool(true);
+        let v_true = Variant::new(true);
         assert_eq!(v_true.get_type(), VariantType::Bool);
 
         assert!(!v_true.is_nil());
-        assert_eq!(v_true.try_to_bool(), Some(true));
-        assert!(v_true.try_to_f64().is_none());
-        assert!(v_true.try_to_array().is_none());
+        assert_eq!(v_true.try_to::<bool>(), Ok(true));
+        assert!(v_true.try_to::<f64>().is_err());
+        assert!(v_true.try_to::<VariantArray>().is_err());
 
-        let v_false = Variant::from_bool(false);
+        let v_false = Variant::new(false);
         assert_eq!(v_false.get_type(), VariantType::Bool);
 
         assert!(!v_false.is_nil());
-        assert_eq!(v_false.try_to_bool(), Some(false));
-        assert!(v_false.try_to_f64().is_none());
-        assert!(v_false.try_to_array().is_none());
+        assert_eq!(v_false.try_to::<bool>(), Ok(false));
+        assert!(v_false.try_to::<f64>().is_err());
+        assert!(v_false.try_to::<VariantArray>().is_err());
 
     }
 );
@@ -973,9 +722,22 @@ godot_test!(
 /// - `Struct { a, b, c }` is represented as a `Dictionary` (`{ "a": a, "b": b, "c": c }`)
 /// - `Unit` is represented as an empty `Dictionary` (`{}`)
 /// - `Enum::Variant(a, b, c)` is represented as an externally tagged `Dictionary`
-///   (`{ "Variant": [a, b, c] }`)
+///   (`{ "Variant": [a, b, c] }`), unless another representation is specified with
+///   `#[variant(enum)]` (see below).
 ///
 /// Behavior of the derive macros can be customized using attributes:
+///
+/// ### Item attributes
+///
+/// - `#[variant(enum = "str")]`
+///
+/// Only applicable to field-less enums. Variants of types annotated with this attribute
+/// are represented as stringified values of their names, i.e. `"Variant"` for `Enum::Variant`.
+///
+/// - `#[variant(enum = "repr")]`
+///
+/// Only applicable to field-less enums with a explicit primitive `#[repr]` type. Variants of
+/// types annotated with this attribute are represented as their primitive integral values.
 ///
 /// ### Field attributes
 ///
@@ -1019,6 +781,8 @@ pub trait ToVariant {
 /// This has a blanket implementation for all types that have `ToVariant`. As such, users
 /// should only derive or implement `OwnedToVariant` when `ToVariant` is not applicable.
 ///
+/// This trait is used for return types of exported methods.
+///
 /// ## Deriving `OwnedToVariant`
 ///
 /// The derive macro behaves the same as `ToVariant`. See the documentation for the latter for
@@ -1031,7 +795,7 @@ pub trait OwnedToVariant {
 ///
 /// This means that for all values `a` and `b`, `a == b` is equivalent to
 /// `a.to_variant() == b.to_variant()`. Most of the time, this means that `to_variant` must
-/// return a "value" type, such as a primitive `i32`, a `GodotString`, or a `TypedArray`.
+/// return a "value" type, such as a primitive `i32`, a `GodotString`, or a `PoolArray`.
 ///
 /// This is mostly useful as a bound for `Dictionary` keys, where the difference between Rust's
 /// structural equality and Godot's referential equality semantics can lead to surprising
@@ -1053,7 +817,15 @@ pub trait OwnedToVariant {
 /// ```
 pub trait ToVariantEq: Eq {}
 
-/// Types that can be converted from a `Variant`.
+/// Types that can be converted from a `Variant`. Conversions are performed in Rust, and can be
+/// implemented for custom types.
+///
+/// [`FromVariant`] generally avoids inexact conversions, favoring returning an error instead,
+/// with the noted exception of integer and float types that are not supported by Godot natively.
+/// This is stricter than what GDScript performs by default. For weakly-typed coercions of GDScript
+/// built-in typed, see [`CoerceFromVariant`] instead.
+///
+/// This trait is used for argument types of exported methods.
 ///
 /// ## `Option<T>` and `MaybeNot<T>`
 ///
@@ -1074,7 +846,21 @@ pub trait FromVariant: Sized {
     fn from_variant(variant: &Variant) -> Result<Self, FromVariantError>;
 }
 
+/// Types that can be coerced from a `Variant`. Coercions are provided by Godot, with results
+/// consistent with GDScript. This cannot be implemented for custom types.
+///
+/// [`CoerceFromVariant`] exposes what Godot presents as "best-effort" conversions. These are
+/// weakly-typed conversions to GDScript built-in types, that will never fail, but can return
+/// empty or "default" values, as defined by Godot.
+///
+/// For strongly-typed conversions that can be implemented or derived for custom types, see
+/// [`FromVariant`].
+pub trait CoerceFromVariant: Sized + private::Sealed {
+    fn coerce_from_variant(variant: &Variant) -> Self;
+}
+
 #[derive(Clone, PartialEq, Eq, Debug)]
+#[non_exhaustive]
 /// Error type returned by `FromVariant::from_variant`.
 pub enum FromVariantError {
     /// An unspecified error.
@@ -1129,7 +915,7 @@ pub enum FromVariantError {
     },
 
     /// Given object is not an instance of the expected NativeClass.
-    InvalidInstance { expected: &'static str },
+    InvalidInstance { expected: Cow<'static, str> },
     /// Collection contains an invalid field.
     InvalidField {
         field_name: &'static str,
@@ -1143,11 +929,13 @@ pub enum FromVariantError {
 }
 
 #[derive(Copy, Clone, PartialEq, Eq, Hash, Debug)]
+#[non_exhaustive]
 pub enum VariantEnumRepr {
     ExternallyTagged,
 }
 
 #[derive(Copy, Clone, PartialEq, Eq, Hash, Debug)]
+#[non_exhaustive]
 pub enum VariantStructRepr {
     Unit,
     Tuple,
@@ -1158,7 +946,7 @@ impl FromVariantError {
     /// Returns a `FromVariantError` with a custom message.
     #[inline]
     pub fn custom<T: fmt::Display>(message: T) -> Self {
-        FromVariantError::Custom(format!("{}", message))
+        FromVariantError::Custom(format!("{message}"))
     }
 }
 
@@ -1169,38 +957,31 @@ impl fmt::Display for FromVariantError {
 
         match self {
             E::Unspecified => write!(f, "unspecified error"),
-            E::Custom(s) => write!(f, "{}", s),
+            E::Custom(s) => write!(f, "{s}"),
             E::InvalidNil => write!(f, "expected non-nullable type, got null"),
             E::InvalidVariantType {
                 variant_type,
                 expected,
             } => write!(
                 f,
-                "invalid variant type: expected {:?}, got {:?}",
-                expected, variant_type
+                "invalid variant type: expected {expected:?}, got {variant_type:?}"
             ),
             E::CannotCast { class, to } => {
-                write!(f, "cannot cast object of class {} to {}", class, to)
+                write!(f, "cannot cast object of class {class} to {to}")
             }
             E::InvalidLength { len, expected } => {
-                write!(f, "expected collection of length {}, got {}", expected, len)
+                write!(f, "expected collection of length {expected}, got {len}")
             }
             E::InvalidEnumRepr { expected, error } => write!(
                 f,
-                "invalid enum representation: expected {:?}, {}",
-                expected, error
+                "invalid enum representation: expected {expected:?}, {error}"
             ),
             E::InvalidStructRepr { expected, error } => write!(
                 f,
-                "invalid struct representation: expected {:?}, {}",
-                expected, error
+                "invalid struct representation: expected {expected:?}, {error}"
             ),
             E::UnknownEnumVariant { variant, expected } => {
-                write!(
-                    f,
-                    "unknown enum variant {}, expected variants are: ",
-                    variant
-                )?;
+                write!(f, "unknown enum variant {variant}, expected variants are: ")?;
                 let mut first = true;
                 for v in *expected {
                     if first {
@@ -1208,39 +989,39 @@ impl fmt::Display for FromVariantError {
                     } else {
                         write!(f, ", ")?;
                     }
-                    write!(f, "{}", v)?;
+                    write!(f, "{v}")?;
                 }
                 Ok(())
             }
             E::InvalidEnumVariant { variant, error } => {
-                write!(f, "invalid value for variant {}: {}", variant, error)
+                write!(f, "invalid value for variant {variant}: {error}")
             }
             E::InvalidInstance { expected } => {
-                write!(f, "object is not an instance of `NativeClass` {}", expected)
+                write!(f, "object is not an instance of `NativeClass` {expected}")
             }
             E::InvalidField { field_name, error } => {
-                write!(f, "invalid value for field {}", field_name)?;
+                write!(f, "invalid value for field {field_name}")?;
 
                 let mut next_error = error.as_ref();
                 loop {
                     match next_error {
                         E::InvalidField { field_name, error } => {
-                            write!(f, ".{}", field_name)?;
+                            write!(f, ".{field_name}")?;
                             next_error = error.as_ref();
                         }
                         E::InvalidItem { index, error } => {
-                            write!(f, "[{}]", index)?;
+                            write!(f, "[{index}]")?;
                             next_error = error.as_ref();
                         }
                         _ => {
-                            write!(f, ": {}", next_error)?;
+                            write!(f, ": {next_error}")?;
                             return Ok(());
                         }
                     }
                 }
             }
             E::InvalidItem { index, error } => {
-                write!(f, "invalid value for item at index {}: {}", index, error)
+                write!(f, "invalid value for item at index {index}: {error}")
             }
         }
     }
@@ -1258,7 +1039,7 @@ impl<T: ToVariant> OwnedToVariant for T {
 impl ToVariant for () {
     #[inline]
     fn to_variant(&self) -> Variant {
-        Variant::new()
+        Variant::nil()
     }
 }
 impl ToVariantEq for () {}
@@ -1266,7 +1047,14 @@ impl ToVariantEq for () {}
 impl FromVariant for () {
     #[inline]
     fn from_variant(variant: &Variant) -> Result<Self, FromVariantError> {
-        variant.try_as_sys_of_type(VariantType::Nil).map(|_| ())
+        if variant.is_nil() {
+            return Ok(());
+        }
+
+        Err(FromVariantError::InvalidVariantType {
+            variant_type: variant.get_type(),
+            expected: VariantType::Nil,
+        })
     }
 }
 
@@ -1318,7 +1106,7 @@ impl<'a, T: GodotObject> ToVariant for TRef<'a, T, Shared> {
 impl<T: GodotObject> FromVariant for Ref<T, Shared> {
     #[inline]
     fn from_variant(variant: &Variant) -> Result<Self, FromVariantError> {
-        variant.try_to_object_with_error::<T>()
+        variant.try_to_object::<T>()
     }
 }
 
@@ -1350,7 +1138,12 @@ from_variant_direct!(
 impl ToVariant for i64 {
     #[inline]
     fn to_variant(&self) -> Variant {
-        Variant::from_i64(*self)
+        unsafe {
+            let api = get_api();
+            let mut dest = sys::godot_variant::default();
+            (api.godot_variant_new_int)(&mut dest, *self);
+            Variant(dest)
+        }
     }
 }
 impl ToVariantEq for i64 {}
@@ -1358,15 +1151,38 @@ impl ToVariantEq for i64 {}
 impl ToVariant for u64 {
     #[inline]
     fn to_variant(&self) -> Variant {
-        Variant::from_u64(*self)
+        unsafe {
+            let api = get_api();
+            let mut dest = sys::godot_variant::default();
+            (api.godot_variant_new_uint)(&mut dest, *self);
+            Variant(dest)
+        }
     }
 }
 impl ToVariantEq for u64 {}
 
+impl ToVariant for bool {
+    #[inline]
+    fn to_variant(&self) -> Variant {
+        unsafe {
+            let api = get_api();
+            let mut dest = sys::godot_variant::default();
+            (api.godot_variant_new_bool)(&mut dest, *self);
+            Variant(dest)
+        }
+    }
+}
+impl ToVariantEq for bool {}
+
 impl ToVariant for f64 {
     #[inline]
     fn to_variant(&self) -> Variant {
-        Variant::from_f64(*self)
+        unsafe {
+            let api = get_api();
+            let mut ret = sys::godot_variant::default();
+            (api.godot_variant_new_real)(&mut ret, *self);
+            Variant(ret)
+        }
     }
 }
 
@@ -1379,6 +1195,14 @@ macro_rules! impl_to_variant_for_num {
                 #[inline]
                 fn to_variant(&self) -> Variant {
                     ((*self) as $src_ty).to_variant()
+                }
+            }
+
+            impl private::Sealed for $ty {}
+            impl CoerceFromVariant for $ty {
+                #[inline]
+                fn coerce_from_variant(variant: &Variant) -> Self {
+                    <$src_ty>::coerce_from_variant(variant) as Self
                 }
             }
 
@@ -1556,7 +1380,7 @@ from_variant_from_sys!(
     impl FromVariant for Dictionary<Shared> as Dictionary : godot_variant_as_dictionary;
 );
 
-impl<T: crate::core_types::typed_array::Element> ToVariant for TypedArray<T> {
+impl<T: crate::core_types::PoolElement> ToVariant for PoolArray<T> {
     #[inline]
     fn to_variant(&self) -> Variant {
         unsafe {
@@ -1567,9 +1391,9 @@ impl<T: crate::core_types::typed_array::Element> ToVariant for TypedArray<T> {
         }
     }
 }
-impl<T: crate::core_types::typed_array::Element + Eq> ToVariantEq for TypedArray<T> {}
+impl<T: crate::core_types::PoolElement + Eq> ToVariantEq for PoolArray<T> {}
 
-impl<T: crate::core_types::typed_array::Element> FromVariant for TypedArray<T> {
+impl<T: crate::core_types::PoolElement> FromVariant for PoolArray<T> {
     #[inline]
     fn from_variant(variant: &Variant) -> Result<Self, FromVariantError> {
         unsafe {
@@ -1584,7 +1408,7 @@ impl<T: crate::core_types::typed_array::Element> FromVariant for TypedArray<T> {
 impl ToVariant for str {
     #[inline]
     fn to_variant(&self) -> Variant {
-        Variant::from_str(self)
+        GodotString::from_str(self).owned_to_variant()
     }
 }
 impl ToVariantEq for str {}
@@ -1592,7 +1416,7 @@ impl ToVariantEq for str {}
 impl ToVariant for String {
     #[inline]
     fn to_variant(&self) -> Variant {
-        Variant::from_str(&self)
+        self.as_str().to_variant()
     }
 }
 impl ToVariantEq for String {}
@@ -1603,14 +1427,6 @@ impl FromVariant for String {
         GodotString::from_variant(variant).map(|s| s.to_string())
     }
 }
-
-impl ToVariant for bool {
-    #[inline]
-    fn to_variant(&self) -> Variant {
-        Variant::from_bool(*self)
-    }
-}
-impl ToVariantEq for bool {}
 
 impl ToVariant for Variant {
     #[inline]
@@ -1629,7 +1445,7 @@ impl FromVariant for Variant {
 impl<T> ToVariant for std::marker::PhantomData<T> {
     #[inline]
     fn to_variant(&self) -> Variant {
-        Variant::new()
+        Variant::nil()
     }
 }
 impl<T> ToVariantEq for std::marker::PhantomData<T> {}
@@ -1637,9 +1453,14 @@ impl<T> ToVariantEq for std::marker::PhantomData<T> {}
 impl<T> FromVariant for std::marker::PhantomData<T> {
     #[inline]
     fn from_variant(variant: &Variant) -> Result<Self, FromVariantError> {
-        variant
-            .try_as_sys_of_type(VariantType::Nil)
-            .map(|_| std::marker::PhantomData)
+        if variant.is_nil() {
+            return Ok(std::marker::PhantomData);
+        }
+
+        Err(FromVariantError::InvalidVariantType {
+            variant_type: variant.get_type(),
+            expected: VariantType::Nil,
+        })
     }
 }
 
@@ -1648,7 +1469,7 @@ impl<T: ToVariant> ToVariant for Option<T> {
     fn to_variant(&self) -> Variant {
         match &self {
             Some(thing) => thing.to_variant(),
-            None => Variant::new(),
+            None => Variant::nil(),
         }
     }
 }
@@ -1800,8 +1621,6 @@ impl<T: ToVariant> ToVariant for Vec<T> {
 impl<T: FromVariant> FromVariant for Vec<T> {
     #[inline]
     fn from_variant(variant: &Variant) -> Result<Self, FromVariantError> {
-        use std::convert::TryInto;
-
         let arr = VariantArray::from_variant(variant)?;
         let len: usize = arr
             .len()
@@ -1817,6 +1636,98 @@ impl<T: FromVariant> FromVariant for Vec<T> {
             vec.push(item);
         }
         Ok(vec)
+    }
+}
+
+/// Converts the hash map to a `Dictionary`, wrapped in a `Variant`.
+///
+/// Note that Rust's `HashMap` is non-deterministically ordered for security reasons, meaning that
+/// the order of the same elements will differ between two program invocations. To provide a
+/// deterministic output in Godot (e.g. UI elements for properties), the elements are sorted by key.
+impl<K: ToVariant + Hash + ToVariantEq, V: ToVariant> ToVariant for HashMap<K, V> {
+    #[inline]
+    fn to_variant(&self) -> Variant {
+        // Note: dictionary currently provides neither a sort() function nor random access (or at least bidirectional)
+        // iterators, making it difficult to sort in-place. Workaround: copy to vector
+
+        let mut intermediate: Vec<(Variant, Variant)> = self
+            .iter()
+            .map(|(k, v)| (k.to_variant(), v.to_variant()))
+            .collect();
+
+        intermediate.sort();
+
+        let dict = Dictionary::new();
+        for (key, value) in intermediate.into_iter() {
+            dict.insert(key, value);
+        }
+
+        dict.owned_to_variant()
+    }
+}
+
+/// Expects a `Variant` populated with a `Dictionary` and tries to convert it into a `HashMap`.
+///
+/// Since Rust's `HashMap` is unordered, there is no guarantee about the resulting element order.
+/// In fact it is possible that two program invocations cause a different output.
+impl<K: FromVariant + Hash + Eq, V: FromVariant> FromVariant for HashMap<K, V> {
+    #[inline]
+    fn from_variant(variant: &Variant) -> Result<Self, FromVariantError> {
+        let dictionary = Dictionary::from_variant(variant)?;
+        let len: usize = dictionary
+            .len()
+            .try_into()
+            .expect("Dictionary length should fit in usize");
+
+        let mut hash_map = HashMap::with_capacity(len);
+        for (key, value) in dictionary.iter() {
+            hash_map.insert(K::from_variant(&key)?, V::from_variant(&value)?);
+        }
+        Ok(hash_map)
+    }
+}
+
+/// Converts the hash set to a `VariantArray`, wrapped in a `Variant`.
+///
+/// Note that Rust's `HashSet` is non-deterministically ordered for security reasons, meaning that
+/// the order of the same elements will differ between two program invocations. To provide a
+/// deterministic output in Godot (e.g. UI elements for properties), the elements are sorted by key.
+impl<T: ToVariant> ToVariant for HashSet<T> {
+    #[inline]
+    fn to_variant(&self) -> Variant {
+        let array = VariantArray::new();
+        for value in self {
+            array.push(value.to_variant());
+        }
+
+        array.sort(); // deterministic order in Godot
+        array.owned_to_variant()
+    }
+}
+
+/// Expects a `Variant` populated with a `VariantArray` and tries to convert it into a `HashSet`.
+///
+/// Since Rust's `HashSet` is unordered, there is no guarantee about the resulting element order.
+/// In fact it is possible that two program invocations cause a different output.
+impl<T: FromVariant + Eq + Hash> FromVariant for HashSet<T> {
+    #[inline]
+    fn from_variant(variant: &Variant) -> Result<Self, FromVariantError> {
+        let arr = VariantArray::from_variant(variant)?;
+        let len: usize = arr
+            .len()
+            .try_into()
+            .expect("VariantArray length should fit in usize");
+
+        let mut set = HashSet::with_capacity(len);
+        for idx in 0..len as i32 {
+            let item =
+                T::from_variant(&arr.get(idx)).map_err(|e| FromVariantError::InvalidItem {
+                    index: idx as usize,
+                    error: Box::new(e),
+                })?;
+            set.insert(item);
+        }
+        Ok(set)
     }
 }
 
@@ -1881,44 +1792,48 @@ macro_rules! impl_variant_for_tuples {
 
 impl_variant_for_tuples!(T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12,);
 
+mod private {
+    pub trait Sealed {}
+}
+
 godot_test!(
     test_variant_option {
         use std::marker::PhantomData;
 
         let variant = Some(42_i64).to_variant();
-        assert_eq!(Some(42), variant.try_to_i64());
+        assert_eq!(Ok(42), variant.try_to::<i64>());
 
         let variant = Option::<bool>::None.to_variant();
         assert!(variant.is_nil());
 
-        let variant = Variant::new();
+        let variant = Variant::nil();
         assert_eq!(Ok(None), Option::<i64>::from_variant(&variant));
         assert_eq!(Ok(None), Option::<bool>::from_variant(&variant));
         assert_eq!(Ok(None), Option::<String>::from_variant(&variant));
 
-        let variant = Variant::from_i64(42);
+        let variant = Variant::new(42);
         assert_eq!(Ok(Some(42)), Option::<i64>::from_variant(&variant));
         assert!(Option::<bool>::from_variant(&variant).is_err());
         assert!(Option::<String>::from_variant(&variant).is_err());
 
-        let variant = Variant::new();
+        let variant = Variant::nil();
         assert_eq!(Ok(Some(())), Option::<()>::from_variant(&variant));
         assert_eq!(Ok(Some(PhantomData)), Option::<PhantomData<*const u8>>::from_variant(&variant));
 
-        let variant = Variant::from_i64(42);
+        let variant = Variant::new(42);
         assert!(Option::<PhantomData<*const u8>>::from_variant(&variant).is_err());
     }
 
     test_variant_result {
         let variant = Result::<i64, ()>::Ok(42_i64).to_variant();
-        let dict = variant.try_to_dictionary().expect("should be dic");
-        assert_eq!(Some(42), dict.get("Ok").and_then(|v| v.try_to_i64()));
+        let dict = variant.try_to::<Dictionary>().expect("should be dic");
+        assert_eq!(Some(42), dict.get("Ok").and_then(|v| v.try_to::<i64>().ok()));
 
         let variant = Result::<(), i64>::Err(54_i64).to_variant();
-        let dict = variant.try_to_dictionary().expect("should be dic");
-        assert_eq!(Some(54), dict.get("Err").and_then(|v| v.try_to_i64()));
+        let dict = variant.try_to::<Dictionary>().expect("should be dic");
+        assert_eq!(Some(54), dict.get("Err").and_then(|v| v.try_to::<i64>().ok()));
 
-        let variant = Variant::from_bool(true);
+        let variant = Variant::new(true);
         assert_eq!(
             Err(FromVariantError::InvalidEnumRepr {
                 expected: VariantEnumRepr::ExternallyTagged,
@@ -1942,18 +1857,18 @@ godot_test!(
     test_to_variant_iter {
         let slice: &[i64] = &[0, 1, 2, 3, 4];
         let variant = slice.to_variant();
-        let array = variant.try_to_array().expect("should be array");
+        let array = variant.try_to::<VariantArray>().expect("should be array");
         assert_eq!(5, array.len());
         for i in 0..5 {
-            assert_eq!(Some(i), array.get(i as i32).try_to_i64());
+            assert_eq!(Ok(i), array.get(i as i32).try_to::<i64>());
         }
 
         let vec = Vec::<i64>::from_variant(&variant).expect("should succeed");
         assert_eq!(slice, vec.as_slice());
 
         let het_array = VariantArray::new();
-        het_array.push(&Variant::from_i64(42));
-        het_array.push(&Variant::new());
+        het_array.push(&Variant::new(42));
+        het_array.push(&Variant::nil());
 
         assert_eq!(
             Err(FromVariantError::InvalidItem {
@@ -1984,15 +1899,104 @@ godot_test!(
         let vec_maybe = Vec::<MaybeNot<i64>>::from_variant(&het_array.into_shared().to_variant()).expect("should succeed");
         assert_eq!(3, vec_maybe.len());
         assert_eq!(Some(&42), vec_maybe[0].as_ref().ok());
-        assert_eq!(Some(&Variant::new()), vec_maybe[1].as_ref().err());
+        assert_eq!(Some(&Variant::nil()), vec_maybe[1].as_ref().err());
         assert_eq!(Some(&f64::to_variant(&54.0)), vec_maybe[2].as_ref().err());
+    }
+
+    test_variant_hash_map {
+        let original_hash_map = HashMap::from([
+            ("Foo".to_string(), 4u32),
+            ("Bar".to_string(), 2u32)
+        ]);
+        let variant = original_hash_map.to_variant();
+        let check_hash_map = variant.try_to::<HashMap<String, u32>>().expect("should be hash map");
+        assert_eq!(original_hash_map, check_hash_map);
+        // Check conversion of heterogeneous dictionary key types
+        let non_homogenous_key_dictonary = Dictionary::new();
+        non_homogenous_key_dictonary.insert("Foo".to_string(), 4u32);
+        non_homogenous_key_dictonary.insert(7, 2u32);
+        assert_eq!(
+            non_homogenous_key_dictonary.owned_to_variant().try_to::<HashMap<String, u32>>(),
+            Err(FromVariantError::InvalidVariantType {
+                variant_type: VariantType::I64,
+                expected: VariantType::GodotString
+            }),
+        );
+        // Check conversion of heterogeneous dictionary value types
+        let non_homogenous_value_dictonary = Dictionary::new();
+        non_homogenous_value_dictonary.insert("Foo".to_string(), 4u32);
+        non_homogenous_value_dictonary.insert("Bar".to_string(), "Unexpected".to_string());
+        assert_eq!(
+            non_homogenous_value_dictonary.owned_to_variant().try_to::<HashMap<String, u32>>(),
+            Err(FromVariantError::InvalidVariantType {
+                variant_type: VariantType::GodotString,
+                expected: VariantType::I64
+            }),
+        );
+    }
+
+    test_variant_hash_set {
+        let original_hash_set = HashSet::from([
+            "Foo".to_string(),
+            "Bar".to_string(),
+        ]);
+        let variant = original_hash_set.to_variant();
+        let check_hash_set = variant.try_to::<HashSet<String>>().expect("should be hash set");
+        assert_eq!(original_hash_set, check_hash_set);
+
+        let duplicate_variant_set = VariantArray::new();
+        duplicate_variant_set.push("Foo".to_string());
+        duplicate_variant_set.push("Bar".to_string());
+        duplicate_variant_set.push("Bar".to_string());
+        let duplicate_hash_set = variant.try_to::<HashSet<String>>().expect("should be hash set");
+        assert_eq!(original_hash_set, duplicate_hash_set);
+
+        // Check conversion of heterogeneous set types
+        let non_homogenous_set = VariantArray::new();
+        non_homogenous_set.push("Foo".to_string());
+        non_homogenous_set.push(7);
+        assert_eq!(
+            non_homogenous_set.owned_to_variant().try_to::<HashSet<String>>(),
+            Err(FromVariantError::InvalidItem {
+                index: 1,
+                error: Box::new(FromVariantError::InvalidVariantType {
+                    variant_type: VariantType::I64,
+                    expected: VariantType::GodotString
+                })
+            }),
+        );
+    }
+
+    test_variant_vec {
+        let original_vec = Vec::from([
+            "Foo".to_string(),
+            "Bar".to_string(),
+        ]);
+        let variant = original_vec.to_variant();
+        let check_vec = variant.try_to::<Vec<String>>().expect("should be hash set");
+        assert_eq!(original_vec, check_vec);
+
+        // Check conversion of heterogeneous vec types
+        let non_homogenous_vec = VariantArray::new();
+        non_homogenous_vec.push("Foo".to_string());
+        non_homogenous_vec.push(7);
+        assert_eq!(
+            non_homogenous_vec.owned_to_variant().try_to::<Vec<String>>(),
+            Err(FromVariantError::InvalidItem {
+                index: 1,
+                error: Box::new(FromVariantError::InvalidVariantType {
+                    variant_type: VariantType::I64,
+                    expected: VariantType::GodotString
+                })
+            }),
+        );
     }
 
     test_variant_tuple {
         let variant = (42i64, 54i64).to_variant();
-        let arr = variant.try_to_array().expect("should be array");
-        assert_eq!(Some(42), arr.get(0).try_to_i64());
-        assert_eq!(Some(54), arr.get(1).try_to_i64());
+        let arr = variant.try_to::<VariantArray>().expect("should be array");
+        assert_eq!(Ok(42), arr.get(0).try_to::<i64>());
+        assert_eq!(Ok(54), arr.get(1).try_to::<i64>());
 
         let tuple = <(i64, i64)>::from_variant(&variant);
         assert_eq!(Ok((42, 54)), tuple);
