@@ -1,18 +1,22 @@
 use proc_macro2::{Literal, Span, TokenStream as TokenStream2};
-use syn::{Fields, Ident, Type};
+use syn::{DataEnum, Fields, Ident, Type};
 
-use super::attr::{Attr, AttrBuilder};
-use super::ToVariantTrait;
+use super::attr::{FieldAttr, FieldAttrBuilder, ItemAttr};
+use super::{parse_attrs, ToVariantTrait};
 
+// Shouldn't matter since this is immediately unpacked anyway.
+// Boxing would add too much noise to the match statements.
+#[allow(clippy::large_enum_variant)]
 #[derive(Clone, Eq, PartialEq, Debug)]
 pub(crate) enum Repr {
-    Struct(VariantRepr),
-    Enum(Vec<(Ident, VariantRepr)>),
+    Struct(StructRepr),
+    Enum(EnumRepr),
 }
 
+#[allow(clippy::large_enum_variant)]
 #[derive(Clone, Eq, PartialEq, Debug)]
 pub(crate) enum VariantRepr {
-    Unit,
+    Unit(Option<syn::Expr>),
     Struct(Vec<Field>),
     Tuple(Vec<Field>),
 }
@@ -21,32 +25,71 @@ pub(crate) enum VariantRepr {
 pub(crate) struct Field {
     pub ident: Ident,
     pub ty: Type,
-    pub attr: Attr,
+    pub attr: FieldAttr,
 }
 
-fn improve_meta_error(err: syn::Error) -> syn::Error {
-    let error = err.to_string();
-    match error.as_str() {
-        "expected literal" => {
-            syn::Error::new(err.span(), "String expected, wrap with double quotes.")
-        }
-        other => syn::Error::new(
-            err.span(),
-            format!("{}, ie: #[variant(with = \"...\")]", other),
-        ),
+#[derive(Clone, Eq, PartialEq, Debug)]
+pub(crate) struct StructRepr(pub VariantRepr);
+
+#[derive(Clone, Eq, PartialEq, Debug)]
+pub(crate) struct EnumRepr {
+    pub kind: EnumReprKind,
+    pub primitive_repr: Option<Type>,
+    pub variants: Vec<(Ident, VariantRepr)>,
+}
+
+#[derive(Copy, Clone, Eq, PartialEq, Debug)]
+pub enum EnumReprKind {
+    /// Externally-tagged objects, i.e. the original behavior.
+    External,
+    /// The integer type specified by the `repr` attribute of the enum.
+    Repr,
+    /// Represent as strings.
+    Str,
+}
+
+impl EnumRepr {
+    pub(crate) fn repr_for(
+        attr: ItemAttr,
+        primitive_repr: Option<syn::Type>,
+        enum_data: &DataEnum,
+    ) -> Result<Self, syn::Error> {
+        let variants = enum_data
+            .variants
+            .iter()
+            .map(|variant| {
+                let mut repr = VariantRepr::repr_for(&variant.fields)?;
+                if let VariantRepr::Unit(discriminant) = &mut repr {
+                    if let Some((_, expr)) = &variant.discriminant {
+                        *discriminant = Some(expr.clone());
+                    }
+                }
+
+                Ok((variant.ident.clone(), repr))
+            })
+            .collect::<Result<_, syn::Error>>()?;
+
+        Ok(EnumRepr {
+            kind: attr
+                .enum_repr_kind
+                .map_or(EnumReprKind::External, |(kind, _)| kind),
+            primitive_repr,
+            variants,
+        })
     }
 }
 
-fn parse_attrs<'a, I>(attrs: I) -> Result<Attr, syn::Error>
-where
-    I: IntoIterator<Item = &'a syn::Attribute>,
-{
-    attrs
-        .into_iter()
-        .filter(|attr| attr.path.is_ident("variant"))
-        .map(|attr| attr.parse_meta().map_err(improve_meta_error))
-        .collect::<Result<AttrBuilder, syn::Error>>()?
-        .done()
+impl StructRepr {
+    pub(crate) fn repr_for(attr: ItemAttr, fields: &Fields) -> Result<Self, syn::Error> {
+        if let Some((_, span)) = attr.enum_repr_kind {
+            return Err(syn::Error::new(
+                span,
+                "`enum` representation can only be set for enums",
+            ));
+        }
+
+        VariantRepr::repr_for(fields).map(StructRepr)
+    }
 }
 
 impl VariantRepr {
@@ -59,7 +102,7 @@ impl VariantRepr {
                     .map(|f| {
                         let ident = f.ident.clone().expect("fields should be named");
                         let ty = f.ty.clone();
-                        let attr = parse_attrs(&f.attrs)?;
+                        let attr = parse_attrs::<FieldAttrBuilder, _>(&f.attrs)?;
                         Ok(Field { ident, ty, attr })
                     })
                     .collect::<Result<Vec<_>, syn::Error>>()?,
@@ -70,14 +113,14 @@ impl VariantRepr {
                     .iter()
                     .enumerate()
                     .map(|(n, f)| {
-                        let ident = Ident::new(&format!("__field_{}", n), Span::call_site());
+                        let ident = Ident::new(&format!("__field_{n}"), Span::call_site());
                         let ty = f.ty.clone();
-                        let attr = parse_attrs(&f.attrs)?;
+                        let attr = parse_attrs::<FieldAttrBuilder, _>(&f.attrs)?;
                         Ok(Field { ident, ty, attr })
                     })
                     .collect::<Result<_, syn::Error>>()?,
             ),
-            Fields::Unit => VariantRepr::Unit,
+            Fields::Unit => VariantRepr::Unit(None),
         };
 
         Ok(this)
@@ -85,7 +128,7 @@ impl VariantRepr {
 
     pub(crate) fn destructure_pattern(&self) -> TokenStream2 {
         match self {
-            VariantRepr::Unit => quote! {},
+            VariantRepr::Unit(_) => quote! {},
             VariantRepr::Tuple(fields) => {
                 let names = fields.iter().map(|f| &f.ident);
                 quote! {
@@ -101,31 +144,31 @@ impl VariantRepr {
         }
     }
 
-    pub(crate) fn to_variant(
+    pub(crate) fn make_to_variant_expr(
         &self,
         trait_kind: ToVariantTrait,
     ) -> Result<TokenStream2, syn::Error> {
         let tokens = match self {
-            VariantRepr::Unit => {
+            VariantRepr::Unit(_) => {
                 quote! { ::gdnative::core_types::Dictionary::new().into_shared().to_variant() }
             }
             VariantRepr::Tuple(fields) => {
                 if fields.len() == 1 {
                     // as newtype
-                    let field = fields.get(0).unwrap();
+                    let field = fields.first().unwrap();
                     if field.attr.skip_to_variant {
                         return Err(syn::Error::new(
                             field.ident.span(),
                             "cannot skip the only field in a tuple",
                         ));
                     }
-                    field.to_variant(trait_kind)
+                    field.make_to_variant_expr(trait_kind)
                 } else {
                     let exprs = fields.iter().filter_map(|f| {
                         if f.attr.skip_to_variant {
                             None
                         } else {
-                            Some(f.to_variant(trait_kind))
+                            Some(f.make_to_variant_expr(trait_kind))
                         }
                     });
 
@@ -150,7 +193,7 @@ impl VariantRepr {
                 let name_string_literals =
                     name_strings.iter().map(|string| Literal::string(string));
 
-                let exprs = fields.iter().map(|f| f.to_variant(trait_kind));
+                let exprs = fields.iter().map(|f| f.make_to_variant_expr(trait_kind));
 
                 quote! {
                     {
@@ -170,13 +213,13 @@ impl VariantRepr {
         Ok(tokens)
     }
 
-    pub(crate) fn from_variant(
+    pub(crate) fn make_from_variant_expr(
         &self,
         variant: &Ident,
         ctor: &TokenStream2,
     ) -> Result<TokenStream2, syn::Error> {
         let tokens = match self {
-            VariantRepr::Unit => {
+            VariantRepr::Unit(_) => {
                 quote! {
                     if #variant.is_nil() {
                         Err(FVE::InvalidStructRepr {
@@ -192,14 +235,14 @@ impl VariantRepr {
             VariantRepr::Tuple(fields) => {
                 if fields.len() == 1 {
                     // as newtype
-                    let field = fields.get(0).unwrap();
+                    let field = fields.first().unwrap();
                     if field.attr.skip_from_variant {
                         return Err(syn::Error::new(
                             field.ident.span(),
                             "cannot skip the only field in a tuple",
                         ));
                     }
-                    let expr = field.from_variant(&quote!(#variant));
+                    let expr = field.make_from_variant_expr(&quote!(#variant));
                     quote! {
                         {
                             #expr.map(#ctor)
@@ -224,14 +267,14 @@ impl VariantRepr {
                     let expr_variant = &quote!(&__array.get(__index));
                     let non_skipped_exprs = non_skipped_fields
                         .iter()
-                        .map(|f| f.from_variant(expr_variant));
+                        .map(|f| f.make_from_variant_expr(expr_variant));
 
                     quote! {
                         {
                             ::gdnative::core_types::VariantArray::from_variant(#variant)
                                 .map_err(|__err| FVE::InvalidStructRepr {
                                     expected: VariantStructRepr::Tuple,
-                                    error: Box::new(__err),
+                                    error: std::boxed::Box::new(__err),
                                 })
                                 .and_then(|__array| {
                                     let __expected = #expected_len;
@@ -245,11 +288,11 @@ impl VariantRepr {
                                             let #non_skipped_idents = #non_skipped_exprs
                                                 .map_err(|err| FVE::InvalidItem {
                                                     index: __index as usize,
-                                                    error: Box::new(err),
+                                                    error: std::boxed::Box::new(err),
                                                 })?;
                                         )*
                                         #(
-                                            let #skipped_idents = ::std::default::Default::default();
+                                            let #skipped_idents = std::default::Default::default();
                                         )*
                                         Ok(#ctor( #(#ctor_idents),* ))
                                     }
@@ -274,7 +317,7 @@ impl VariantRepr {
 
                 let name_strings: Vec<String> = non_skipped_idents
                     .iter()
-                    .map(|ident| format!("{}", ident))
+                    .map(|ident| format!("{ident}"))
                     .collect();
 
                 let name_string_literals =
@@ -283,14 +326,14 @@ impl VariantRepr {
                 let expr_variant = &quote!(&__dict.get_or_nil(&__key));
                 let exprs = non_skipped_fields
                     .iter()
-                    .map(|f| f.from_variant(expr_variant));
+                    .map(|f| f.make_from_variant_expr(expr_variant));
 
                 quote! {
                     {
                         ::gdnative::core_types::Dictionary::from_variant(#variant)
                             .map_err(|__err| FVE::InvalidStructRepr {
                                 expected: VariantStructRepr::Struct,
-                                error: Box::new(__err),
+                                error: std::boxed::Box::new(__err),
                             })
                             .and_then(|__dict| {
                                 #(
@@ -299,11 +342,11 @@ impl VariantRepr {
                                     let #non_skipped_idents = #exprs
                                         .map_err(|err| FVE::InvalidField {
                                             field_name: __field_name,
-                                            error: Box::new(err),
+                                            error: std::boxed::Box::new(err),
                                         })?;
                                 )*
                                 #(
-                                    let #skipped_idents = ::std::default::Default::default();
+                                    let #skipped_idents = std::default::Default::default();
                                 )*
                                 Ok(#ctor { #( #ctor_idents ),* })
                             })
@@ -317,7 +360,7 @@ impl VariantRepr {
 }
 
 impl Field {
-    fn to_variant(&self, trait_kind: ToVariantTrait) -> TokenStream2 {
+    fn make_to_variant_expr(&self, trait_kind: ToVariantTrait) -> TokenStream2 {
         let Field { ident, attr, .. } = self;
         if let Some(to_variant_with) = &attr.to_variant_with {
             quote!(#to_variant_with(#ident))
@@ -327,7 +370,7 @@ impl Field {
         }
     }
 
-    fn from_variant(&self, variant: &TokenStream2) -> TokenStream2 {
+    fn make_from_variant_expr(&self, variant: &TokenStream2) -> TokenStream2 {
         if let Some(from_variant_with) = &self.attr.from_variant_with {
             quote!(#from_variant_with(#variant))
         } else {
