@@ -2,7 +2,7 @@ use std::env;
 
 fn main() {
     let manifest_dir = env::var("CARGO_MANIFEST_DIR").unwrap();
-    let api_json_file = format!("{}/godot_headers/gdnative_api.json", manifest_dir);
+    let api_json_file = format!("{manifest_dir}/godot_headers/gdnative_api.json");
     let out_dir = env::var("OUT_DIR").unwrap();
 
     header_binding::generate(&manifest_dir, &out_dir);
@@ -11,7 +11,7 @@ fn main() {
 
     // Only re-run build.rs if the gdnative_api.json file has been updated.
     // Manually rebuilding the crate will ignore this.
-    println!("cargo:rerun-if-changed={}", api_json_file);
+    println!("cargo:rerun-if-changed={api_json_file}");
 }
 
 mod header_binding {
@@ -23,7 +23,7 @@ mod header_binding {
         let target = std::env::var("TARGET").unwrap();
         let platform = if target.contains("apple-darwin") {
             "macosx"
-        } else if target == "x86_64-apple-ios" {
+        } else if target == "x86_64-apple-ios" || target == "aarch64-apple-ios-sim" {
             "iphonesimulator"
         } else if target == "aarch64-apple-ios" {
             "iphoneos"
@@ -33,7 +33,7 @@ mod header_binding {
 
         // run `xcrun --sdk iphoneos --show-sdk-path`
         let output = Command::new("xcrun")
-            .args(&["--sdk", platform, "--show-sdk-path"])
+            .args(["--sdk", platform, "--show-sdk-path"])
             .output()?
             .stdout;
         let prefix = std::str::from_utf8(&output)
@@ -41,7 +41,7 @@ mod header_binding {
             .trim_end();
 
         let suffix = "usr/include";
-        let directory = format!("{}/{}", prefix, suffix);
+        let directory = format!("{prefix}/{suffix}");
 
         Ok(directory)
     }
@@ -52,11 +52,8 @@ mod header_binding {
 
         assert_eq!("android", &target_os);
 
-        let java_home =
-            std::env::var("JAVA_HOME").expect("JAVA_HOME and ANDROID_SDK_ROOT must be set");
-        let java_home = Path::new(&java_home).to_path_buf();
         let android_sdk_root =
-            std::env::var("ANDROID_SDK_ROOT").expect("JAVA_HOME and ANDROID_SDK_ROOT must be set");
+            std::env::var("ANDROID_SDK_ROOT").expect("ANDROID_SDK_ROOT must be set");
         let android_sdk_root = Path::new(&android_sdk_root).to_path_buf();
 
         // Note: cfg!(target_os) and cfg!(target_arch) refer to the target of the build script:
@@ -64,55 +61,81 @@ mod header_binding {
         // and have been erroneously used for target platforms in this library in the past. Make sure
         // to double-check them wherever they occur.
 
-        if !cfg!(target_arch = "x86_64") {
-            panic!("unsupported host architecture: build from x86_64 instead");
+        assert!(
+            cfg!(target_os = "macos") || // All macOS architectures are supported
+            cfg!(target_arch = "x86_64"),
+            "unsupported host architecture: build from x86_64 instead"
+        );
+
+        let mut android_ndk_root: Option<PathBuf> = None;
+
+        let android_ndk_folder = Path::join(&android_sdk_root, "ndk");
+        if android_ndk_folder.exists() {
+            // New NDK
+            let available_ndk_versions: Vec<_> = std::fs::read_dir(android_ndk_folder.clone())
+                .unwrap()
+                .map(|dir| dir.unwrap().path())
+                .collect();
+
+            if !available_ndk_versions.is_empty() {
+                let ndk_version = std::env::var("ANDROID_NDK_VERSION");
+
+                if let Ok(ndk_version) = ndk_version {
+                    if available_ndk_versions
+                        .iter()
+                        .map(|p| p.file_name())
+                        .any(|p| {
+                            p.is_some() && p.unwrap().to_string_lossy().eq(ndk_version.as_str())
+                        })
+                    {
+                        // Asked version is available
+                        android_ndk_root = Some(Path::join(&android_ndk_folder, ndk_version))
+                    } else {
+                        panic!(
+                            "no available android ndk versions matches {ndk_version}. Available versions: {available_ndk_versions:?}"
+                        )
+                    }
+                } else {
+                    // No NDK version chosen, chose the most recent one and issue a warning
+                    println!("cargo:warning=Multiple android ndk versions have been detected.");
+                    println!("cargo:warning=You should chose one using ANDROID_NDK_VERSION environment variable to have reproducible builds.");
+                    println!("cargo:warning=Available versions: {available_ndk_versions:?}");
+
+                    let ndk_version = available_ndk_versions
+                        .iter()
+                        .filter_map(|p| p.file_name())
+                        .filter_map(|v| semver::Version::parse(v.to_string_lossy().as_ref()).ok())
+                        .max()
+                        .unwrap();
+
+                    println!("cargo:warning=Automatically chosen version: {ndk_version} (latest)");
+
+                    android_ndk_root =
+                        Some(Path::join(&android_ndk_folder, ndk_version.to_string()));
+                }
+            }
         }
+
+        let android_ndk_bundle_folder = Path::join(&android_sdk_root, "ndk-bundle");
+        if android_ndk_root.is_none() && android_ndk_bundle_folder.exists() {
+            // Old NDK
+            android_ndk_root = Some(android_ndk_bundle_folder);
+        }
+
+        let android_ndk_root = android_ndk_root.expect("Android ndk needs to be installed");
 
         builder = builder
             .clang_arg("-I")
-            .clang_arg(Path::join(&java_home, "include/").to_string_lossy());
-
+            .clang_arg(Path::join(&android_ndk_root, "sysroot/usr/include").to_string_lossy());
         builder = builder.clang_arg("-I").clang_arg(
-            Path::join(
-                &java_home,
-                format!("include/{}/", {
-                    if cfg!(target_os = "windows") {
-                        "win32"
-                    } else if cfg!(target_os = "macos") {
-                        "darwin"
-                    } else if cfg!(target_os = "linux") {
-                        "linux"
-                    } else {
-                        panic!("unsupported host OS: build from Windows, MacOS, or Linux instead");
-                    }
-                }),
-            )
-            .to_string_lossy(),
-        );
-
-        builder = builder.clang_arg("-I").clang_arg(
-            Path::join(&android_sdk_root, "ndk-bundle/sysroot/usr/include").to_string_lossy(),
+            Path::join(&android_ndk_root, "sources/cxx-stl/llvm-libc++/include").to_string_lossy(),
         );
         builder = builder.clang_arg("-I").clang_arg(
-            Path::join(
-                &android_sdk_root,
-                "ndk-bundle/sources/cxx-stl/llvm-libc++/include",
-            )
-            .to_string_lossy(),
+            Path::join(&android_ndk_root, "sources/cxx-stl/llvm-libc++abi/include")
+                .to_string_lossy(),
         );
         builder = builder.clang_arg("-I").clang_arg(
-            Path::join(
-                &android_sdk_root,
-                "ndk-bundle/sources/cxx-stl/llvm-libc++abi/include",
-            )
-            .to_string_lossy(),
-        );
-        builder = builder.clang_arg("-I").clang_arg(
-            Path::join(
-                &android_sdk_root,
-                "ndk-bundle/sources/android/support/include",
-            )
-            .to_string_lossy(),
+            Path::join(&android_ndk_root, "sources/android/support/include").to_string_lossy(),
         );
 
         let host_tag = {
@@ -129,27 +152,27 @@ mod header_binding {
 
         builder = builder.clang_arg("-I").clang_arg(
             Path::join(
-                &android_sdk_root,
+                &android_ndk_root,
+                format!("toolchains/llvm/prebuilt/{}/sysroot/usr/include", &host_tag),
+            )
+            .to_string_lossy(),
+        );
+
+        let target_triple = match target_triple.as_str() {
+            "armv7-linux-androideabi" => "arm-linux-androideabi", // Workaround for different naming scheme with NDK
+            _ => &target_triple,
+        };
+
+        builder = builder.clang_arg("-I").clang_arg(
+            Path::join(
+                &android_ndk_root,
                 format!(
-                    "ndk-bundle/toolchains/llvm/prebuilt/{}/sysroot/usr/include",
-                    &host_tag
+                    "toolchains/llvm/prebuilt/{host}/sysroot/usr/include/{target_triple}",
+                    host = &host_tag,
                 ),
             )
             .to_string_lossy(),
         );
-
-        builder = builder.clang_arg("-I").clang_arg(
-            Path::join(
-                &android_sdk_root,
-                format!(
-            "ndk-bundle/toolchains/llvm/prebuilt/{host}/sysroot/usr/include/{target_triple}",
-            host = &host_tag,
-            target_triple = &target_triple,
-        ),
-            )
-            .to_string_lossy(),
-        );
-
         builder
     }
 
@@ -168,7 +191,7 @@ mod header_binding {
             .ignore_functions()
             .size_t_is_usize(true)
             .ctypes_prefix("libc")
-            .clang_arg(format!("-I{}/godot_headers", manifest_dir));
+            .clang_arg(format!("-I{manifest_dir}/godot_headers"));
 
         let target_os = std::env::var("CARGO_CFG_TARGET_OS").unwrap();
         let target_arch = std::env::var("CARGO_CFG_TARGET_ARCH").unwrap();
@@ -187,7 +210,11 @@ mod header_binding {
         // Workaround for https://github.com/rust-lang/rust-bindgen/issues/1211: manually set
         // target triple to `arm64-apple-ios` in place of `aarch64-apple-ios`.
         if target_arch == "aarch64" && target_os == "ios" {
-            builder = builder.clang_arg("--target=arm64-apple-ios");
+            if target_env == "sim" {
+                builder = builder.clang_arg("--target=arm64-apple-ios-sim");
+            } else {
+                builder = builder.clang_arg("--target=arm64-apple-ios");
+            }
         }
 
         // Workaround: Microsoft extensions aren't enabled by default for the `gnu` toolchain
@@ -202,13 +229,13 @@ mod header_binding {
                 "x86_64" => "_M_X64",
                 "arm" => "_M_ARM",
                 "aarch64" => "_M_ARM64",
-                _ => panic!("architecture {} not supported on Windows", target_arch),
+                _ => panic!("architecture {target_arch} not supported on Windows"),
             };
 
             builder = builder
                 .clang_arg("-fms-extensions")
                 .clang_arg("-fmsc-version=1300")
-                .clang_arg(format!("-D{}=100", arch_macro));
+                .clang_arg(format!("-D{arch_macro}=100"));
         }
 
         if target_os == "android" {
@@ -227,7 +254,6 @@ mod header_binding {
 mod api_wrapper {
     use proc_macro2::{Ident, TokenStream};
     use quote::{format_ident, quote, ToTokens};
-    use std::convert::AsRef;
     use std::fs::File;
     use std::io::Write as _;
     use std::path;
@@ -365,44 +391,44 @@ mod api_wrapper {
         }
     }
 
+    impl de::Visitor for Place<Argument> {
+        fn seq(&mut self) -> miniserde::Result<Box<dyn de::Seq + '_>> {
+            Ok(Box::new(ArgumentBuilder {
+                out: &mut self.out,
+                tuple: (None, None),
+            }))
+        }
+    }
+
+    struct ArgumentBuilder<'a> {
+        out: &'a mut Option<Argument>,
+        tuple: (Option<String>, Option<String>),
+    }
+
+    impl<'a> de::Seq for ArgumentBuilder<'a> {
+        fn element(&mut self) -> miniserde::Result<&mut dyn de::Visitor> {
+            if self.tuple.0.is_none() {
+                Ok(Deserialize::begin(&mut self.tuple.0))
+            } else if self.tuple.1.is_none() {
+                Ok(Deserialize::begin(&mut self.tuple.1))
+            } else {
+                Err(miniserde::Error)
+            }
+        }
+
+        fn finish(&mut self) -> miniserde::Result<()> {
+            if let (Some(a), Some(b)) = (self.tuple.0.take(), self.tuple.1.take()) {
+                *self.out = Some(Argument { type_: a, name: b });
+                Ok(())
+            } else {
+                Err(miniserde::Error)
+            }
+        }
+    }
+
     // Used to convert [String, String] in JSON into the Argument struct.
     impl Deserialize for Argument {
         fn begin(out: &mut Option<Self>) -> &mut dyn de::Visitor {
-            impl de::Visitor for Place<Argument> {
-                fn seq(&mut self) -> miniserde::Result<Box<dyn de::Seq + '_>> {
-                    Ok(Box::new(ArgumentBuilder {
-                        out: &mut self.out,
-                        tuple: (None, None),
-                    }))
-                }
-            }
-
-            struct ArgumentBuilder<'a> {
-                out: &'a mut Option<Argument>,
-                tuple: (Option<String>, Option<String>),
-            }
-
-            impl<'a> de::Seq for ArgumentBuilder<'a> {
-                fn element(&mut self) -> miniserde::Result<&mut dyn de::Visitor> {
-                    if self.tuple.0.is_none() {
-                        Ok(Deserialize::begin(&mut self.tuple.0))
-                    } else if self.tuple.1.is_none() {
-                        Ok(Deserialize::begin(&mut self.tuple.1))
-                    } else {
-                        Err(miniserde::Error)
-                    }
-                }
-
-                fn finish(&mut self) -> miniserde::Result<()> {
-                    if let (Some(a), Some(b)) = (self.tuple.0.take(), self.tuple.1.take()) {
-                        *self.out = Some(Argument { type_: a, name: b });
-                        Ok(())
-                    } else {
-                        Err(miniserde::Error)
-                    }
-                }
-            }
-
             Place::new(out)
         }
     }
@@ -425,7 +451,7 @@ mod api_wrapper {
             ("VIDEODECODER", 0, 1) => format_ident!("godot_gdnative_ext_videodecoder_api_struct"),
             ("NET", 3, 1) => format_ident!("godot_gdnative_ext_net_api_struct"),
             ("NET", maj, min) => format_ident!("godot_gdnative_ext_net_{}_{}_api_struct", maj, min),
-            api => panic!("Unknown API type and version: {:?}", api),
+            api => panic!("Unknown API type and version: {api:?}"),
         }
     }
 
@@ -438,7 +464,7 @@ mod api_wrapper {
             "ARVR" => format_ident!("GDNATIVE_API_TYPES_GDNATIVE_EXT_ARVR"),
             "VIDEODECODER" => format_ident!("GDNATIVE_API_TYPES_GDNATIVE_EXT_VIDEODECODER"),
             "NET" => format_ident!("GDNATIVE_API_TYPES_GDNATIVE_EXT_NET"),
-            other => panic!("Unknown API type: {:?}", other),
+            other => panic!("Unknown API type: {other:?}"),
         }
     }
 
@@ -450,17 +476,26 @@ mod api_wrapper {
         let from_json = from_json.as_ref();
         let to = to.as_ref();
         let api_json_file = std::fs::read_to_string(from_json)
-            .unwrap_or_else(|_| panic!("No such file: {:?}", from_json));
+            .unwrap_or_else(|_| panic!("No such file: {from_json:?}"));
         eprintln!("{}", &api_json_file);
         let api_root: ApiRoot = miniserde::json::from_str(&api_json_file)
-            .unwrap_or_else(|_| panic!("Could not parse ({:?}) into ApiRoot", from_json));
+            .unwrap_or_else(|_| panic!("Could not parse ({from_json:?}) into ApiRoot"));
 
-        for api in api_root.all_apis() {
-            // Currently don't support Godot 4.0
-            if api.version.major == 1 && api.version.minor == 3 {
-                panic!("GodotEngine v4.* is not yet supported. See https://github.com/godot-rust/godot-rust/issues/396");
-            }
-        }
+        // Note: this code ensured that Godot 4 (which was back then GDNative 1.3) wasn't actually used.
+        // Godot uses now GDExtension, so this no longer applies. In fact, different module APIs all have different versions.
+        // See also: https://github.com/godot-rust/godot-rust/issues/904
+
+        // Listed versions for Godot 3.5 RC:
+        //  * CORE 1.0
+        //  * CORE 1.1
+        //  * CORE 1.2
+        //  * NATIVESCRIPT 1.0
+        //  * NATIVESCRIPT 1.1
+        //  * PLUGINSCRIPT 1.0
+        //  * ARVR 1.1
+        //  * VIDEODECODER 0.1
+        //  * NET 3.1
+        //  * NET 3.2
 
         let struct_fields = godot_api_functions(&api_root);
         let impl_constructor = api_constructor(&api_root);
@@ -474,7 +509,7 @@ mod api_wrapper {
         };
         let mut wrapper_file = File::create(to.join(file_name))
             .unwrap_or_else(|_| panic!("Couldn't create output file: {:?}", to.join(file_name)));
-        write!(wrapper_file, "{}", wrapper).unwrap();
+        write!(wrapper_file, "{wrapper}").unwrap();
     }
 
     fn godot_api_functions(api: &ApiRoot) -> TokenStream {
@@ -521,10 +556,10 @@ mod api_wrapper {
             }
         }
         quote! {
-            pub unsafe fn from_raw(core_api_struct: *const godot_gdnative_core_api_struct) -> Result<Self, InitError> {
+            pub unsafe fn from_raw(core_api_struct: *const godot_gdnative_core_api_struct) -> std::result::Result<Self, InitError> {
                 #godot_apis
                 #struct_field_bindings
-                Ok(GodotApi{
+                std::result::Result::Ok(GodotApi{
                     #constructed_struct_fields
                 })
             }
@@ -553,7 +588,7 @@ mod api_wrapper {
             (1, true) => quote!(*const ),
             (1, false) => quote!(*mut ),
             (2, true) => quote!(*mut *const ),
-            _ => panic!("Unknown C type (Too many pointers?): {:?}", c_type),
+            _ => panic!("Unknown C type (Too many pointers?): {c_type:?}"),
         };
         let rust_type = match base_c_type {
             "void" => {
