@@ -8,16 +8,19 @@ const OUT_BANDWIDTH: i64 = 1000;
 
 #[derive(NativeClass)]
 #[inherit(Node)]
-pub struct Server;
+pub struct Server {
+    #[property(rpc = "master", set = "Self::set_foo")]
+    foo: i32,
+}
 
 #[methods]
 impl Server {
     fn new(_owner: &Node) -> Self {
-        Self
+        Self { foo: 0 }
     }
 
-    #[export]
-    fn _ready(&mut self, owner: &Node) {
+    #[method]
+    fn _ready(&mut self, #[base] owner: &Node) {
         let peer = NetworkedMultiplayerENet::new();
         peer.create_server(PORT, MAX_CLIENTS, IN_BANDWIDTH, OUT_BANDWIDTH)
             .unwrap();
@@ -28,8 +31,8 @@ impl Server {
         tree.set_network_peer(peer);
     }
 
-    #[export(rpc = "master")]
-    fn greet_server(&mut self, owner: &Node, msg: GodotString) {
+    #[method(rpc = "master")]
+    fn greet_server(&mut self, #[base] owner: &Node, msg: GodotString) {
         godot_print!("Client says: {}", msg);
 
         let tree = owner.get_tree().expect("could not retreive Scene Tree");
@@ -38,7 +41,13 @@ impl Server {
         owner.rpc_id(
             tree.get_rpc_sender_id(),
             "return_greeting",
-            &[Variant::from_str("hello")],
+            &[Variant::new("hello")],
         );
+    }
+
+    #[method]
+    fn set_foo(&mut self, #[base] _owner: TRef<Node>, value: i32) {
+        godot_print!("Client sets foo to: {}", value);
+        self.foo = value;
     }
 }
