@@ -1,119 +1,178 @@
 use crate::core_types::{IsEqualApprox, Quat, Vector3};
-use core::ops::Mul;
+use crate::globalscope::lerp;
 use glam::Mat3;
+use std::ops::Mul;
 
-/// A 3x3 matrix.
+/// A 3x3 matrix, typically used as an orthogonal basis for [`Transform`][crate::core_types::Transform].
+///
+/// The basis vectors are the column vectors of the matrix, while the [`elements`][Self::elements]
+/// field represents the row vectors.
+///
+/// See also [Basis](https://docs.godotengine.org/en/stable/classes/class_basis.html) in the Godot API doc.
 #[repr(C)]
 #[derive(Copy, Clone, Debug, PartialEq)]
+#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
 pub struct Basis {
+    /// Matrix rows. These are **not** the basis vectors!
+    ///
+    /// This is a transposed view for performance.<br>
+    /// To read basis vectors, see [`a()`][Self::a], [`b()`][Self::b], [`c()`][Self::c].<br>
+    /// To write them, see [`set_a()`][Self::set_a], [`set_b()`][Self::set_b], [`set_c()`][Self::set_c].
     pub elements: [Vector3; 3],
 }
 
 impl Default for Basis {
     #[inline]
     fn default() -> Self {
-        Self::identity()
+        Self::IDENTITY
     }
 }
 
 impl Basis {
-    /// The identity basis.
+    /// The identity basis. Basis vectors are unit vectors along each axis X, Y and Z.
     ///
-    /// Identical to calling [`Basis::default()`](#method.default).
-    #[inline]
-    pub const fn identity() -> Basis {
-        Self {
-            elements: [
-                Vector3::new(1.0, 0.0, 0.0),
-                Vector3::new(0.0, 1.0, 0.0),
-                Vector3::new(0.0, 0.0, 1.0),
-            ],
-        }
-    }
+    /// Equivalent to calling [`Basis::default()`].
+    pub const IDENTITY: Self = Self {
+        elements: [
+            Vector3::new(1.0, 0.0, 0.0),
+            Vector3::new(0.0, 1.0, 0.0),
+            Vector3::new(0.0, 0.0, 1.0),
+        ],
+    };
 
-    #[doc(hidden)]
-    #[inline]
-    pub fn sys(&self) -> *const sys::godot_basis {
-        unsafe { std::mem::transmute::<*const Basis, *const sys::godot_basis>(self as *const _) }
-    }
-
-    #[doc(hidden)]
-    #[inline]
-    pub fn from_sys(c: sys::godot_basis) -> Self {
-        unsafe { std::mem::transmute::<sys::godot_basis, Self>(c) }
-    }
-
-    /// The basis that will flip something along the **X Axis** when used in a transformation.
-    #[inline]
-    pub const fn flip_x() -> Basis {
-        Basis::from_diagonal(Vector3::new(-1.0, 1.0, 1.0))
-    }
+    /// The basis that will flip something along the **X axis** when used in a transformation.
+    pub const FLIP_X: Self = Self::from_diagonal(Vector3::new(-1.0, 1.0, 1.0));
 
     /// The basis that will flip something along the **Y axis** when used in a transformation.
-    #[inline]
-    pub const fn flip_y() -> Basis {
-        Basis::from_diagonal(Vector3::new(1.0, -1.0, 1.0))
-    }
+    pub const FLIP_Y: Self = Self::from_diagonal(Vector3::new(1.0, -1.0, 1.0));
 
     /// The basis that will flip something along the **Z axis** when used in a transformation.
-    #[inline]
-    pub const fn flip_z() -> Basis {
-        Basis::from_diagonal(Vector3::new(1.0, 1.0, -1.0))
-    }
+    pub const FLIP_Z: Self = Self::from_diagonal(Vector3::new(1.0, 1.0, -1.0));
 
-    /// Creates a Basis from the given [`Vector3`](./type.Vector3.html)
+    /// Constructs a basis matrix from 3 linearly independent basis vectors (matrix columns).
+    ///
+    /// This is the typical way to construct a basis. If you want to fill in the elements one-by-one,
+    /// consider using [`Self::from_rows()`].
     #[inline]
-    pub const fn from_diagonal(p_diag: Vector3) -> Basis {
-        Basis {
+    pub const fn from_basis_vectors(a: Vector3, b: Vector3, c: Vector3) -> Self {
+        Self {
             elements: [
-                Vector3::new(p_diag.x, 0.0, 0.0),
-                Vector3::new(0.0, p_diag.y, 0.0),
-                Vector3::new(0.0, 0.0, p_diag.z),
+                Vector3::new(a.x, b.x, c.x),
+                Vector3::new(a.y, b.y, c.y),
+                Vector3::new(a.z, b.z, c.z),
             ],
         }
     }
 
-    /// Creates a `Basis` from 3 [`Vector3`](./type.Vector3.html)s
+    /// Creates a basis from 3 row vectors. These are **not** basis vectors.
+    ///
+    /// This constructor is mostly useful if you want to write matrix elements
+    /// in matrix syntax:
+    /// ```
+    /// # use gdnative::core_types::{Vector3, Basis};
+    /// # let a = Vector3::RIGHT;
+    /// # let b = Vector3::UP;
+    /// # let c = Vector3::BACK;
+    /// let basis = Basis::from_rows(
+    ///     Vector3::new(a.x, b.x, c.x),
+    ///     Vector3::new(a.y, b.y, c.y),
+    ///     Vector3::new(a.z, b.z, c.z),
+    /// );
+    /// ```
+    /// The vectors `a`, `b` and `c` are the basis vectors.<br>
+    /// In this particular case, you could also use [`Self::from_basis_vectors(a, b, c)`][Self::from_basis_vectors()] instead.
     #[inline]
-    pub const fn from_elements(elements: [Vector3; 3]) -> Self {
-        Self { elements }
+    pub const fn from_rows(
+        x_components: Vector3,
+        y_components: Vector3,
+        z_components: Vector3,
+    ) -> Self {
+        Self {
+            elements: [x_components, y_components, z_components],
+        }
     }
 
-    /// Creates a rotation matrix.
+    /// Creates a diagonal matrix from the given vector.
+    ///
+    /// Can be used as a basis for a scaling transform.
+    /// Each component of `scale` represents the scale factor in the corresponding direction.
     #[inline]
-    pub fn from_euler(euler: Vector3) -> Self {
-        let mut b = Basis::default();
-        b.set_euler_yxz(&euler);
+    pub const fn from_diagonal(scale: Vector3) -> Self {
+        Self {
+            elements: [
+                Vector3::new(scale.x, 0.0, 0.0),
+                Vector3::new(0.0, scale.y, 0.0),
+                Vector3::new(0.0, 0.0, scale.z),
+            ],
+        }
+    }
+
+    /// Creates a rotation matrix from Euler angles.
+    ///
+    /// The angle vector has XYZ components. However, the angles are applied in YXZ convention:
+    /// first **Z**, then **X**, and **Y** last.
+    #[inline]
+    pub fn from_euler(euler_angles: Vector3) -> Self {
+        let mut b = Self::default();
+        b.set_euler_yxz(euler_angles);
         b
+    }
+
+    #[inline]
+    fn set_euler_yxz(&mut self, euler: Vector3) {
+        let c = euler.x.cos();
+        let s = euler.x.sin();
+        let xmat = Self::from_rows(
+            Vector3::new(1.0, 0.0, 0.0),
+            Vector3::new(0.0, c, -s),
+            Vector3::new(0.0, s, c),
+        );
+        let c = euler.y.cos();
+        let s = euler.y.sin();
+        let ymat = Self::from_rows(
+            Vector3::new(c, 0.0, s),
+            Vector3::new(0.0, 1.0, 0.0),
+            Vector3::new(-s, 0.0, c),
+        );
+
+        let c = euler.z.cos();
+        let s = euler.z.sin();
+        let zmat = Self::from_rows(
+            Vector3::new(c, -s, 0.0),
+            Vector3::new(s, c, 0.0),
+            Vector3::new(0.0, 0.0, 1.0),
+        );
+
+        *self = ymat * xmat * zmat;
     }
 
     /// Constructs a pure rotation basis matrix from the given quaternion.
     #[inline]
     pub fn from_quat(quat: Quat) -> Self {
         let basis = Mat3::from_quat(quat.glam()).to_cols_array_2d();
-        let basis = [
+
+        Self::from_rows(
             Vector3::new(basis[0][0], basis[1][0], basis[2][0]),
             Vector3::new(basis[0][1], basis[1][1], basis[2][1]),
             Vector3::new(basis[0][2], basis[1][2], basis[2][2]),
-        ];
-        Basis::from_elements(basis)
+        )
     }
 
     /// Rotation matrix from axis and angle.
     ///
-    /// See <https://en.wikipedia.org/wiki/Rotation_matrix#Rotation_matrix_from_axis_angle>
+    /// See <https://en.wikipedia.org/wiki/Rotation_matrix#Rotation_matrix_from_axis_and_angle>
     ///
     /// # Panics
     ///
     /// If `axis` is not normalized.
     #[inline]
-    pub fn from_axis_angle(axis: &Vector3, phi: f32) -> Self {
+    pub fn from_axis_angle(axis: Vector3, phi: f32) -> Self {
         assert!(
             axis.length().is_equal_approx(1.0),
             "The axis Vector3 must be normalized."
         );
 
-        let mut basis = Basis::default();
+        let mut basis = Self::default();
         let [x, y, z] = &mut basis.elements;
 
         let axis_sq = Vector3::new(axis.x * axis.x, axis.y * axis.y, axis.z * axis.z);
@@ -150,7 +209,7 @@ impl Basis {
     ///
     /// If the determinant of `self` is zero.
     #[inline]
-    pub fn invert(&mut self) {
+    fn invert(&mut self) {
         let [x, y, z] = self.elements;
 
         let co = [
@@ -164,13 +223,13 @@ impl Basis {
 
         let s: f32 = 1.0 / det;
 
-        self.set_x(Vector3::new(co[0] * s, co[1] * s, co[2] * s));
-        self.set_y(Vector3::new(
+        self.set_a(Vector3::new(co[0] * s, co[1] * s, co[2] * s));
+        self.set_b(Vector3::new(
             (x.z * z.y - x.y * z.z) * s,
             (x.x * z.z - x.z * z.x) * s,
             (x.y * z.x - x.x * z.y) * s,
         ));
-        self.set_z(Vector3::new(
+        self.set_c(Vector3::new(
             (x.y * y.z - x.z * y.y) * s,
             (x.z * y.x - x.x * y.z) * s,
             (x.x * y.y - x.y * y.x) * s,
@@ -183,14 +242,49 @@ impl Basis {
     ///
     /// If the determinant of `self` is zero.
     #[inline]
-    pub fn inverted(mut self) -> Basis {
-        self.invert();
-        self
+    pub fn inverse(&self) -> Self {
+        let mut copy = *self;
+        copy.invert();
+        copy
+    }
+
+    #[inline]
+    #[deprecated = "Use `inverse` instead."]
+    pub fn inverted(&self) -> Self {
+        self.inverse()
+    }
+
+    /// Returns linear interpolation on a sphere between two basis by weight amount (on the range of 0.0 to 1.0).
+    #[inline]
+    pub fn slerp(&self, other: &Basis, weight: f32) -> Self {
+        let from = self.to_quat();
+        let to = other.to_quat();
+        let mut result = Basis::from_quat(from.slerp(to, weight));
+
+        for i in 0..3 {
+            result.elements[i] *= lerp(
+                self.elements[i].length()..=other.elements[i].length(),
+                weight,
+            );
+        }
+
+        result
+    }
+
+    /// Returns linear interpolation between two basis by weight amount (on the range of 0.0 to 1.0).
+    #[inline]
+    pub fn lerp(&self, other: &Basis, weight: f32) -> Self {
+        // this is how godot is doing it at https://github.com/godotengine/godot/blob/master/core/math/basis.cpp#L964
+        // but Godot engine output for me differs than godot-rust
+        let a = self.elements[0].linear_interpolate(other.elements[0], weight);
+        let b = self.elements[1].linear_interpolate(other.elements[1], weight);
+        let c = self.elements[2].linear_interpolate(other.elements[2], weight);
+        Basis::from_rows(a, b, c)
     }
 
     /// Transposes the matrix.
     #[inline]
-    pub fn transpose(&mut self) {
+    fn transpose(&mut self) {
         std::mem::swap(&mut self.elements[0].y, &mut self.elements[1].x);
         std::mem::swap(&mut self.elements[0].z, &mut self.elements[2].x);
         std::mem::swap(&mut self.elements[1].z, &mut self.elements[2].y);
@@ -198,9 +292,10 @@ impl Basis {
 
     /// Returns the transposed version of the matrix.
     #[inline]
-    pub fn transposed(mut self) -> Basis {
-        self.transpose();
-        self
+    pub fn transposed(&self) -> Self {
+        let mut copy = *self;
+        copy.transpose();
+        copy
     }
 
     /// Returns the determinant of the matrix.
@@ -221,16 +316,16 @@ impl Basis {
     ///
     /// If the determinant of `self` is zero.
     #[inline]
-    pub fn orthonormalize(&mut self) {
+    fn orthonormalize(&mut self) {
         assert!(
             !self.determinant().is_equal_approx(0.0),
             "Determinant should not be zero."
         );
 
         // Gram-Schmidt Process
-        let mut x = self.x();
-        let mut y = self.y();
-        let mut z = self.z();
+        let mut x = self.a();
+        let mut y = self.b();
+        let mut z = self.c();
 
         x = x.normalized();
         y = y - x * (x.dot(y));
@@ -238,18 +333,37 @@ impl Basis {
         z = z - x * (x.dot(z)) - y * (y.dot(z));
         z = z.normalized();
 
-        self.set_x(x);
-        self.set_y(y);
-        self.set_z(z);
+        self.set_a(x);
+        self.set_b(y);
+        self.set_c(z);
     }
 
-    /// Returns an orthonormalized version of the matrix.
-    ///
-    /// See [`Basis::orthonormalize()`](#method.orthonormalize)
     #[inline]
-    pub fn orthonormalized(mut self) -> Basis {
+    fn is_orthogonal(&self) -> bool {
+        let m = (*self) * self.transposed();
+        m.is_equal_approx(&Self::IDENTITY)
+    }
+
+    #[inline]
+    fn orthogonalize(&mut self) {
+        let scale = self.scale();
         self.orthonormalize();
-        self
+        self.scale_local(scale);
+    }
+
+    #[inline]
+    pub fn orthogonalized(&self) -> Self {
+        let mut copy = *self;
+        copy.orthogonalize();
+        copy
+    }
+
+    /// Returns an orthonormalized version of the matrix: 3 orthogonal basis vectors of unit length.
+    #[inline]
+    pub fn orthonormalized(&self) -> Self {
+        let mut copy = *self;
+        copy.orthonormalize();
+        copy
     }
 
     /// Returns `true` if `self` and `other` are approximately equal.
@@ -260,46 +374,82 @@ impl Basis {
             && self.elements[2].is_equal_approx(other.elements[2])
     }
 
+    /// Multiplies the matrix from left with the rotation matrix: M -> R·M
+    ///
+    /// The main use of `Basis` is as a `Transform.basis`, which is used as the transformation matrix
+    /// of the 3D object. `rotated()` here refers to rotation of the object (which is `R * self`), not the matrix itself.
     #[inline]
-    fn is_orthogonal(&self) -> bool {
-        let identity = Self::identity();
-        let m = (*self) * self.transposed();
-        m.is_equal_approx(&identity)
+    pub fn rotated(&self, axis: Vector3, phi: f32) -> Self {
+        let mut copy = *self;
+        copy.rotate(axis, phi);
+        copy
     }
 
+    /// Rotates the matrix.
+    ///
+    /// If object rotation is needed, see [`Basis::rotated()`]
+    #[inline]
+    fn rotate(&mut self, axis: Vector3, phi: f32) {
+        let rot = Self::from_axis_angle(axis, phi);
+        *self = rot * *self;
+    }
+
+    /// Returns true if this basis represents a rotation matrix (orthogonal and no scaling).
     #[inline]
     fn is_rotation(&self) -> bool {
         let det = self.determinant();
         det.is_equal_approx(1.0) && self.is_orthogonal()
     }
 
-    /// Multiplies the matrix from left by the rotation matrix: M -> R.M
-    ///
-    /// The main use of `Basis` is as a `Transform.basis`, which is used as the transformation matrix
-    /// of the 3D object. `rotated()` here refers to rotation of the object (which is `R * self`), not the matrix itself.
+    /// Returns the scale of the matrix.
     #[inline]
-    pub fn rotated(self, axis: Vector3, phi: f32) -> Basis {
-        let rot = Basis::from_axis_angle(&axis, phi);
-        rot * self
+    pub fn scale(&self) -> Vector3 {
+        let det = self.determinant();
+        let det_sign = if det < 0.0 { -1.0 } else { 1.0 };
+
+        Vector3::new(self.a().length(), self.b().length(), self.c().length()) * det_sign
     }
 
-    /// Rotates the matrix.
-    ///
-    /// If object rotation is needed, see [`Basis::rotated()`](#method.rotated)
+    /// Introduce an additional scaling specified by the given 3D scaling factor.
     #[inline]
-    pub fn rotate(&mut self, axis: Vector3, phi: f32) {
-        *self = self.rotated(axis, phi);
+    pub fn scaled(&self, scale: Vector3) -> Self {
+        let mut copy = *self;
+        copy.scale_self(scale);
+        copy
     }
 
-    /// Converts matrix into a [Quaternion](./type.Quat.html)
+    /// In-place basis scaling in object-local coordinate system
+    #[inline]
+    fn scale_local(&mut self, scale: Vector3) {
+        *self = self.scaled_local(scale);
+    }
+
+    /// Returns basis scaled in object-local coordinate system
+    #[inline]
+    fn scaled_local(&self, scale: Vector3) -> Self {
+        (*self) * Basis::from_diagonal(scale)
+    }
+
+    /// Multiplies the matrix from left with the scaling matrix: M -> S·M
     ///
-    /// Quaternions are frequently used in 3D graphics, because they enable easy and cheap interpolation. However, they are less human-readable. For Euler angles, see [`Basis::to_euler()`](#method.to_euler).
+    /// See the comment for [Basis::rotated](#method.rotated) for further explanation.
+    #[inline]
+    fn scale_self(&mut self, s: Vector3) {
+        self.elements[0] *= s.x;
+        self.elements[1] *= s.y;
+        self.elements[2] *= s.z;
+    }
+
+    /// Converts matrix into a quaternion.
+    ///
+    /// Quaternions are frequently used in 3D graphics, because they enable easy and cheap interpolation.
+    /// However, they are less human-readable. For Euler angles, see [`Basis::to_euler()`].
     ///
     /// # Panics
     ///
     /// If `self` is not normalized.
     #[inline]
-    pub fn to_quat(self) -> Quat {
+    pub fn to_quat(&self) -> Quat {
         // Assumes that the matrix can be decomposed into a proper rotation and scaling matrix as M = R.S,
         // and returns the Euler angles corresponding to the rotation part, complementing get_scale().
         // See the comment in get_scale() for further information.
@@ -307,7 +457,7 @@ impl Basis {
         let det = matrix.determinant();
         if det < 0.0 {
             // Ensure that the determinant is 1, such that result is a proper rotation matrix which can be represented by Euler angles.
-            matrix.scale(&Vector3::new(-1.0, -1.0, -1.0));
+            matrix.scale_self(Vector3::new(-1.0, -1.0, -1.0));
         }
 
         assert!(matrix.is_rotation(), "Basis must be normalized in order to be casted to a Quaternion. Use to_quat() or call orthonormalized() instead.");
@@ -359,35 +509,6 @@ impl Basis {
         Quat::new(a, b, c, r)
     }
 
-    /// Returns the scale of the matrix.
-    #[inline]
-    pub fn to_scale(self) -> Vector3 {
-        let det = self.determinant();
-        let det_sign = if det < 0.0 { -1.0 } else { 1.0 };
-        Vector3::new(
-            Vector3::new(self.elements[0].x, self.elements[1].x, self.elements[2].x).length(),
-            Vector3::new(self.elements[0].y, self.elements[1].y, self.elements[2].y).length(),
-            Vector3::new(self.elements[0].z, self.elements[1].z, self.elements[2].z).length(),
-        ) * det_sign
-    }
-
-    /// Multiplies the matrix from left by the scaling matrix: M -> S.M
-    ///
-    /// See the comment for [Basis::rotated](#method.rotated) for further explanation.
-    #[inline]
-    fn scale(&mut self, s: &Vector3) {
-        self.elements[0] *= s.x;
-        self.elements[1] *= s.y;
-        self.elements[2] *= s.z;
-    }
-
-    /// Introduce an additional scaling specified by the given 3D scaling factor.
-    #[inline]
-    pub fn scaled(mut self, scale: &Vector3) -> Basis {
-        self.scale(scale);
-        self
-    }
-
     /// Returns the `Basis`’s rotation in the form of Euler angles.
     ///
     /// In the YXZ convention: first **Z**, then **X**, and **Y** last.
@@ -396,7 +517,7 @@ impl Basis {
     ///
     /// See [`Basis::to_quat`](#method.to_quat) if you need a quaternion instead.
     #[inline]
-    pub fn to_euler(self) -> Vector3 {
+    pub fn to_euler(&self) -> Vector3 {
         let mut euler = Vector3::ZERO;
 
         let m12 = self.elements[1].z;
@@ -434,34 +555,6 @@ impl Basis {
         euler
     }
 
-    #[inline]
-    fn set_euler_yxz(&mut self, euler: &Vector3) {
-        let c = euler.x.cos();
-        let s = euler.x.sin();
-        let xmat = Basis::from_elements([
-            Vector3::new(1.0, 0.0, 0.0),
-            Vector3::new(0.0, c, -s),
-            Vector3::new(0.0, s, c),
-        ]);
-        let c = euler.y.cos();
-        let s = euler.y.sin();
-        let ymat = Basis::from_elements([
-            Vector3::new(c, 0.0, s),
-            Vector3::new(0.0, 1.0, 0.0),
-            Vector3::new(-s, 0.0, c),
-        ]);
-
-        let c = euler.z.cos();
-        let s = euler.z.sin();
-        let zmat = Basis::from_elements([
-            Vector3::new(c, -s, 0.0),
-            Vector3::new(s, c, 0.0),
-            Vector3::new(0.0, 0.0, 1.0),
-        ]);
-
-        *self = ymat * xmat * zmat;
-    }
-
     /// Returns a vector transformed (multiplied) by the matrix.
     #[inline]
     pub fn xform(&self, v: Vector3) -> Vector3 {
@@ -477,97 +570,79 @@ impl Basis {
     /// Note: This results in a multiplication by the inverse of the matrix only if it represents a rotation-reflection.
     #[inline]
     pub fn xform_inv(&self, v: Vector3) -> Vector3 {
-        Vector3::new(
-            (self.elements[0].x * v.x) + (self.elements[1].x * v.y) + (self.elements[2].x * v.z),
-            (self.elements[0].y * v.x) + (self.elements[1].y * v.y) + (self.elements[2].y * v.z),
-            (self.elements[0].z * v.x) + (self.elements[1].z * v.y) + (self.elements[2].z * v.z),
-        )
+        Vector3::new(self.a().dot(v), self.b().dot(v), self.c().dot(v))
     }
 
-    /*
-    /// Creates a `Basis` from the rotation and scaling of the provided transform.
+    /// Transposed dot product with the **X basis vector** of the matrix.
     #[inline]
-    pub fn from_transform(transform: &default::Transform3D<f32>) -> Basis {
-        Self::from_typed_transform::<UnknownUnit, UnknownUnit>(transform)
+    pub(crate) fn tdotx(&self, v: Vector3) -> f32 {
+        self.a().dot(v)
     }
 
-    /// Creates a `Basis` from the rotation and scaling of the provided transform, in `Dst` space.
+    /// Transposed dot product with the **Y basis vector** of the matrix.
     #[inline]
-    pub fn from_typed_transform<Src, Dst>(transform: &Transform3D<f32, Src, Dst>) -> Basis {
-        Basis {
-            elements: [
-                transform
-                    .transform_vector3d(Vector3D::<_, Src>::new(1.0, 0.0, 0.0))
-                    .to_untyped(),
-                transform
-                    .transform_vector3d(Vector3D::<_, Src>::new(0.0, 1.0, 0.0))
-                    .to_untyped(),
-                transform
-                    .transform_vector3d(Vector3D::<_, Src>::new(0.0, 0.0, 1.0))
-                    .to_untyped(),
-            ],
-        }
-    }
-    */
-
-    /// Transposed dot product with the **X Axis** of the matrix.
-    #[inline]
-    pub fn tdotx(&self, v: Vector3) -> f32 {
-        self.elements[0].x * v.x + self.elements[1].x * v.y + self.elements[2].x * v.z
+    pub(crate) fn tdoty(&self, v: Vector3) -> f32 {
+        self.b().dot(v)
     }
 
-    /// Transposed dot product with the **Y axis** of the matrix.
+    /// Transposed dot product with the **Z basis vector** of the matrix.
     #[inline]
-    pub fn tdoty(&self, v: Vector3) -> f32 {
-        self.elements[0].y * v.x + self.elements[1].y * v.y + self.elements[2].y * v.z
+    pub(crate) fn tdotz(&self, v: Vector3) -> f32 {
+        self.c().dot(v)
     }
 
-    /// Transposed dot product with the **Z axis** of the matrix.
+    /// Get the **1st basis vector** (first column vector of the matrix).
     #[inline]
-    pub fn tdotz(&self, v: Vector3) -> f32 {
-        self.elements[0].z * v.x + self.elements[1].z * v.y + self.elements[2].z * v.z
-    }
-
-    /// Get the **X Axis** of the matrix
-    #[inline]
-    pub fn x(&self) -> Vector3 {
+    pub fn a(&self) -> Vector3 {
         Vector3::new(self.elements[0].x, self.elements[1].x, self.elements[2].x)
     }
 
-    /// Set the **X Axis** of the matrix
+    /// Set the **1st basis vector** (first column vector of the matrix).
     #[inline]
-    pub fn set_x(&mut self, v: Vector3) {
+    pub fn set_a(&mut self, v: Vector3) {
         self.elements[0].x = v.x;
         self.elements[1].x = v.y;
         self.elements[2].x = v.z;
     }
 
-    /// Get the **Y axis** of the matrix
+    /// Get the **2nd basis vector** (second column vector of the matrix).
     #[inline]
-    pub fn y(&self) -> Vector3 {
+    pub fn b(&self) -> Vector3 {
         Vector3::new(self.elements[0].y, self.elements[1].y, self.elements[2].y)
     }
 
-    /// Set the **Y axis** of the matrix
+    /// Set the **2nd basis vector** (second column vector of the matrix).
     #[inline]
-    pub fn set_y(&mut self, v: Vector3) {
+    pub fn set_b(&mut self, v: Vector3) {
         self.elements[0].y = v.x;
         self.elements[1].y = v.y;
         self.elements[2].y = v.z;
     }
 
-    /// Get the **Z axis** of the matrix
+    /// Get the **3rd basis vector** (third column vector of the matrix).
     #[inline]
-    pub fn z(&self) -> Vector3 {
+    pub fn c(&self) -> Vector3 {
         Vector3::new(self.elements[0].z, self.elements[1].z, self.elements[2].z)
     }
 
-    /// Set the **Z axis** of the matrix
+    /// Set the **3rd basis vector** (third column vector of the matrix).
     #[inline]
-    pub fn set_z(&mut self, v: Vector3) {
+    pub fn set_c(&mut self, v: Vector3) {
         self.elements[0].z = v.x;
         self.elements[1].z = v.y;
         self.elements[2].z = v.z;
+    }
+
+    #[doc(hidden)]
+    #[inline]
+    pub fn sys(&self) -> *const sys::godot_basis {
+        unsafe { std::mem::transmute::<*const Basis, *const sys::godot_basis>(self as *const _) }
+    }
+
+    #[doc(hidden)]
+    #[inline]
+    pub fn from_sys(c: sys::godot_basis) -> Self {
+        unsafe { std::mem::transmute::<sys::godot_basis, Self>(c) }
     }
 }
 
@@ -576,7 +651,7 @@ impl Mul<Basis> for Basis {
 
     #[inline]
     fn mul(self, rhs: Self) -> Self {
-        Basis::from_elements([
+        Basis::from_rows(
             Vector3::new(
                 rhs.tdotx(self.elements[0]),
                 rhs.tdoty(self.elements[0]),
@@ -592,7 +667,7 @@ impl Mul<Basis> for Basis {
                 rhs.tdoty(self.elements[2]),
                 rhs.tdotz(self.elements[2]),
             ),
-        ])
+        )
     }
 }
 
@@ -622,9 +697,9 @@ mod tests {
 
         let vector = Vector3::new(4.0, 5.0, 6.0);
 
-        assert!((basis.tdotx(vector) - 32.0).abs() < std::f32::EPSILON);
-        assert!((basis.tdoty(vector) - 47.0).abs() < std::f32::EPSILON);
-        assert!((basis.tdotz(vector) - 62.0).abs() < std::f32::EPSILON);
+        assert!((basis.tdotx(vector) - 32.0).abs() < f32::EPSILON);
+        assert!((basis.tdoty(vector) - 47.0).abs() < f32::EPSILON);
+        assert!((basis.tdotz(vector) - 62.0).abs() < f32::EPSILON);
     }
 
     #[test]
@@ -637,9 +712,9 @@ mod tests {
             ],
         };
 
-        assert!(basis.x() == Vector3::new(1.0, 4.0, 7.0));
-        assert!(basis.y() == Vector3::new(2.0, 5.0, 8.0));
-        assert!(basis.z() == Vector3::new(3.0, 6.0, 9.0));
+        assert!(basis.a() == Vector3::new(1.0, 4.0, 7.0));
+        assert!(basis.b() == Vector3::new(2.0, 5.0, 8.0));
+        assert!(basis.c() == Vector3::new(3.0, 6.0, 9.0));
     }
 
     #[test]
@@ -648,9 +723,9 @@ mod tests {
             elements: [Vector3::ZERO, Vector3::ZERO, Vector3::ZERO],
         };
 
-        basis.set_x(Vector3::new(1.0, 4.0, 7.0));
-        basis.set_y(Vector3::new(2.0, 5.0, 8.0));
-        basis.set_z(Vector3::new(3.0, 6.0, 9.0));
+        basis.set_a(Vector3::new(1.0, 4.0, 7.0));
+        basis.set_b(Vector3::new(2.0, 5.0, 8.0));
+        basis.set_c(Vector3::new(3.0, 6.0, 9.0));
 
         assert!(basis.elements[0] == Vector3::new(1.0, 2.0, 3.0));
         assert!(basis.elements[1] == Vector3::new(4.0, 5.0, 6.0));
@@ -686,11 +761,11 @@ mod tests {
     fn orthonormalized() {
         let (b, _bn) = test_inputs();
 
-        let expected = Basis::from_elements([
+        let expected = Basis::from_rows(
             Vector3::new(0.077431, -0.165055, 0.98324),
             Vector3::new(-0.288147, 0.94041, 0.180557),
             Vector3::new(-0.95445, -0.297299, 0.025257),
-        ]);
+        );
         assert!(expected.is_equal_approx(&b.orthonormalized()));
     }
 
@@ -698,12 +773,12 @@ mod tests {
     fn scaled() {
         let (b, _bn) = test_inputs();
 
-        let expected = Basis::from_elements([
+        let expected = Basis::from_rows(
             Vector3::new(0.052484, -0.111876, 0.666453),
             Vector3::new(0.012407, -0.040492, -0.007774),
             Vector3::new(-0.682131, -0.212475, 0.018051),
-        ]);
-        assert!(expected.is_equal_approx(&b.scaled(&Vector3::new(0.677813, -0.043058, 0.714685))));
+        );
+        assert!(expected.is_equal_approx(&b.scaled(Vector3::new(0.677813, -0.043058, 0.714685))));
     }
 
     #[test]
@@ -711,11 +786,11 @@ mod tests {
         let (b, _bn) = test_inputs();
 
         let r = Vector3::new(-50.167156, 60.67781, -70.04305).normalized();
-        let expected = Basis::from_elements([
+        let expected = Basis::from_rows(
             Vector3::new(-0.676245, 0.113805, 0.727833),
             Vector3::new(-0.467094, 0.697765, -0.54309),
             Vector3::new(-0.569663, -0.707229, -0.418703),
-        ]);
+        );
         assert!(expected.is_equal_approx(&b.rotated(r, 1.0)));
     }
 
@@ -730,7 +805,7 @@ mod tests {
     fn scale() {
         let (b, _bn) = test_inputs();
 
-        assert!(Vector3::new(1.0, 1.0, 1.0).is_equal_approx(b.to_scale()));
+        assert!(Vector3::new(1.0, 1.0, 1.0).is_equal_approx(b.scale()));
     }
 
     #[test]
@@ -742,11 +817,11 @@ mod tests {
     #[test]
     fn transposed() {
         let (b, _bn) = test_inputs();
-        let expected = Basis::from_elements([
+        let expected = Basis::from_rows(
             Vector3::new(0.077431, -0.288147, -0.95445),
             Vector3::new(-0.165055, 0.94041, -0.297299),
             Vector3::new(0.98324, 0.180557, 0.025257),
-        ]);
+        );
         assert!(expected.is_equal_approx(&b.transposed()));
     }
 
@@ -770,11 +845,11 @@ mod tests {
     fn inverse() {
         let (b, _bn) = test_inputs();
 
-        let expected = Basis::from_elements([
+        let expected = Basis::from_rows(
             Vector3::new(0.077431, -0.288147, -0.95445),
             Vector3::new(-0.165055, 0.94041, -0.297299),
             Vector3::new(0.98324, 0.180557, 0.025257),
-        ]);
-        assert!(expected.is_equal_approx(&b.inverted()));
+        );
+        assert!(expected.is_equal_approx(&b.inverse()));
     }
 }
