@@ -1,9 +1,9 @@
+use crate::core_types::Variant;
+use crate::object::NewRef;
 use crate::private::get_api;
 use crate::sys;
-use crate::NewRef;
-
 use std::cmp::Ordering;
-use std::convert::TryFrom;
+
 use std::ffi::CStr;
 use std::fmt;
 use std::mem::forget;
@@ -201,15 +201,49 @@ impl GodotString {
         unsafe { (get_api().godot_string_find_last)(&self.0, what.0) }
     }
 
-    /// Returns the internal ffi representation of the string and consumes
-    /// the rust object without running the destructor.
+    /// Formats the string by replacing all occurrences of a key in the string with the
+    /// corresponding value. The method can handle arrays or dictionaries for the key/value pairs.
     ///
-    /// This should be only used when certain that the receiving side is
-    /// responsible for running the destructor for the object, otherwise
-    /// it is leaked.
+    /// Arrays can be used as key, index, or mixed style (see below examples). Order only matters
+    /// when the index or mixed style of Array is used.
+    ///
+    /// # Examples
+    ///
+    /// ```no_run
+    /// # use gdnative::prelude::*;
+    /// // Array values, index style
+    /// let template = GodotString::from("{0} {1}");
+    /// let data = VariantArray::new();
+    /// data.push("foo");
+    /// data.push("bar");
+    ///
+    /// let formatted = template.format(&data.into_shared().to_variant());
+    /// godot_print!("{}", formatted); // "foo bar"
+    /// ```
+    ///
+    /// ```no_run
+    /// # use gdnative::prelude::*;
+    /// // Dictionary values
+    /// let template = GodotString::from("foo {bar}");
+    /// let data = Dictionary::new();
+    /// data.insert("bar", "baz");
+    ///
+    /// let formatted = template.format(&data.into_shared().to_variant());
+    /// godot_print!("{}", formatted); // "foo baz"
+    /// ```
+    #[inline]
+    pub fn format(&self, values: &Variant) -> Self {
+        Self(unsafe { (get_api().godot_string_format)(&self.0, &values.0) })
+    }
+
+    /// Returns the internal FFI representation of the string and consumes
+    /// the Rust object without running the destructor.
+    ///
+    /// The returned object has no `Drop` implementation. The caller is
+    /// responsible of manually ensuring destruction.
     #[doc(hidden)]
     #[inline]
-    pub fn forget(self) -> sys::godot_string {
+    pub fn leak(self) -> sys::godot_string {
         let v = self.0;
         forget(self);
         v
@@ -249,11 +283,9 @@ impl GodotString {
     pub fn clone_from_sys(sys: sys::godot_string) -> Self {
         let sys_string = GodotString(sys);
         let this = sys_string.clone();
-        sys_string.forget();
+        sys_string.leak();
         this
     }
-
-    // TODO: many missing methods.
 }
 
 impl Clone for GodotString {
@@ -267,6 +299,7 @@ impl_basic_traits_as_sys!(
     for GodotString as godot_string {
         Drop => godot_string_destroy;
         Eq => godot_string_operator_equal;
+        Ord => godot_string_operator_less;
         Default => godot_string_new;
         NewRef => godot_string_new_copy;
     }
@@ -348,26 +381,6 @@ where
     }
 }
 
-impl PartialOrd for GodotString {
-    #[inline]
-    fn partial_cmp(&self, other: &GodotString) -> Option<Ordering> {
-        Some(self.cmp(other))
-    }
-}
-
-impl Ord for GodotString {
-    #[inline]
-    fn cmp(&self, other: &GodotString) -> Ordering {
-        if self == other {
-            Ordering::Equal
-        } else if unsafe { (get_api().godot_string_operator_less)(&self.0, &other.0) } {
-            Ordering::Less
-        } else {
-            Ordering::Greater
-        }
-    }
-}
-
 /// Type representing a character in Godot's native encoding. Can be converted to and
 /// from `char`. Depending on the platform, this might not always be able to represent
 /// a full code point.
@@ -377,6 +390,7 @@ pub struct GodotChar(libc::wchar_t);
 
 /// Error indicating that a `GodotChar` cannot be converted to a `char`.
 #[derive(Debug)]
+#[non_exhaustive]
 pub enum GodotCharError {
     /// The character cannot be represented as a Unicode code point.
     InvalidCodePoint,
@@ -396,6 +410,7 @@ impl TryFrom<GodotChar> for char {
             1 => std::char::from_u32(c.0 as u32).ok_or(GodotCharError::IncompleteSequence),
             4 => std::char::from_u32(c.0 as u32).ok_or(GodotCharError::InvalidCodePoint),
             2 => {
+                #[allow(clippy::unnecessary_cast)] // type wchar_t may be platform-dependent
                 let mut iter = std::char::decode_utf16(std::iter::once(c.0 as u16));
                 let c = iter
                     .next()
@@ -439,7 +454,7 @@ impl Index<usize> for GodotString {
     }
 }
 
-// TODO: Is it useful to expose this type?
+// TODO(#993): Is it useful to expose this type?
 // Could just make it an internal detail of how to convert to a rust string.
 #[doc(hidden)]
 pub struct Utf8String(pub(crate) sys::godot_char_string);
@@ -508,6 +523,12 @@ impl fmt::Debug for Utf8String {
     }
 }
 
+/// Interned string.
+///
+/// Like [`GodotString`], but unique: two `StringName`s with the same string value share the same
+/// internal object. Just like the `GodotString` struct, this type is immutable.
+///
+/// Use [`Self::from_godot_string()`] and [`Self::to_godot_string()`] for conversions.
 pub struct StringName(pub(crate) sys::godot_string_name);
 
 impl StringName {
@@ -545,13 +566,8 @@ impl StringName {
     }
 
     #[inline]
-    pub fn get_name(&self) -> GodotString {
+    pub fn to_godot_string(&self) -> GodotString {
         unsafe { GodotString((get_api().godot_string_name_get_name)(&self.0)) }
-    }
-
-    #[inline]
-    pub fn operator_less(&self, s: &StringName) -> bool {
-        unsafe { (get_api().godot_string_name_operator_less)(&self.0, &s.0) }
     }
 
     #[doc(hidden)]
@@ -576,29 +592,43 @@ impl StringName {
 impl_basic_traits_as_sys! {
     for StringName as godot_string_name {
         Drop => godot_string_name_destroy;
-        Eq => godot_string_name_operator_equal;
+
+        // Note: Godot's equal/less implementations contained a bug until Godot 3.5, see https://github.com/godot-rust/godot-rust/pull/912
+        // Thus explicit impl as a workaround for now
+        // Eq => godot_string_name_operator_equal;
+        // Ord => godot_string_name_operator_less;
+    }
+}
+
+impl PartialEq for StringName {
+    #[inline]
+    fn eq(&self, other: &Self) -> bool {
+        // Slow but correct -- see comment above
+        self.to_godot_string() == other.to_godot_string()
+    }
+}
+
+impl Eq for StringName {}
+
+impl PartialOrd for StringName {
+    #[inline]
+    fn partial_cmp(&self, other: &Self) -> Option<Ordering> {
+        Some(Ord::cmp(self, other))
+    }
+}
+
+impl Ord for StringName {
+    #[inline]
+    fn cmp(&self, other: &Self) -> Ordering {
+        // Slow but correct -- see comment above
+        Ord::cmp(&self.to_godot_string(), &other.to_godot_string())
     }
 }
 
 impl fmt::Debug for StringName {
     #[inline]
     fn fmt(&self, f: &mut fmt::Formatter) -> Result<(), fmt::Error> {
-        self.get_name().to_string().fmt(f)
-    }
-}
-
-impl PartialOrd for StringName {
-    #[inline]
-    fn partial_cmp(&self, other: &StringName) -> Option<Ordering> {
-        unsafe {
-            let native = (get_api().godot_string_name_operator_less)(&self.0, &other.0);
-
-            if native {
-                Some(Ordering::Less)
-            } else {
-                Some(Ordering::Greater)
-            }
-        }
+        self.to_godot_string().fmt(f)
     }
 }
 
@@ -612,8 +642,59 @@ where
     }
 }
 
+#[cfg(feature = "serde")]
+mod serialize {
+    use super::*;
+    use serde::{
+        de::{Error, Visitor},
+        Deserialize, Deserializer, Serialize, Serializer,
+    };
+    use std::fmt::Formatter;
+
+    impl Serialize for GodotString {
+        #[inline]
+        fn serialize<S>(
+            &self,
+            serializer: S,
+        ) -> Result<<S as Serializer>::Ok, <S as Serializer>::Error>
+        where
+            S: Serializer,
+        {
+            serializer.serialize_str(&self.to_string())
+        }
+    }
+
+    #[cfg(feature = "serde")]
+    impl<'de> serialize::Deserialize<'de> for GodotString {
+        #[inline]
+        fn deserialize<D>(deserializer: D) -> Result<Self, <D as Deserializer<'de>>::Error>
+        where
+            D: Deserializer<'de>,
+        {
+            struct GodotStringVisitor;
+            impl<'de> Visitor<'de> for GodotStringVisitor {
+                type Value = GodotString;
+
+                fn expecting(&self, formatter: &mut Formatter) -> fmt::Result {
+                    formatter.write_str("a GodotString")
+                }
+
+                fn visit_str<E>(self, s: &str) -> Result<Self::Value, E>
+                where
+                    E: Error,
+                {
+                    Ok(GodotString::from(s))
+                }
+            }
+
+            deserializer.deserialize_str(GodotStringVisitor)
+        }
+    }
+}
+
 godot_test!(test_string {
-    use crate::core_types::{GodotString, Variant, VariantType, ToVariant};
+    use crate::core_types::{GodotString, Variant, VariantType, ToVariant, VariantArray, Dictionary};
+    use std::cmp::Ordering;
 
     let foo: GodotString = "foo".into();
     assert_eq!(foo.len(), 3);
@@ -645,17 +726,78 @@ godot_test!(test_string {
     assert_eq!(index_string[1], 'a');
     assert_eq!(index_string[2], 'r');
 
-    let variant = Variant::from_godot_string(&foo);
+    let variant = Variant::new(&foo);
     assert!(variant.get_type() == VariantType::GodotString);
 
     let variant2: Variant = "foo".to_variant();
     assert!(variant == variant2);
 
-    if let Some(foo_variant) = variant.try_to_godot_string() {
+    if let Ok(foo_variant) = variant.try_to::<GodotString>() {
         assert!(foo_variant == foo);
     } else {
         panic!("variant should be a GodotString");
     }
 
     assert_eq!(foo.to_utf8().as_str(), "foo");
+
+    let fmt_string = GodotString::from("foo {bar}");
+    let fmt_data = Dictionary::new();
+    fmt_data.insert("bar", "baz");
+
+    let fmt = fmt_string.format(&fmt_data.into_shared().to_variant());
+    assert_eq!(fmt, GodotString::from("foo baz"));
+    assert_eq!(fmt_string, GodotString::from("foo {bar}"));
+
+    let fmt_string2 = GodotString::from("{0} {1}");
+    let fmt_data2 = VariantArray::new();
+    fmt_data2.push("foo");
+    fmt_data2.push("bar");
+
+    let fmt2 = fmt_string2.format(&fmt_data2.into_shared().to_variant());
+    assert_eq!(fmt2, GodotString::from("foo bar"));
+    assert_eq!(fmt_string2, GodotString::from("{0} {1}"));
+});
+
+godot_test!(test_string_name_eq {
+    use crate::core_types::{GodotString, StringName};
+
+    let a: StringName = StringName::from_str("some string");
+    let b: StringName = StringName::from_godot_string(&GodotString::from("some string"));
+    let c: StringName = StringName::from_str(String::from("some other string"));
+    let d: StringName = StringName::from_str("yet another one");
+
+    // test Eq
+    assert_eq!(a, b);
+    assert_ne!(a, c);
+    assert_ne!(a, d);
+    assert_ne!(b, c);
+    assert_ne!(b, d);
+    assert_ne!(c, d);
+
+    let back = b.to_godot_string();
+    assert_eq!(back, GodotString::from("some string"));
+});
+
+godot_test!(test_string_name_ord {
+    use crate::core_types::{GodotString, StringName};
+
+    let a: StringName = StringName::from_str("some string");
+    let b: StringName = StringName::from_godot_string(&GodotString::from("some string"));
+    let c: StringName = StringName::from_str(String::from("some other string"));
+
+    // test Ord
+    let a_b = Ord::cmp(&a, &b);
+    let b_a = Ord::cmp(&b, &a);
+    assert_eq!(a_b, Ordering::Equal);
+    assert_eq!(b_a, Ordering::Equal);
+
+    let a_c = Ord::cmp(&a, &c);
+    let c_a = Ord::cmp(&c, &a);
+    let b_c = Ord::cmp(&b, &c);
+    let c_b = Ord::cmp(&c, &b);
+    assert_ne!(a_c, Ordering::Equal);
+    assert_eq!(a_c, b_c);
+    assert_eq!(c_a, c_b);
+    assert_eq!(a_c, c_a.reverse());
+    assert_eq!(b_c, c_b.reverse());
 });
