@@ -14,49 +14,55 @@
 //!
 //! Since it is easy to expect containers and other types to allocate a copy of their
 //! content when using the `Clone` trait, some types do not implement `Clone` and instead
-//! implement [`NewRef`](./trait.NewRef.html) which provides a `new_ref(&self) -> Self` method
+//! implement [`NewRef`](object::NewRef) which provides a `new_ref(&self) -> Self` method
 //! to create references to the same collection or object.
 //!
 //! [thread-safety]: https://docs.godotengine.org/en/stable/tutorials/threads/thread_safe_apis.html
 
-#![deny(clippy::missing_inline_in_public_items)]
-#![allow(clippy::transmute_ptr_to_ptr)]
-#![cfg_attr(feature = "gd_test", allow(clippy::blacklisted_name))]
+#![warn(clippy::missing_inline_in_public_items, clippy::exhaustive_enums)]
+#![allow(
+    clippy::transmute_ptr_to_ptr,
+    clippy::missing_safety_doc,
+    clippy::non_send_fields_in_send_ty
+)]
+#![cfg_attr(
+    any(test, feature = "gd-test"),
+    allow(
+        clippy::excessive_precision,
+        clippy::disallowed_names,
+        clippy::redundant_clone
+    )
+)]
 
 #[doc(hidden)]
 pub extern crate gdnative_sys as sys;
+
 #[doc(hidden)]
 pub extern crate libc;
 
-#[cfg(feature = "gd_test")]
+#[cfg(feature = "gd-test")]
 #[macro_use]
 extern crate approx;
+
+#[doc(inline)]
+pub use gdnative_derive::godot_wrap_method;
+
+/// Derive macros and macro attributes.
+#[doc(inline)]
+pub use gdnative_derive as derive;
 
 // Macros have to be processed before they are used.
 mod macros;
 
 pub mod core_types;
-mod init;
 
-#[cfg(feature = "nativescript")]
-pub mod nativescript;
-
+pub mod export;
+pub mod globalscope;
+pub mod init;
 pub mod log;
-mod new_ref;
 pub mod object;
-pub mod ref_kind;
-pub mod thread_access;
+pub mod profiler;
 
 /// Internal low-level API for use by macros and generated bindings. Not a part of the public API.
 #[doc(hidden)]
 pub mod private;
-
-//
-// Re-exports
-//
-
-pub use init::{InitializeInfo, TerminateInfo};
-pub use new_ref::NewRef;
-pub use object::{GodotObject, Null, Ref, TRef};
-
-pub type GodotResult = Result<(), core_types::error::GodotError>;
