@@ -1,103 +1,5 @@
 #![macro_use]
 
-/// Declare the API endpoint to initialize the gdnative API on startup.
-///
-/// By default this declares an extern function named `godot_gdnative_init`.
-/// This can be overridden, for example:
-///
-/// ```ignore
-/// // Declares an extern function named custom_gdnative_init instead of
-/// // godot_gdnative_init.
-/// godot_gdnative_init!(my_init_callback as custom_gdnative_init);
-/// ```
-///
-/// Overriding the default entry point names can be useful if several gdnative
-/// libraries are linked statically  to avoid name clashes.
-#[macro_export]
-macro_rules! godot_gdnative_init {
-    () => {
-        fn godot_gdnative_init_empty(_options: &$crate::InitializeInfo) {}
-        $crate::godot_gdnative_init!(godot_gdnative_init_empty);
-    };
-    (_ as $fn_name:ident) => {
-        fn godot_gdnative_init_empty(_options: &$crate::InitializeInfo) {}
-        $crate::godot_gdnative_init!(godot_gdnative_init_empty as $fn_name);
-    };
-    ($callback:ident) => {
-        $crate::godot_gdnative_init!($callback as godot_gdnative_init);
-    };
-    ($callback:ident as $fn_name:ident) => {
-        #[no_mangle]
-        #[doc(hidden)]
-        #[allow(unused_unsafe)]
-        pub unsafe extern "C" fn $fn_name(options: *mut $crate::sys::godot_gdnative_init_options) {
-            if !$crate::private::bind_api(options) {
-                // Can't use godot_error here because the API is not bound.
-                // Init errors should be reported by bind_api.
-                return;
-            }
-
-            let __result = ::std::panic::catch_unwind(|| {
-                let callback_options = $crate::InitializeInfo::new(options);
-                $callback(&callback_options)
-            });
-            if __result.is_err() {
-                $crate::godot_error!("gdnative-core: gdnative_init callback panicked");
-            }
-        }
-    };
-}
-
-/// Declare the API endpoint invoked during shutdown.
-///
-/// By default this declares an extern function named `godot_gdnative_terminate`.
-/// This can be overridden, for example:
-///
-/// ```ignore
-/// // Declares an extern function named custom_gdnative_terminate instead of
-/// // godot_gdnative_terminate.
-/// godot_gdnative_terminate!(my_shutdown_callback as custom_gdnative_terminate);
-/// ```
-///
-/// Overriding the default entry point names can be useful if several gdnative
-/// libraries are linked statically  to avoid name clashes.
-#[macro_export]
-macro_rules! godot_gdnative_terminate {
-    () => {
-        fn godot_gdnative_terminate_empty(_term_info: &$crate::TerminateInfo) {}
-        $crate::godot_gdnative_terminate!(godot_gdnative_terminate_empty);
-    };
-    ($callback:ident) => {
-        $crate::godot_gdnative_terminate!($callback as godot_gdnative_terminate);
-    };
-    (_ as $fn_name:ident) => {
-        fn godot_gdnative_terminate_empty(_term_info: &$crate::TerminateInfo) {}
-        $crate::godot_gdnative_terminate!(godot_gdnative_terminate_empty as $fn_name);
-    };
-    ($callback:ident as $fn_name:ident) => {
-        #[no_mangle]
-        #[doc(hidden)]
-        #[allow(unused_unsafe)]
-        pub unsafe extern "C" fn $fn_name(
-            options: *mut $crate::sys::godot_gdnative_terminate_options,
-        ) {
-            if !$crate::private::is_api_bound() {
-                return;
-            }
-
-            let __result = ::std::panic::catch_unwind(|| {
-                let term_info = $crate::TerminateInfo::new(options);
-                $callback(&term_info)
-            });
-            if __result.is_err() {
-                $crate::godot_error!("gdnative-core: nativescript_init callback panicked");
-            }
-
-            $crate::private::cleanup_internal_state();
-        }
-    };
-}
-
 /// Print a message using the engine's logging system (visible in the editor).
 #[macro_export]
 macro_rules! godot_print {
@@ -133,16 +35,19 @@ macro_rules! godot_dbg {
     };
 }
 
-/// Creates a `gdnative::log::Site` value from the current position in code,
+/// Creates a [`Site`][crate::log::Site] value from the current position in code,
 /// optionally with a function path for identification.
 ///
 /// # Examples
 ///
 /// ```ignore
+/// use gdnative::log;
+///
 /// // WARN: <unset>: foo At: path/to/file.rs:123
-/// gdnative::log::warn(godot_site!(), "foo");
+/// log::warn(log::godot_site!(), "foo");
+///
 /// // WARN: Foo::my_func: bar At: path/to/file.rs:123
-/// gdnative::log::error(godot_site!(Foo::my_func), "bar");
+/// log::error(log::godot_site!(Foo::my_func), "bar");
 /// ```
 #[macro_export]
 macro_rules! godot_site {
@@ -264,13 +169,32 @@ macro_rules! impl_basic_trait_as_sys {
     (
         Eq for $Type:ty as $GdType:ident : $gd_method:ident
     ) => {
-        impl PartialEq for $Type {
+        impl_basic_trait_as_sys!(PartialEq for $Type as $GdType : $gd_method);
+        impl Eq for $Type {}
+    };
+
+    (
+        Ord for $Type:ty as $GdType:ident : $gd_method:ident
+    ) => {
+        impl PartialOrd for $Type {
             #[inline]
-            fn eq(&self, other: &Self) -> bool {
-                unsafe { (get_api().$gd_method)(self.sys(), other.sys()) }
+            fn partial_cmp(&self, other: &Self) -> Option<std::cmp::Ordering> {
+                Some(self.cmp(other))
             }
         }
-        impl Eq for $Type {}
+        impl Ord for $Type {
+            #[inline]
+            fn cmp(&self, other: &Self) -> std::cmp::Ordering {
+                let op_less = get_api().$gd_method;
+                if unsafe { op_less(&self.0, &other.0) } {
+                    std::cmp::Ordering::Less
+                } else if unsafe { op_less(&other.0, &self.0) } {
+                    std::cmp::Ordering::Greater
+                } else {
+                    std::cmp::Ordering::Equal
+                }
+            }
+        }
     };
 
     (
@@ -303,12 +227,15 @@ macro_rules! impl_basic_traits_as_sys {
     )
 }
 
-macro_rules! godot_test {
-    ($($test_name:ident $body:block)*) => {
+#[doc(hidden)]
+#[macro_export]
+macro_rules! godot_test_impl {
+    ( $( $test_name:ident $body:block $($attrs:tt)* )* ) => {
         $(
-            #[cfg(feature = "gd_test")]
+            $($attrs)*
             #[doc(hidden)]
             #[inline]
+            #[must_use]
             pub fn $test_name() -> bool {
                 let str_name = stringify!($test_name);
                 println!("   -- {}", str_name);
@@ -318,11 +245,47 @@ macro_rules! godot_test {
                 ).is_ok();
 
                 if !ok {
-                    $crate::godot_error!("   !! Test {} failed", str_name);
+                    if ::std::panic::catch_unwind(|| {
+                        $crate::godot_error!("   !! Test {} failed", str_name);
+                    }).is_err() {
+                        eprintln!("   !! Test {} failed", str_name);
+                        eprintln!("   !! And failed to call Godot API to log error message");
+                    }
                 }
 
                 ok
             }
+        )*
+    }
+}
+
+/// Declares a test to be run with the Godot engine (i.e. not a pure Rust unit test).
+///
+/// Creates a wrapper function that catches panics, prints errors and returns true/false.
+/// To be manually invoked in higher-level test routine.
+///
+/// This macro is designed to be used within the current crate only, hence the #[cfg] attribute.
+#[doc(hidden)]
+macro_rules! godot_test {
+    ($($test_name:ident $body:block)*) => {
+        $(
+            godot_test_impl!($test_name $body #[cfg(feature = "gd-test")]);
+        )*
+    }
+}
+
+/// Declares a test to be run with the Godot engine (i.e. not a pure Rust unit test).
+///
+/// Creates a wrapper function that catches panics, prints errors and returns true/false.
+/// To be manually invoked in higher-level test routine.
+///
+/// This macro is designed to be used within the `test` crate, hence the method is always declared (not only in certain features).
+#[doc(hidden)]
+#[macro_export]
+macro_rules! godot_itest {
+    ($($test_name:ident $body:block)*) => {
+        $(
+            $crate::godot_test_impl!($test_name $body);
         )*
     }
 }
