@@ -1,9 +1,11 @@
+use syn::punctuated::Punctuated;
 use syn::visit::Visit;
-use syn::Generics;
+use syn::{GenericParam, Generics};
 
-use crate::extend_bounds::{with_visitor, BoundsVisitor};
+use crate::utils::extend_bounds::{with_visitor, BoundsVisitor};
+use crate::variant::repr::StructRepr;
 
-use super::repr::{Field, Repr, VariantRepr};
+use super::repr::{EnumRepr, Field, Repr, VariantRepr};
 use super::Direction;
 
 pub(crate) fn extend_bounds(
@@ -12,7 +14,7 @@ pub(crate) fn extend_bounds(
     bound: &syn::Path,
     dir: Direction,
 ) -> Generics {
-    with_visitor(generics, bound, |visitor| {
+    with_visitor(generics, Some(bound), None, |visitor| {
         // iterate through parsed variant representations and visit the types of each field
         fn visit_var_repr<'ast>(
             visitor: &mut BoundsVisitor<'ast>,
@@ -20,7 +22,7 @@ pub(crate) fn extend_bounds(
             dir: Direction,
         ) {
             match repr {
-                VariantRepr::Unit => {}
+                VariantRepr::Unit(_) => {}
                 VariantRepr::Tuple(tys) => {
                     for Field { ty, attr, .. } in tys.iter() {
                         if !attr.skip_bounds(dir) {
@@ -39,14 +41,31 @@ pub(crate) fn extend_bounds(
         }
 
         match repr {
-            Repr::Enum(ref variants) => {
+            Repr::Enum(EnumRepr { ref variants, .. }) => {
                 for (_, var_repr) in variants.iter() {
                     visit_var_repr(visitor, var_repr, dir);
                 }
             }
-            Repr::Struct(var_repr) => {
+            Repr::Struct(StructRepr(var_repr)) => {
                 visit_var_repr(visitor, var_repr, dir);
             }
         }
     })
+}
+
+pub(crate) fn remove_bounds(mut generics: Generics) -> Generics {
+    for param in generics.params.iter_mut() {
+        match param {
+            GenericParam::Type(ty) => {
+                ty.colon_token = None;
+                ty.bounds = Punctuated::new();
+            }
+            GenericParam::Lifetime(lt) => {
+                lt.colon_token = None;
+                lt.bounds = Punctuated::new();
+            }
+            GenericParam::Const(_) => {}
+        }
+    }
+    generics
 }
