@@ -1,75 +1,126 @@
 //! # Rust bindings for the Godot game engine
 //!
-//! This crate contains high-level wrappers around the Godot game engine's gdnative API.
+//! This crate contains high-level wrappers around the Godot game engine's GDNative API.
 //! Some of the types were automatically generated from the engine's JSON API description,
 //! and some other types are hand made wrappers around the core C types.
 //!
-//! ## Memory management for core types
+//! ## Core types
 //!
-//! Wrappers for most core types expose safe Rust interfaces, and it's unnecessary to mind
-//! memory management most of the times. The exceptions are `VariantArray` and `Dictionary`,
-//! internally reference-counted collections with "interior mutability" in Rust parlance. These
-//! types are modelled using the "typestate" pattern to enforce that the official
+//! Wrappers for most core types expose safe Rust interfaces, and it's unnecessary
+//! to mind memory management most of the times. The exceptions are
+//! [`VariantArray`](core_types::VariantArray) and [`Dictionary`](core_types::Dictionary),
+//! internally reference-counted collections with _interior mutability_ in Rust parlance.
+//! These types are modelled using the _typestate_ pattern to enforce that the official
 //! [thread-safety guidelines][thread-safety]. For more information, read the type-level
 //! documentation for these types.
 //!
 //! Since it is easy to expect containers and other types to allocate a copy of their
 //! content when using the `Clone` trait, some types do not implement `Clone` and instead
-//! implement [`NewRef`](./trait.NewRef.html) which provides a `new_ref(&self) -> Self` method
+//! implement [`NewRef`](object::NewRef) which provides a `new_ref(&self) -> Self` method
 //! to create references to the same collection or object.
 //!
 //! ## Generated API types
 //!
-//! The `api` module contains high-level wrappers for all the API types generated from a
-//! JSON description of the API. The generated types are tied to a specific version, which
-//! is currently `3.2.3-stable` for the crates.io version. If you want to use the bindings
-//! with another version of the engine, see the instructions [here][custom-version] on
-//! generating custom bindings.
+//! The [`api`] module contains high-level wrappers for all the API types generated from a
+//! JSON description of the API. The generated types are tied to a specific version, typically
+//! the latest Godot 3.x release (at the time of the godot-rust release).
+//! If you want to use the bindings with another version of the engine, read the notes on
+//! the `custom-godot` feature flag below.
 //!
 //! ### Memory management
 //!
 //! API types may be reference-counted or manually-managed. This is indicated by the
 //! `RefCounted` and `ManuallyManaged` marker traits.
 //!
-//! The API types can exist in three reference forms: bare, `TRef` and `Ref`. Bare references
-//! to API types, like `&'a Node`, represent valid and safe references to Godot objects.
+//! The API types can exist in three reference forms: bare, [`TRef`](object::TRef) and [`Ref`](object::Ref).
+//! Bare references to API types, like `&'a Node`, represent valid and safe references to Godot objects.
 //! As such, API methods may be called safely on them. `TRef` adds typestate tracking, which
 //! enable additional abilities like being able to be passed to the engine. `Ref`, or
-//! "persistent" references, have `'static` lifetime, but are not always safe to use. For more
-//! information on how to use persistent references safely, see the type-level documentation
-//! on `Ref`.
+//! _persistent_ references, have `'static` lifetime, but are not always safe to use. For more
+//! information on how to use persistent references safely, see the [`object`] module documentation
+//! or the corresponding [book chapter][gdnative-overview].
 //!
 //! ## Feature flags
+//! All features are disabled by default.
 //!
-//! ### `bindings`
+//! Functionality toggles:
 //!
-//! *Enabled* by default. Includes the crates.io version of the bindings in the `api` module.
+//! * **`async`**<br>
+//!   Activates async functionality, see [`tasks`] module for details.
+//!
+//! * **`serde`**<br>
+//!   Enable for `serde` support of several core types. See also [`Variant`](core_types::Variant).
+//!
+//! * **`inventory`**<br>
+//!   Enables automatic class registration via `inventory`.
+//!
+//!   **Attention:** Automatic registration is unsupported on some platforms, notably WASM. `inventory`
+//!   can still be used for iterative development if such platforms are targeted, in which case the
+//!   run-time diagnostic [`init::diagnostics::missing_manual_registration`] may be helpful.
+//!
+//!   Please refer to [the `rust-ctor` README][ctor-repo] for an up-to-date listing of platforms
+//!   that *do* support automatic registration.
+//!
+//! Bindings generation:
+//!
+//! * **`custom-godot`**<br>
+//!   When active, tries to locate a Godot executable on your system, in this order:
+//!   1. If a `GODOT_BIN` environment variable is defined, it will interpret it as a path to the binary
+//!      (not directory).
+//!   2. An executable called `godot`, accessible in your system's PATH, is used.
+//!   3. If neither of the above is found, an error is generated.
+//!
+//!   The symbols in [`api`] will be generated in a way compatible with that engine version.
+//!   This allows to use Godot versions older than the currently supported ones.
+//!
+//!   See [Custom Godot builds][custom-godot] for detailed instructions.
+//!
+//! * **`formatted`**<br>
+//!   Enable if the generated binding source code should be human-readable and split
+//!   into multiple files. This can also help IDEs that struggle with a single huge file.
+//!
+//! * **`ptrcall`**<br>
+//!   Enables the `ptrcall` convention for calling Godot API methods. This increases performance, at the
+//!   cost of forward binary compatibility with the engine. Binaries built with `ptrcall` enabled
+//!   **may exhibit undefined behavior** when loaded by a different version of Godot, even when there are
+//!   no breaking API changes as far as GDScript is concerned. Notably, the addition of new default
+//!   parameters breaks any code using `ptrcall`.
+//!
+//!   Cargo features are additive, and as such, it's only necessary to enable this feature for the final
+//!   `cdylib` crates, whenever desired.
 //!
 //! [thread-safety]: https://docs.godotengine.org/en/stable/tutorials/threads/thread_safe_apis.html
-//! [custom-version]: https://github.com/godot-rust/godot-rust/#other-versions-or-custom-builds
+//! [gdnative-overview]: https://godot-rust.github.io/book/gdnative-overview.html
+//! [custom-godot]: https://godot-rust.github.io/book/advanced-guides/custom-godot.html
+//! [ctor-repo]: https://github.com/mmastrac/rust-ctor
+//!
 //!
 
-// TODO: add logo using #![doc(html_logo_url = "https://<url>")]
+#![doc(html_logo_url = "https://github.com/godot-rust/gdnative/raw/master/assets/godot-ferris.svg")]
 
-// Workaround: rustdoc currently shows hidden items in the original crate when they are
-//             re-exported. Manually re-exporting the public items works around that.
+// Workaround (rustdoc 1.55):
+// Items, which are #[doc(hidden)] in their original crate and re-exported with a wildcard, lose
+// their hidden status. Re-exporting them manually and hiding the wildcard solves this.
 #[doc(inline)]
 pub use gdnative_core::{
-    core_types, godot_dbg, godot_error, godot_gdnative_init, godot_gdnative_terminate, godot_init,
-    godot_nativescript_init, godot_print, godot_warn, godot_wrap_method, nativescript, object,
-    ref_kind, thread_access, GodotObject, GodotResult, NewRef, Null, Ref, TRef,
+    core_types, derive, export, godot_dbg, godot_error, godot_print, godot_site, init, log, object,
+    profiler,
 };
 
-#[doc(hidden)]
-pub use gdnative_core::*;
+pub mod globalscope;
 
-#[doc(inline)]
-pub use gdnative_derive::*;
+// Implementation details (e.g. used by macros).
+// However, do not re-export macros (on crate level), thus no wildcard
+#[doc(hidden)]
+pub use gdnative_core::{libc, private, sys};
 
 /// Curated re-exports of common items.
 pub mod prelude;
 
-#[doc(inline)]
-#[cfg(feature = "bindings")]
 /// Bindings for the Godot Class API.
+#[doc(inline)]
 pub use gdnative_bindings as api;
+
+#[doc(inline)]
+#[cfg(feature = "async")]
+pub use gdnative_async as tasks;
