@@ -1,41 +1,41 @@
-pub use gdnative_core::core_types::{
-    self, error::GodotError, Aabb, Basis, ByteArray, Color, ColorArray, Dictionary, Float32Array,
-    GodotString, Int32Array, NodePath, Plane, Quat, Rect2, Rid, StringArray, StringName, Transform,
-    Transform2D, TypedArray, Variant, VariantArray, VariantDispatch, VariantOperator, VariantType,
-    Vector2, Vector2Array, Vector3, Vector3Array,
-};
-pub use gdnative_core::core_types::{
-    FromVariant, FromVariantError, OwnedToVariant, ToVariant, ToVariantEq,
-};
-
-pub use gdnative_core::object::{
-    AsArg, GodotObject, Instanciable, Null, QueueFree, Ref, SubClass, TRef,
-};
-pub use gdnative_core::ref_kind::{ManuallyManaged, RefCounted};
-pub use gdnative_core::thread_access::{Shared, ThreadLocal, Unique};
-pub use gdnative_core::NewRef;
-
-pub use gdnative_core::nativescript::{
-    self,
-    class::{Instance, RefInstance},
-    init::{ClassBuilder, InitHandle, Method, MethodBuilder, Signal, SignalArgument},
-    user_data::{self, Aether, ArcData, LocalCellData, MutexData, RwLockData},
-    ExportInfo, NativeClass, NativeClassMethods, PropertyUsage,
-};
-
-pub use gdnative_core::{
-    godot_dbg, godot_error, godot_gdnative_init, godot_gdnative_terminate, godot_init,
-    godot_nativescript_init, godot_print, godot_site, godot_warn,
-};
-
-pub use gdnative_derive::*;
-
-#[cfg(feature = "bindings")]
+pub use gdnative_bindings::utils::*;
 pub use gdnative_bindings::{
     Button, CanvasItem, CanvasLayer, ColorRect, Control, Image, Input, InputEvent, InputEventKey,
     KinematicBody, KinematicBody2D, Label, Node, Node2D, Object, PackedScene, Reference,
     ResourceLoader, SceneTree, Shader, Spatial, Sprite, Texture, Timer, Tween, Viewport,
 };
+pub use gdnative_core::core_types::{
+    Aabb, Basis, Color, Dictionary, GodotError, GodotString, NodePath, Plane, PoolArray, Quat,
+    Rect2, Rid, StringName, Transform, Transform2D, Variant, VariantArray, VariantType, Vector2,
+    Vector3,
+};
+pub use gdnative_core::core_types::{
+    FromVariant, FromVariantError, OwnedToVariant, ToVariant, ToVariantEq,
+};
+pub use gdnative_core::export::{
+    ClassBuilder, ExportInfo, Method, MethodBuilder, NativeClass, NativeClassMethods, Property,
+    PropertyUsage, SignalBuilder, SignalParam,
+};
+pub use gdnative_core::init::{GDNativeCallbacks, InitHandle};
+pub use gdnative_core::object::{
+    memory::{ManuallyManaged, RefCounted},
+    ownership::{Shared, ThreadLocal, Unique},
+    AsArg, GodotObject, Instance, Instanciable, NewRef, Null, QueueFree, Ref, SubClass, TInstance,
+    TRef,
+};
+pub use gdnative_core::{godot_dbg, godot_error, godot_print, godot_warn};
+#[allow(deprecated)]
+pub use gdnative_core::{
+    godot_gdnative_init, godot_gdnative_terminate, godot_init, godot_nativescript_init,
+};
+pub use gdnative_derive::*;
 
-#[cfg(feature = "bindings")]
-pub use gdnative_bindings::utils::*;
+/// User-data attributes from [`export::user_data`][crate::export::user_data] module.
+pub mod user_data {
+    // Re-export selected user_data types, but keep qualified due to rather generic names
+    pub use gdnative_core::export::user_data::{
+        Aether, ArcData, LocalCellData, MutexData, RwLockData,
+    };
+}
+#[doc(inline)]
+pub use crate::globalscope::load;
