@@ -3,17 +3,20 @@ fn ui_tests() {
     let t = trybuild::TestCases::new();
 
     // NativeClass
+    t.pass("tests/ui/derive_flexible_self_types.rs");
+    t.pass("tests/ui/derive_no_inherit.rs");
     t.pass("tests/ui/derive_pass.rs");
     t.pass("tests/ui/derive_property_basic.rs");
-    t.pass("tests/ui/derive_no_inherit.rs");
-    t.compile_fail("tests/ui/derive_fail_property_hint.rs");
     t.compile_fail("tests/ui/derive_fail_inherit_param.rs");
-    t.compile_fail("tests/ui/derive_fail_methods.rs");
-    t.compile_fail("tests/ui/derive_fail_methods_param.rs");
+    t.compile_fail("tests/ui/derive_fail_lifetime.rs");
     t.compile_fail("tests/ui/derive_fail_methods_list.rs");
     t.compile_fail("tests/ui/derive_fail_methods_missing_new.rs");
-    t.compile_fail("tests/ui/derive_fail_userdata.rs");
+    t.compile_fail("tests/ui/derive_fail_methods_param.rs");
+    t.compile_fail("tests/ui/derive_fail_methods_special_args.rs");
+    t.compile_fail("tests/ui/derive_fail_methods.rs");
     t.compile_fail("tests/ui/derive_fail_property_empty_hint.rs");
+    t.compile_fail("tests/ui/derive_fail_property_hint.rs");
+    t.compile_fail("tests/ui/derive_fail_userdata.rs");
 
     // Variants
     t.pass("tests/ui/variant_pass.rs");
@@ -25,6 +28,8 @@ fn ui_tests() {
     t.compile_fail("tests/ui/to_variant_fail_05.rs");
     t.compile_fail("tests/ui/to_variant_fail_06.rs");
     t.compile_fail("tests/ui/to_variant_fail_07.rs");
+    t.compile_fail("tests/ui/to_variant_fail_08.rs");
+    t.compile_fail("tests/ui/to_variant_fail_09.rs");
 
     // FromVariant
     t.compile_fail("tests/ui/from_variant_fail_01.rs");
@@ -33,6 +38,8 @@ fn ui_tests() {
     t.compile_fail("tests/ui/from_variant_fail_05.rs");
     t.compile_fail("tests/ui/from_variant_fail_06.rs");
     t.compile_fail("tests/ui/from_variant_fail_07.rs");
+    t.compile_fail("tests/ui/from_variant_fail_08.rs");
+    t.compile_fail("tests/ui/from_variant_fail_09.rs");
 }
 
 // FIXME(rust/issues/54725): Full path spans are only available on nightly as of now
