@@ -1,30 +1,54 @@
-use std::cell::Cell;
+use std::cell::{self, Cell, RefCell};
+use std::collections::HashMap;
+use std::rc::Rc;
 
+use gdnative::export::Property;
 use gdnative::prelude::*;
 
 pub(crate) fn run_tests() -> bool {
     let mut status = true;
 
     status &= test_derive_to_variant();
+    status &= test_derive_to_variant_repr();
+    status &= test_derive_to_variant_str();
     status &= test_derive_owned_to_variant();
-    status &= test_derive_nativeclass_with_property_hooks();
+    status &= test_derive_nativeclass();
     status &= test_derive_nativeclass_without_constructor();
+    status &= test_derive_nativeclass_without_inherit();
+    status &= test_derive_nativeclass_godot_attr_without_base();
+    status &= test_derive_nativeclass_godot_attr_with_base();
+    status &= test_derive_nativeclass_godot_attr_deref_return();
+    status &= test_derive_nativeclass_godot_attr_rename_method();
+    status &= test_derive_nativeclass_godot_attr_all_arguments();
+    status &= test_derive_nativeclass_with_property_get_set();
+    status &= test_derive_nativeclass_property_with_only_getter();
 
     status
 }
 
+#[cfg(not(feature = "no-manual-register"))]
 pub(crate) fn register(handle: InitHandle) {
-    handle.add_class::<PropertyHooks>();
+    handle.add_class::<MinimalDerive>();
     handle.add_class::<EmplacementOnly>();
+    handle.add_class::<WithoutInherit>();
+    handle.add_class::<GodotAttrWithoutBase>();
+    handle.add_class::<GodotAttrWithBase>();
+    handle.add_class::<GodotAttrDerefReturn>();
+    handle.add_class::<GodotAttrRenameMethod>();
+    handle.add_class::<GodotAttrAllArguments>();
+    handle.add_class::<CustomGetSet>();
+    handle.add_class::<MyVec>();
 }
 
-fn test_derive_to_variant() -> bool {
-    println!(" -- test_derive_to_variant");
+#[cfg(feature = "no-manual-register")]
+pub(crate) fn register(_handle: InitHandle) {}
 
+// ----------------------------------------------------------------------------------------------------------------------------------------------
+
+crate::godot_itest! { test_derive_to_variant {
     #[derive(Clone, Eq, PartialEq, Debug, ToVariant, FromVariant)]
-    struct ToVar<T, R>
+    struct ToVar<T: Associated, R>
     where
-        T: Associated,
         R: Default,
     {
         foo: T::A,
@@ -37,7 +61,7 @@ fn test_derive_to_variant() -> bool {
     }
 
     #[derive(Clone, Eq, PartialEq, Debug, ToVariant, FromVariant)]
-    enum ToVarEnum<T> {
+    enum ToVarEnum<T: Bound> {
         Foo(T),
         Bar,
         Baz { baz: u8 },
@@ -49,9 +73,12 @@ fn test_derive_to_variant() -> bool {
         T: Associated,
         R: Default;
 
+    trait Bound {}
+    impl Bound for bool {}
+
     trait Associated {
         type A;
-        type B;
+        type B : Bound;
     }
 
     impl Associated for f64 {
@@ -72,217 +99,230 @@ fn test_derive_to_variant() -> bool {
         }
     }
 
-    let ok = std::panic::catch_unwind(|| {
-        let data = ToVar::<f64, i128> {
+    let data = ToVar::<f64, i128> {
+        foo: 42,
+        bar: 54.0,
+        baz: ToVarEnum::Foo(true),
+        ptr: std::ptr::null_mut(),
+        skipped: 42,
+    };
+
+    let variant = data.to_variant();
+    let dictionary = variant
+        .try_to::<Dictionary>()
+        .expect("should be dictionary");
+    assert_eq!(Some(42), dictionary.get("foo").and_then(|v| v.to::<i64>()));
+    assert_eq!(
+        Some(54.0),
+        dictionary.get("bar").and_then(|v| v.to::<f64>())
+    );
+    assert_eq!(
+        Some("*mut ()".into()),
+        dictionary.get("ptr").and_then(|v| v.to::<String>())
+    );
+    assert!(!dictionary.contains("skipped"));
+
+    let enum_dict = dictionary
+        .get("baz")
+        .and_then(|v| v.to::<Dictionary>())
+        .expect("should be dictionary");
+    assert_eq!(
+        Some(true),
+        enum_dict.get("Foo").and_then(|v| v.to::<bool>())
+    );
+
+    assert_eq!(
+        Ok(ToVar::<f64, i128> {
             foo: 42,
             bar: 54.0,
             baz: ToVarEnum::Foo(true),
             ptr: std::ptr::null_mut(),
-            skipped: 42,
-        };
+            skipped: 0,
+        }),
+        ToVar::from_variant(&variant)
+    );
 
-        let variant = data.to_variant();
-        let dictionary = variant.try_to_dictionary().expect("should be dictionary");
-        assert_eq!(Some(42), dictionary.get("foo").and_then(|v| v.try_to_i64()));
-        assert_eq!(
-            Some(54.0),
-            dictionary.get("bar").and_then(|v| v.try_to_f64())
-        );
-        assert_eq!(
-            Some("*mut ()".into()),
-            dictionary.get("ptr").and_then(|v| v.try_to_string())
-        );
-        assert!(!dictionary.contains("skipped"));
+    let data = ToVarTuple::<f64, i128>(1, 2, false);
+    let variant = data.to_variant();
+    let tuple_array = variant.to::<VariantArray>().expect("should be array");
 
-        let enum_dict = dictionary
-            .get("baz")
-            .and_then(|v| v.try_to_dictionary())
-            .expect("should be dictionary");
-        assert_eq!(
-            Some(true),
-            enum_dict.get("Foo").and_then(|v| v.try_to_bool())
-        );
+    assert_eq!(2, tuple_array.len());
+    assert_eq!(Some(1), tuple_array.get(0).to::<i64>());
+    assert_eq!(Some(false), tuple_array.get(1).to::<bool>());
+    assert_eq!(
+        Ok(ToVarTuple::<f64, i128>(1, 0, false)),
+        ToVarTuple::from_variant(&variant)
+    );
 
-        assert_eq!(
-            Ok(ToVar::<f64, i128> {
-                foo: 42,
-                bar: 54.0,
-                baz: ToVarEnum::Foo(true),
-                ptr: std::ptr::null_mut(),
-                skipped: 0,
-            }),
-            ToVar::from_variant(&variant)
-        );
+    // Derive on uninhabitable enum results an error
+    #[derive(Debug, PartialEq, FromVariant)]
+    enum NoVariant {}
 
-        let data = ToVarTuple::<f64, i128>(1, 2, false);
-        let variant = data.to_variant();
-        let tuple_array = variant.try_to_array().expect("should be array");
+    let input = HashMap::from_iter([("foo", "bar")]).to_variant();
+    assert_eq!(
+        NoVariant::from_variant(&input),
+        Err(FromVariantError::UnknownEnumVariant {
+            variant: "foo".into(),
+            expected: &[]
+        })
+    );
+}}
 
-        assert_eq!(2, tuple_array.len());
-        assert_eq!(Some(1), tuple_array.get(0).try_to_i64());
-        assert_eq!(Some(false), tuple_array.get(1).try_to_bool());
-        assert_eq!(
-            Ok(ToVarTuple::<f64, i128>(1, 0, false)),
-            ToVarTuple::from_variant(&variant)
-        );
-    })
-    .is_ok();
+// ----------------------------------------------------------------------------------------------------------------------------------------------
 
-    if !ok {
-        gdnative::godot_error!("   !! Test test_derive_to_variant failed");
+crate::godot_itest! { test_derive_to_variant_repr {
+    const ANSWER: u8 = 42;
+
+    #[derive(Copy, Clone, Eq, PartialEq, Debug, ToVariant, FromVariant)]
+    #[variant(enum = "repr")]
+    #[repr(u8)]
+    enum ToVarRepr {
+        A = 0,
+        B,
+        C,
+        D = 128 - 1,
+        E,
+        F = ANSWER,
     }
 
-    ok
-}
+    #[derive(Clone, Eq, PartialEq, Debug, OwnedToVariant, FromVariant)]
+    #[variant(enum = "repr")]
+    #[repr(u8)]
+    enum ToVarReprOwned {
+        A = 0,
+        B,
+        C,
+        D = 128 - 1,
+        E,
+        F = ANSWER,
+    }
 
-fn test_derive_owned_to_variant() -> bool {
-    println!(" -- test_derive_owned_to_variant");
+    let variant = ToVarRepr::A.to_variant();
+    assert_eq!(Some(0), variant.to::<u8>());
 
+    let variant = ToVarRepr::B.to_variant();
+    assert_eq!(Some(1), variant.to::<u8>());
+
+    let variant = ToVarRepr::E.to_variant();
+    assert_eq!(Some(128), variant.to::<u8>());
+
+    let variant = ToVarReprOwned::A.owned_to_variant();
+    assert_eq!(Some(0), variant.to::<u8>());
+
+    let variant = ToVarReprOwned::C.owned_to_variant();
+    assert_eq!(Some(2), variant.to::<u8>());
+
+    let variant = ToVarReprOwned::F.owned_to_variant();
+    assert_eq!(Some(42), variant.to::<u8>());
+
+    assert_eq!(Some(ToVarRepr::A), Variant::new(0).to::<ToVarRepr>());
+    assert_eq!(Some(ToVarRepr::B), Variant::new(1).to::<ToVarRepr>());
+    assert_eq!(Some(ToVarRepr::C), Variant::new(2).to::<ToVarRepr>());
+    assert_eq!(Some(ToVarRepr::D), Variant::new(127).to::<ToVarRepr>());
+    assert_eq!(Some(ToVarRepr::E), Variant::new(128).to::<ToVarRepr>());
+    assert_eq!(Some(ToVarRepr::F), Variant::new(42).to::<ToVarRepr>());
+    assert_eq!(None, Variant::new(48).to::<ToVarRepr>());
+    assert_eq!(None, Variant::new(192).to::<ToVarRepr>());
+}}
+
+// ----------------------------------------------------------------------------------------------------------------------------------------------
+
+crate::godot_itest! { test_derive_to_variant_str {
+    #[derive(Clone, Eq, PartialEq, Debug, ToVariant, FromVariant)]
+    #[variant(enum = "str")]
+    enum ToVarStr {
+        A,
+        B,
+        C,
+    }
+
+    #[derive(Clone, Eq, PartialEq, Debug, OwnedToVariant, FromVariant)]
+    #[variant(enum = "str")]
+    enum ToVarStrOwned {
+        A,
+        B,
+        C,
+    }
+
+    let variant = ToVarStr::A.to_variant();
+    assert_eq!(Some("A"), variant.to::<String>().as_deref());
+
+    let variant = ToVarStr::B.to_variant();
+    assert_eq!(Some("B"), variant.to::<String>().as_deref());
+
+    let variant = ToVarStr::C.to_variant();
+    assert_eq!(Some("C"), variant.to::<String>().as_deref());
+
+    let variant = ToVarStrOwned::A.owned_to_variant();
+    assert_eq!(Some("A"), variant.to::<String>().as_deref());
+
+    let variant = ToVarStrOwned::B.owned_to_variant();
+    assert_eq!(Some("B"), variant.to::<String>().as_deref());
+
+    let variant = ToVarStrOwned::C.owned_to_variant();
+    assert_eq!(Some("C"), variant.to::<String>().as_deref());
+
+    assert_eq!(Some(ToVarStr::A), Variant::new("A").to::<ToVarStr>());
+    assert_eq!(Some(ToVarStr::B), Variant::new("B").to::<ToVarStr>());
+    assert_eq!(Some(ToVarStr::C), Variant::new("C").to::<ToVarStr>());
+    assert_eq!(None, Variant::new("").to::<ToVarStr>());
+    assert_eq!(None, Variant::new("D").to::<ToVarStr>());
+}}
+
+// ----------------------------------------------------------------------------------------------------------------------------------------------
+
+crate::godot_itest! { test_derive_owned_to_variant {
     #[derive(OwnedToVariant)]
     struct ToVar {
         arr: VariantArray<Unique>,
     }
 
-    let ok = std::panic::catch_unwind(|| {
-        let data = ToVar {
-            arr: [1, 2, 3].iter().collect(),
-        };
+    let data = ToVar {
+        arr: [1, 2, 3].iter().collect(),
+    };
 
-        let variant = data.owned_to_variant();
-        let dictionary = variant.try_to_dictionary().expect("should be dictionary");
-        let array = dictionary
-            .get("arr")
-            .and_then(|v| v.try_to_array())
-            .expect("should be array");
-        assert_eq!(3, array.len());
-        assert_eq!(
-            &[1, 2, 3],
-            array
-                .iter()
-                .map(|v| v.try_to_i64().unwrap())
-                .collect::<Vec<_>>()
-                .as_slice()
-        );
-    })
-    .is_ok();
+    let variant = data.owned_to_variant();
+    let dictionary = variant.to::<Dictionary>().expect("should be dictionary");
+    let array = dictionary
+        .get("arr")
+        .and_then(|v| v.to::<VariantArray>())
+        .expect("should be array");
+    assert_eq!(3, array.len());
+    assert_eq!(
+        &[1, 2, 3],
+        array
+            .iter()
+            .map(|v| v.to::<i64>().unwrap())
+            .collect::<Vec<_>>()
+            .as_slice()
+    );
+}}
 
-    if !ok {
-        gdnative::godot_error!("   !! Test test_derive_owned_to_variant failed");
+// ----------------------------------------------------------------------------------------------------------------------------------------------
+
+#[derive(NativeClass)]
+#[inherit(Reference)]
+struct MinimalDerive(i64);
+
+#[methods]
+impl MinimalDerive {
+    fn new(_owner: &Reference) -> Self {
+        Self(54)
     }
 
-    ok
-}
-
-#[derive(gdnative::NativeClass)]
-#[inherit(Node)]
-struct PropertyHooks {
-    #[property(
-        before_get = "Self::before_get",
-        after_get = "Self::after_get",
-        before_set = "Self::before_set",
-        after_set = "Self::after_set"
-    )]
-    value: u32,
-
-    pub before_get_called: Cell<u32>,
-    pub after_get_called: Cell<u32>,
-    pub before_set_value: Option<u32>,
-    pub after_set_value: Option<u32>,
-}
-
-#[gdnative_derive::methods]
-impl PropertyHooks {
-    fn new(_owner: &Node) -> Self {
-        Self {
-            value: 0,
-            before_get_called: Cell::new(0),
-            after_get_called: Cell::new(0),
-            before_set_value: None,
-            after_set_value: None,
-        }
-    }
-
-    fn before_get(&self, _owner: TRef<Node, Shared>) {
-        assert_eq!(self.before_get_called.get(), self.after_get_called.get());
-        self.before_get_called.set(self.before_get_called.get() + 1);
-    }
-
-    fn after_get(&self, _owner: TRef<Node, Shared>) {
-        assert_eq!(
-            self.before_get_called.get(),
-            self.after_get_called.get() + 1
-        );
-        self.after_get_called.set(self.after_get_called.get() + 1);
-    }
-
-    fn assert_get_calls(&self, times: u32) {
-        assert_eq!(times, self.before_get_called.get());
-        assert_eq!(times, self.after_get_called.get());
-    }
-
-    fn before_set(&mut self, _owner: TRef<Node, Shared>) {
-        self.before_set_value = Some(self.value);
-    }
-
-    fn after_set(&mut self, _owner: TRef<Node, Shared>) {
-        self.after_set_value = Some(self.value);
-    }
-
-    fn reset_set_value(&mut self) {
-        self.before_set_value = None;
-        self.after_set_value = None;
+    #[export] // deliberately use old attribute
+    fn answer(&self, _owner: &Reference) -> i64 {
+        self.0
     }
 }
 
-fn test_derive_nativeclass_with_property_hooks() -> bool {
-    println!(" -- test_derive_nativeclass_with_property_hooks");
+crate::godot_itest! { test_derive_nativeclass {
+    let thing = Instance::<MinimalDerive, _>::new();
+    let base: Ref<Reference, Unique> = thing.into_base();
+    assert_eq!(unsafe { base.call("answer", &[]).to::<i64>() }, Some(54));
+}}
 
-    let ok = std::panic::catch_unwind(|| {
-        use gdnative::nativescript::user_data::MapMut;
-
-        let thing = Instance::<PropertyHooks, _>::new();
-        let (owner, script) = thing.decouple();
-
-        owner.set("value", 42);
-        script
-            .map_mut(|script| {
-                assert_eq!(Some(0), script.before_set_value);
-                assert_eq!(Some(42), script.after_set_value);
-                script.reset_set_value();
-            })
-            .unwrap();
-
-        script
-            .map_mut(|script| {
-                script.assert_get_calls(0);
-            })
-            .unwrap();
-        assert_eq!(42, u32::from_variant(&owner.get("value")).unwrap());
-        script
-            .map_mut(|script| {
-                script.assert_get_calls(1);
-            })
-            .unwrap();
-
-        owner.set("value", 12345);
-        script
-            .map_mut(|script| {
-                assert_eq!(Some(42), script.before_set_value);
-                assert_eq!(Some(12345), script.after_set_value);
-                script.reset_set_value();
-            })
-            .unwrap();
-
-        owner.free();
-    })
-    .is_ok();
-
-    if !ok {
-        gdnative::godot_error!("   !! Test test_derive_owned_to_variant failed");
-    }
-
-    ok
-}
+// ----------------------------------------------------------------------------------------------------------------------------------------------
 
 #[derive(NativeClass)]
 #[inherit(Reference)]
@@ -291,32 +331,274 @@ struct EmplacementOnly(i64);
 
 #[methods]
 impl EmplacementOnly {
-    #[export]
-    fn answer(&self, _owner: &Reference) -> i64 {
+    #[method]
+    fn answer(&self, #[base] _base: &Reference) -> i64 {
         self.0
     }
 }
 
-fn test_derive_nativeclass_without_constructor() -> bool {
-    println!(" -- test_derive_nativeclass_without_constructor");
+crate::godot_itest! { test_derive_nativeclass_without_constructor {
+    let foo = Instance::emplace(EmplacementOnly(54));
+    assert_eq!(Ok(54), foo.map(|foo, base| { foo.answer(&base) }));
 
-    let ok = std::panic::catch_unwind(|| {
-        let foo = Instance::emplace(EmplacementOnly(54));
+    let base = foo.into_base();
+    assert_eq!(Some(54), unsafe { base.call("answer", &[]).to::<i64>() });
 
-        assert_eq!(Ok(54), foo.map(|foo, owner| { foo.answer(&*owner) }));
+    let foo = Instance::<EmplacementOnly, _>::try_from_base(base)
+        .expect("should be able to downcast");
+    assert_eq!(Ok(54), foo.map(|foo, base| { foo.answer(&base) }));
+}}
 
-        let base = foo.into_base();
-        assert_eq!(Some(54), unsafe { base.call("answer", &[]).try_to_i64() });
+// ----------------------------------------------------------------------------------------------------------------------------------------------
 
-        let foo = Instance::<EmplacementOnly, _>::try_from_base(base)
-            .expect("should be able to downcast");
-        assert_eq!(Ok(54), foo.map(|foo, owner| { foo.answer(&*owner) }));
-    })
-    .is_ok();
+#[derive(NativeClass)]
+struct WithoutInherit(i64);
 
-    if !ok {
-        gdnative::godot_error!("   !! Test test_derive_nativeclass_without_constructor failed");
+#[methods]
+impl WithoutInherit {
+    fn new(_owner: &Reference) -> Self {
+        Self(54)
     }
 
-    ok
+    #[method]
+    fn answer(&self) -> i64 {
+        self.0
+    }
 }
+
+crate::godot_itest! { test_derive_nativeclass_without_inherit {
+    let thing = Instance::<WithoutInherit, _>::new();
+    let base = thing.into_base();
+    assert_eq!(unsafe { base.call("answer", &[]).to::<i64>() }, Some(54));
+}}
+
+// ----------------------------------------------------------------------------------------------------------------------------------------------
+
+#[derive(NativeClass)]
+#[inherit(Reference)]
+struct GodotAttrWithoutBase(i64);
+
+#[methods]
+impl GodotAttrWithoutBase {
+    fn new(_owner: &Reference) -> Self {
+        Self(54)
+    }
+
+    #[method]
+    fn answer(&self) -> i64 {
+        self.0
+    }
+}
+
+crate::godot_itest! { test_derive_nativeclass_godot_attr_without_base {
+    let thing = Instance::<GodotAttrWithoutBase, _>::new();
+    let base = thing.into_base();
+    assert_eq!(unsafe { base.call("answer", &[]).to::<i64>() }, Some(54));
+}}
+
+// ----------------------------------------------------------------------------------------------------------------------------------------------
+
+#[derive(NativeClass)]
+#[inherit(Reference)]
+struct GodotAttrWithBase(i64);
+
+#[methods]
+impl GodotAttrWithBase {
+    fn new(_owner: &Reference) -> Self {
+        Self(54)
+    }
+
+    #[method]
+    fn answer(&self, #[base] _base: &Reference) -> i64 {
+        self.0
+    }
+}
+
+crate::godot_itest! { test_derive_nativeclass_godot_attr_with_base {
+    let thing = Instance::<GodotAttrWithBase, _>::new();
+    let base = thing.into_base();
+    assert_eq!(unsafe { base.call("answer", &[]).to::<i64>() }, Some(54));
+}}
+
+// ----------------------------------------------------------------------------------------------------------------------------------------------
+
+#[derive(NativeClass)]
+#[inherit(Reference)]
+struct GodotAttrDerefReturn(Rc<RefCell<Vec<i64>>>);
+
+#[methods]
+impl GodotAttrDerefReturn {
+    fn new(_owner: &Reference) -> Self {
+        let vec = Vec::from([12, 34]);
+        let rc_ref = Rc::new(RefCell::new(vec));
+        Self(rc_ref)
+    }
+
+    #[method(deref_return)]
+    fn answer(&self) -> cell::Ref<Vec<i64>> {
+        self.0.borrow()
+    }
+}
+
+crate::godot_itest! { test_derive_nativeclass_godot_attr_deref_return {
+    let thing = Instance::<GodotAttrDerefReturn, _>::new();
+    let base = thing.into_base();
+
+    let res = unsafe { base.call("answer", &[]).to::<Vec<i64>>() };
+    assert_eq!(res, Some([12, 34].into()));
+}}
+
+// ----------------------------------------------------------------------------------------------------------------------------------------------
+
+#[derive(NativeClass)]
+#[inherit(Reference)]
+struct GodotAttrRenameMethod(i64);
+
+#[methods]
+impl GodotAttrRenameMethod {
+    fn new(_owner: &Reference) -> Self {
+        Self(54)
+    }
+
+    #[method(name = "ask")]
+    fn answer(&self) -> i64 {
+        self.0
+    }
+}
+
+crate::godot_itest! { test_derive_nativeclass_godot_attr_rename_method {
+    let thing = Instance::<GodotAttrRenameMethod, _>::new();
+    let base = thing.into_base();
+    assert_eq!(unsafe { base.call("ask", &[]).to::<i64>() }, Some(54));
+}}
+
+// ----------------------------------------------------------------------------------------------------------------------------------------------
+
+#[derive(NativeClass)]
+#[inherit(Reference)]
+struct GodotAttrAllArguments(Rc<RefCell<Vec<i64>>>);
+
+#[methods]
+impl GodotAttrAllArguments {
+    fn new(_owner: &Reference) -> Self {
+        let vec = Vec::from([12, 34]);
+        let rc_ref = Rc::new(RefCell::new(vec));
+        Self(rc_ref)
+    }
+
+    #[method(rpc = "disabled", name = "ask", deref_return)]
+    fn answer(&self, #[base] _base: &Reference) -> cell::Ref<Vec<i64>> {
+        self.0.borrow()
+    }
+}
+
+crate::godot_itest! { test_derive_nativeclass_godot_attr_all_arguments {
+    let thing = Instance::<GodotAttrAllArguments, _>::new();
+    let base = thing.into_base();
+
+    let res = unsafe { base.call("ask", &[]).to::<Vec<i64>>() };
+    assert_eq!(res, Some([12, 34].into()));
+}}
+
+// ----------------------------------------------------------------------------------------------------------------------------------------------
+
+#[derive(NativeClass)]
+#[inherit(Node)]
+struct CustomGetSet {
+    pub get_called: Cell<i32>,
+    pub set_called: Cell<i32>,
+    #[allow(dead_code)]
+    #[property(get_ref = "Self::get_foo", set = "Self::set_foo")]
+    pub foo: Property<i32>,
+    pub _foo: i32,
+}
+
+#[methods]
+impl CustomGetSet {
+    fn new(_onwer: &Node) -> Self {
+        Self {
+            get_called: Cell::new(0),
+            set_called: Cell::new(0),
+            foo: Property::default(),
+            _foo: 0,
+        }
+    }
+
+    fn get_foo(&self, _owner: TRef<Node>) -> &i32 {
+        self.get_called.set(self.get_called.get() + 1);
+        &self._foo
+    }
+
+    fn set_foo(&mut self, _owner: TRef<Node>, value: i32) {
+        self.set_called.set(self.set_called.get() + 1);
+        self._foo = value;
+    }
+}
+
+crate::godot_itest! { test_derive_nativeclass_with_property_get_set {
+    use gdnative::export::user_data::Map;
+    let (owner, script) = CustomGetSet::new_instance().decouple();
+    script
+        .map(|script| {
+            assert_eq!(0, script.get_called.get());
+            assert_eq!(0, script.set_called.get());
+        })
+        .unwrap();
+    owner.set("foo", 1);
+    script
+        .map(|script| {
+            assert_eq!(0, script.get_called.get());
+            assert_eq!(1, script.set_called.get());
+            assert_eq!(1, script._foo);
+        })
+        .unwrap();
+    assert_eq!(1, i32::from_variant(&owner.get("foo")).unwrap());
+    script
+        .map(|script| {
+            assert_eq!(1, script.get_called.get());
+            assert_eq!(1, script.set_called.get());
+        })
+        .unwrap();
+    owner.free();
+}}
+
+// ----------------------------------------------------------------------------------------------------------------------------------------------
+
+#[derive(NativeClass)]
+struct MyVec {
+    vec: Vec<i32>,
+    #[allow(dead_code)]
+    #[property(get = "Self::get_size")]
+    size: Property<u32>,
+}
+
+#[methods]
+impl MyVec {
+    fn new(_owner: TRef<Reference>) -> Self {
+        Self {
+            vec: Vec::new(),
+            size: Property::default(),
+        }
+    }
+
+    fn add(&mut self, val: i32) {
+        self.vec.push(val);
+    }
+
+    fn get_size(&self, _owner: TRef<Reference>) -> u32 {
+        self.vec.len() as u32
+    }
+}
+
+crate::godot_itest! { test_derive_nativeclass_property_with_only_getter {
+    use gdnative::export::user_data::MapMut;
+    let (owner, script) = MyVec::new_instance().decouple();
+    assert_eq!(u32::from_variant(&owner.get("size")).unwrap(), 0);
+
+    script.map_mut(|script| script.add(42)).unwrap();
+    assert_eq!(u32::from_variant(&owner.get("size")).unwrap(), 1);
+
+    // check the setter doesn't work for `size`
+    let _ = std::panic::catch_unwind(|| owner.set("size", 3));
+    assert_eq!(u32::from_variant(&owner.get("size")).unwrap(), 1);
+}}
