@@ -11,24 +11,11 @@ pub(crate) fn run_tests() -> bool {
 
 pub(crate) fn register(_handle: InitHandle) {}
 
-fn test_vararray_return_crash() -> bool {
-    println!(" -- test_vararray_return_crash");
+crate::godot_itest! { test_vararray_return_crash {
+    // See https://github.com/godot-rust/godot-rust/issues/422
+    let camera = Camera::new();
 
-    let ok = std::panic::catch_unwind(|| {
-        // See https://github.com/godot-rust/godot-rust/issues/422
-        let camera = Camera::new();
-
-        camera.set_frustum(5.0, Vector2::new(1.0, 2.0), 0.0, 1.0);
-
-        camera.get_frustum(); // this should not crash!
-
-        camera.free();
-    })
-    .is_ok();
-
-    if !ok {
-        gdnative::godot_error!("   !! Test test_vararray_return_crash failed");
-    }
-
-    ok
-}
+    camera.set_frustum(5.0, Vector2::new(1.0, 2.0), 0.0, 1.0);
+    camera.get_frustum(); // this should not crash!
+    camera.free();
+}}
